### 2024-11-23 #7520 @editorialbot
[PRE REVIEW]: cellular_raza: Cellular Agent-based Modeling from a Clean Slate
**Submitting author:** @jonaspleyer (<a href="http://orcid.org/0009-0001-0613-7978">Jonas Pleyer</a>)
**Repository:** https://github.com/jonaspleyer/cellular_raza
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.6
