### 2024-11-23 #7520 @danielskatz
👋 @jonaspleyer - Thanks for your submission.

However, note that your paper is about 3 times longer than the maximum recommended JOSS paper length. Please take a at the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). Please use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. You can also use the command `@editorialbot check references` to check a number of things, including the paper length. editorialbot commands need to be the first entry in a new comment.

In addition, the branch with your paper does not seem to include the software as well, but only has the paper. Typically, the paper branch is the same as the main branch, with the addition of the paper. Can you fix this?
