+++ paper.md
@@ -1,0 +1,422 @@
+---
+title: 'cellular_raza: Cellular Agent-based Modeling from a Clean Slate'
+tags:
+  - rust
+  - biology
+  - agent-based
+  - cellular
+authors:
+  - name: Jonas Pleyer
+    orcid: 0009-0001-0613-7978
+    affiliation: 1
+  - name: Christian Fleck
+    affiliation: 1
+affiliations:
+ - name: Freiburg Center for Data-Analysis and Modeling
+   index: 1
+date: 01 June 2024
+bibliography: paper.bib
+
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+# aas-journal: Astrophysical Journal <- The name of the AAS journal.
+---
+
+# Summary
+
+`cellular_raza` is a cellular agent-based modeling framework which allows researchers to construct
+models from a clean slate.
+In contrast to other agent-based modeling toolkits, `cellular_raza` was designed to be free of
+assumptions about the underlying cellular representation.
+This enables researchers to build up complex models while retaining full control over every
+parameter introduced.
+It comes with predefined building blocks for agents and their physical domain to quickly
+construct new simulations bottom-up.
+Furthermore, `cellular_raza` can be used with the `pyo3` and `maturin` packages and thus act as a
+numerical backend to a python package.
+
+# Statement of Need
+
+Agent-based models have become popular in cellular biology
+[@Mogilner2016; @Cess2022; @Delile2017; @Delile_Herrmann_Peyrieras_Doursat_2017].
+<!-- and many tools have been developed so far to asses specific questions in specialized fields -->
+While these tools have proven to be effective for targeted research questions,
+they often lack the ability to be applied for multiple distinct use-cases in a more generic context.
+Nevertheless, core functionalities such as numerical solvers, storage solutions, domain
+decomposition methods and functions to construct these simulations could be shared between models
+if written generically.
+In order to address this issue and construct models from first principles without any assumptions
+regarding the underlying complexity or abstraction level, we developed `cellular_raza`.
+
+# State of Field
+## General-Purpose Agent-Based Modeling Toolkits
+
+<!-- There exist a wide variety of many general-purpose agent-based simulation toolkits which are being
+actively applied in a different fields of study [@Abar2017; @Datseris2022; @Wilensky:1999]. -->
+General-purpose agent-based toolkits are often designed without specific applications in mind
+[@Abar2017; @Datseris2022; @Wilensky:1999].
+They are often able to define agents bottom-up and can be a good choice if they allow for the
+desired cellular representation.
+However, they lack the explicit forethough to be applied in cellular systems.
+Since they are required to solve a wider range of problems they are not able to make assumptions on
+the type of agent or the nature of their interactions and thus miss out on possible
+performance optimizations and advanced numerical solvers.
+
+## Cellular Agent-Based Frameworks
+
+In our previous efforts [@Pleyer2023], we assessed the overall state of modelling toolkits for
+individual-based cellular simulations.
+The frameworks reviewed are all designed for specific use cases and often require a large number of
+parameters which are often unknown in practice and difficult to determine experimentally.
+This is an inherent problem for the applicability of the software and the ability to properly
+interpret results.
+Few modelling frameworks exist that provide a significant degree of flexibility and customisation in
+the definition of cell agents.
+Chaste [@Cooper2020] allows reuse of individual components of its simulation code, such as ODE and
+PDE solvers, but is only partially cell-based.
+Biocellion [@Kang2014] has support for different cell shapes such as spheres and cylinders, but
[…]
