+++ paper.bib
@@ -1,0 +1,358 @@
+@article{Pleyer2023,
+  title = {Agent-based models in cellular systems},
+  volume = {10},
+  ISSN = {2296-424X},
+  url = {http://dx.doi.org/10.3389/fphy.2022.968409},
+  DOI = {10.3389/fphy.2022.968409},
+  journal = {Frontiers in Physics},
+  publisher = {Frontiers Media SA},
+  author = {Pleyer,  Jonas and Fleck,  Christian},
+  year = {2023},
+  month = jan 
+}
+
+@article{Abar2017,
+  title = {Agent Based Modelling and Simulation tools: A review of the state-of-art software},
+  volume = {24},
+  ISSN = {1574-0137},
+  url = {http://dx.doi.org/10.1016/j.cosrev.2017.03.001},
+  DOI = {10.1016/j.cosrev.2017.03.001},
+  journal = {Computer Science Review},
+  publisher = {Elsevier BV},
+  author = {Abar,  Sameera and Theodoropoulos,  Georgios K. and Lemarinier,  Pierre and O’Hare,  Gregory M.P.},
+  year = {2017},
+  month = may,
+  pages = {13–33}
+}
+
+@article{Rodgers1985,
+  title = {Improvements in multiprocessor system design},
+  volume = {13},
+  ISSN = {0163-5964},
+  url = {http://dx.doi.org/10.1145/327070.327215},
+  DOI = {10.1145/327070.327215},
+  number = {3},
+  journal = {ACM SIGARCH Computer Architecture News},
+  publisher = {Association for Computing Machinery (ACM)},
+  author = {Rodgers,  David P.},
+  year = {1985},
+  month = jun,
+  pages = {225–231}
+}
+
+@article{kawasakiModelingSpatioTemporalPatterns1997,
+  title = {Modeling {{Spatio-Temporal Patterns Generated byBacillus}} Subtilis},
+  author = {Kawasaki, K. and Mochizuki, A. and Matsushita, M. and Umeda, T. and Shigesada, N.},
+  year = {1997},
+  month = sep,
+  doi = {10.1006/jtbi.1997.0462},
+}
+
+@article{matsushitaInterfaceGrowthPattern1998,
+  title = {Interface Growth and Pattern Formation in Bacterial Colonies},
+  author = {Matsushita, M. and Wakita, J. and Itoh, H. and R{\`a}fols, I. and Matsuyama, T. and Sakaguchi, H. and Mimura, M.},
+  year = {1998},
+  month = jan,
+  doi = {10.1016/S0378-4371(97)00511-6},
+}
+
+@article{Bouyer2008,
+  title = {Two-Dimensional Patterning by a Trapping/Depletion Mechanism: The Role of TTG1 and GL3 in Arabidopsis Trichome Formation},
+  volume = {6},
+  ISSN = {1545-7885},
+  url = {http://dx.doi.org/10.1371/journal.pbio.0060141},
+  DOI = {10.1371/journal.pbio.0060141},
+  number = {6},
+  journal = {PLoS Biology},
+  publisher = {Public Library of Science (PLoS)},
+  author = {Bouyer,  Daniel and Geier,  Florian and Kragler,  Friedrich and Schnittger,  Arp and Pesch,  Martina and Wester,  Katja and Balkunde,  Rachappa and Timmer,  Jens and Fleck,  Christian and H\"{u}lskamp,  Martin},
+  editor = {Weigel,  Detlef},
+  year = {2008},
+  month = jun,
+  pages = {e141}
+}
+
+@article{Arajo2017,
+  title = {Stochastic gene expression in Arabidopsis thaliana},
+  volume = {8},
[…]
