+++ docker.sh
@@ -1,0 +1,6 @@
+docker run --rm -it \
+    -v $PWD:/data \
+    -u $(id -u):$(id -g) \
+    openjournals/inara \
+    -o pdf \
+    paper.md
