### 2024-11-23 #7520 @jonaspleyer
@danielskatz Thanks for initiating the pre-review process. I was aware that my paper is longer than the expected maximum size. I am very happy to trim it down. Is it reasonable to make this decision via the review process or would you like me to make these changes before advancing?
