### 2024-11-23 #7520 @jonaspleyer
> In addition, the branch with your paper does not seem to include the software as well, but only has the paper. Typically, the paper branch is the same as the main branch, with the addition of the paper. Can you fix this?

@danielskatz
I have updated the `paper` branch to also include the most recent version of `cellular_raza`. The editorialbot seems to pick up the corresponding information correctly.
