+++ paper.md
@@ -17,10 +17,6 @@ affiliations:
 date: 01 June 2024
 bibliography: paper.bib
 
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
@@ -52,8 +48,6 @@ regarding the underlying complexity or abstraction level, we developed `cellular
 # State of Field
 ## General-Purpose Agent-Based Modeling Toolkits
 
-<!-- There exist a wide variety of many general-purpose agent-based simulation toolkits which are being
-actively applied in a different fields of study [@Abar2017; @Datseris2022; @Wilensky:1999]. -->
 General-purpose agent-based toolkits are often designed without specific applications in mind
 [@Abar2017; @Datseris2022; @Wilensky:1999].
 They are often able to define agents bottom-up and can be a good choice if they allow for the
@@ -99,106 +93,6 @@ Each simulation aspect is formulated as a trait in Rust's type system, which pro
 abstractions.
 The getting-started guide provides a good entry point and explains every step from building, running
 to visualising.
-
-<!--
-# Underlying Assumptions and Internals
-
-## List of Simulation Aspects
-
-`cellular_raza` assumes that all dynamics can be categorized into what we call "simulation
-aspects".
-They represent cellular processes, interactions, changes of the simulation domain and interactions
-with the external environment.
-
-| Aspect | Description | Depends on |
-| --- | --- | --- |
-| **Cellular Agent** | | |
-| `Position` | Spatial representation of the cell | |
-| `Velocity` | Spatial velocity of the cell | |
-| `Mechanics` | Calculates the next increment from given force, velocity and position. | `Position` and `Velocity` |
-| `Interaction` | Calculates force acting between agents. Also reacts to neighbors. | `Position` and `Velocity` |
-| `Cycle` | Changes core properties of the cell. Responsible for cell-division and death. | |
-| `Intracellular` | Intracellular representation of the cell. | |
-| `Reactions` | Intracellular reactions | `Intracellular` |
-| `ReactionsExtra` | Couples intra- & extracellular reactions | `DomainReactions` |
-| `ReactionsContact` | Models reactions between cells purely by contact | `Position`, `Intracellular` |
-| **Simulation Domain** | | |
-| `Domain` | Represents the physical simulation domain. | |
-| `DomainMechanics` | Apply boundary conditions to agents. | `Position`, `Velocity` |
-| `DomainForce` | Apply a spatially-dependent force onto the cell. | `Mechanics` |
-| `DomainReactions` | Calculate extracellular reactions and effects such as diffusion. | `ReactionsExtra` |
-| **Other** | | |
-| `Controller` | Externally apply changes to the cells. | |
-
-## Spatially Localized Interactions
-
-One useful assumption within `cellular_raza` is that each and every interaction is of finite range.
-This means that cellular agents only interact with a limited amount of neighbors and close
-environment.
-Any long-ranged interactions must be the result of a collection of short-ranged interactions.
-This assumption enables us to split the simulation domain into chunks and process them individually
-although some communication is needed in order to deal with boundary conditions.
-In practice, this means that any interaction force should be given a cutoff.
-It also means that any interactions which need to be evaluated between agents should in theory scale
-linearly with the number of agents $\mathcal{O}(n_\text{agents})$.
-
-## Code Structure
-
-`cellular_raza` consists of multiple crates working in tandem.
-It was designed to have clear separations between conceptual choices and implementation details.
[…]
