### 2024-11-25 #7520 @jonaspleyer
@danielskatz It seems that the editorialbot has counted words which were commented out. I had left some previous sections in comments  in the source code of the `paper.md` file in case we need access to this information. When removing these paragraphs, the total word-count is approximately half (1414 words)
It seems that only the source file `paper.md` is scanned which also includes the preamble and other commands that will be compiled (such as figures). Without checking, it seems that the `wc` utility of the GNU coreutils is being used. When accounting for some of these additional contributions and gradually reducing the file, we obtain the following table.

| | Wordcount |
| --- | --- |
| Full | 2826 |
| Without Comments | 1345 |
| Without Preamble | 1293 |
| Without Figures | 1266 |
| Without Headings | 1220 |

It is clear that we are still over the desired 1000 words but much closer than thought initially. We will nevertheless try to compact the paper.
