### 2024-11-25 f5b64ce
remove commented sentence
