### 2024-11-25 7f84ceb
strike through text which could be omitted
