+++ paper.md
@@ -37,61 +37,63 @@ numerical backend to a python package.
 Agent-based models have become popular in cellular biology
 [@Mogilner2016; @Cess2022; @Delile2017; @Delile_Herrmann_Peyrieras_Doursat_2017].
 While these tools have proven to be effective for targeted research questions,
-they often lack the ability to be applied for multiple distinct use-cases in a more generic context.
-Nevertheless, core functionalities such as numerical solvers, storage solutions, domain
-decomposition methods and functions to construct these simulations could be shared between models
-if written generically.
+they often lack the ability to be applied ~~for multiple distinct use-cases in a more generic
+context.~~ more generically.
+Nevertheless, core functionalities such as numerical solvers, storage solutions or domain
+decomposition methods ~~and functions to construct these simulations~~ could be shared between models
+~~if written generically~~.
 In order to address this issue and construct models from first principles without any assumptions
 regarding the underlying complexity or abstraction level, we developed `cellular_raza`.
 
 # State of Field
-## General-Purpose Agent-Based Modeling Toolkits
 
 General-purpose agent-based toolkits are often designed without specific applications in mind
 [@Abar2017; @Datseris2022; @Wilensky:1999].
 They are often able to define agents bottom-up and can be a good choice if they allow for the
 desired cellular representation.
-However, they lack the explicit forethough to be applied in cellular systems.
+However, they lack the explicit forethough to be applied in cellular systems~~.
 Since they are required to solve a wider range of problems they are not able to make assumptions on
 the type of agent or the nature of their interactions and thus miss out on possible
-performance optimizations and advanced numerical solvers.
+performance optimizations and advanced numerical solvers.~~
+and may not be able to describe every cellular aspect.
 
 ## Cellular Agent-Based Frameworks
 
 In our previous efforts [@Pleyer2023], we assessed the overall state of modelling toolkits for
 individual-based cellular simulations.
-The frameworks reviewed are all designed for specific use cases and often require a large number of
-parameters which are often unknown in practice and difficult to determine experimentally.
-This is an inherent problem for the applicability of the software and the ability to properly
-interpret results.
-Few modelling frameworks exist that provide a significant degree of flexibility and customisation in
+These frameworks ~~reviewed~~ are ~~all~~ designed for specific us~~e~~ages ~~cases~~ and often 
+require ~~a large number of~~ many
+parameters which are ~~often~~ unknown or ~~in practice and~~ difficult to determine experimentally.
+This poses an inherent problem for their applicability ~~of the software~~ and the ability to
+properly interpret results.
+Few modelling frameworks exist that provide a significant degree of flexibility and customization in
 the definition of cell agents.
-Chaste [@Cooper2020] allows reuse of individual components of its simulation code, such as ODE and
-PDE solvers, but is only partially cell-based.
-Biocellion [@Kang2014] has support for different cell shapes such as spheres and cylinders, but
-admits that their current approach lacks flexibility in the subcellular description.
+Chaste [@Cooper2020] allows reuse of individual components ~~of its simulation code~~, such as ODE
+and PDE solvers, but is only partially cell-based.
+Biocellion [@Kang2014] ~~has~~ supports ~~for~~ different cell shapes such as spheres and cylinders,
+but admits that their current approach lacks flexibility in the subcellular description.
 BioDynaMo [@breitwieser_biodynamo_2022] offers some modularity in the choice of components for
-cellular agents, but cannot freely customise the cellular representation.
+cellular agents, but cannot ~~freely~~ deeply customize the cellular representation.
 
 # cellular_raza
 
[…]
