### 2024-11-25 9130749
remove strikethrough text; reformulate; format
