+++ paper.md
@@ -37,11 +37,10 @@ numerical backend to a python package.
 Agent-based models have become popular in cellular biology
 [@Mogilner2016; @Cess2022; @Delile2017; @Delile_Herrmann_Peyrieras_Doursat_2017].
 While these tools have proven to be effective for targeted research questions,
-they often lack the ability to be applied ~~for multiple distinct use-cases in a more generic
-context.~~ more generically.
+they often lack the ability to be applied more generically.
 Nevertheless, core functionalities such as numerical solvers, storage solutions or domain
-decomposition methods ~~and functions to construct these simulations~~ could be shared between models
-~~if written generically~~.
+decomposition methods could be shared between models
+.
 In order to address this issue and construct models from first principles without any assumptions
 regarding the underlying complexity or abstraction level, we developed `cellular_raza`.
 
@@ -51,47 +50,42 @@ General-purpose agent-based toolkits are often designed without specific applica
 [@Abar2017; @Datseris2022; @Wilensky:1999].
 They are often able to define agents bottom-up and can be a good choice if they allow for the
 desired cellular representation.
-However, they lack the explicit forethough to be applied in cellular systems~~.
-Since they are required to solve a wider range of problems they are not able to make assumptions on
-the type of agent or the nature of their interactions and thus miss out on possible
-performance optimizations and advanced numerical solvers.~~
+However, they lack the explicit forethough to be applied in cellular systems
 and may not be able to describe every cellular aspect.
 
 ## Cellular Agent-Based Frameworks
 
 In our previous efforts [@Pleyer2023], we assessed the overall state of modelling toolkits for
 individual-based cellular simulations.
-These frameworks ~~reviewed~~ are ~~all~~ designed for specific us~~e~~ages ~~cases~~ and often 
-require ~~a large number of~~ many
-parameters which are ~~often~~ unknown or ~~in practice and~~ difficult to determine experimentally.
-This poses an inherent problem for their applicability ~~of the software~~ and the ability to
+These frameworks are designed for specific usages and often 
+require many
+parameters which are unknown or difficult to determine experimentally.
+This poses an inherent problem for their applicability and the ability to
 properly interpret results.
 Few modelling frameworks exist that provide a significant degree of flexibility and customization in
 the definition of cell agents.
-Chaste [@Cooper2020] allows reuse of individual components ~~of its simulation code~~, such as ODE
+Chaste [@Cooper2020] allows reuse of individual components , such as ODE
 and PDE solvers, but is only partially cell-based.
-Biocellion [@Kang2014] ~~has~~ supports ~~for~~ different cell shapes such as spheres and cylinders,
+Biocellion [@Kang2014] supports different cell shapes such as spheres and cylinders,
 but admits that their current approach lacks flexibility in the subcellular description.
 BioDynaMo [@breitwieser_biodynamo_2022] offers some modularity in the choice of components for
-cellular agents, but cannot ~~freely~~ deeply customize the cellular representation.
+cellular agents, but cannot deeply customize the cellular representation.
 
 # cellular_raza
 
-We distinguish between different simulation aspects, i.e.,  mechanics, interaction, or cell cycle.
+We distinguish between different simulation aspects, i.e., mechanics, interaction, or cell cycle.
 These aspects are directly related to the properties of the cells, domain, or other external
 interactions.
 The user selects a cellular representation, which can be built from pre-existing building blocks or
-~~a~~ fully customized bottom-up ~~approach~~, if desired.
-'cellular_raza' utilizes macros to generate code contingent on the simulation aspects ~~being solved
-numerically~~.
+ fully customized bottom-up , if desired.
[…]
