### 2024-11-26 ca801a4
minor reformulations to enhance flow of wording
