### 2024-12-02 5574fa7
simplify equations
