### 2024-12-02 534313c
redefine MiePotential and fix implementation
