+++ cellular_raza-building-blocks/src/cell_building_blocks/interaction.rs
@@ -251,114 +251,98 @@ macro_rules! implement_morse_potential(
 implement_morse_potential!(MorsePotential, f64);
 implement_morse_potential!(MorsePotentialF32, f32);
 
-/// Generalizeation of the [BoundLennardJones] potential for arbitrary floating point types.
-///
-/// \\begin{align}
-///     U(r) &= C\epsilon\left[ \left(\frac{\sigma}{r}\right)^n -
-///         \left(\frac{\sigma}{r}\right)^m\right]\\\\
-///     C &= \frac{n}{n-m}\left(\frac{n}{m}\right)^{\frac{n}{n-m}}\\\\
-///     V(r) &= \min(U(r), \beta)\theta(r-\zeta)
-/// \\end{align}
-///
-// This struct itself does not provide python bindings.
-// We provide specialized types for different floating-point types.
-//
-// | Name | Float Type |
-// | --- | --- |
-// | [MiePotentialF64] | COMING |
-// | [MiePotentialF32] | COMING |
-//
-/// # References
-/// [1]
-/// G. Mie, “Zur kinetischen Theorie der einatomigen Körper,”
-/// Annalen der Physik, vol. 316, no. 8. Wiley, pp. 657–697, Jan. 1903.
-/// doi: [10.1002/andp.19033160802](https://doi.org/10.1002/andp.19033160802).
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
-pub struct MiePotential<const N: usize, const M: usize, F = f64> {
-    /// Interaction strength $\epsilon$ of the potential.
-    pub radius: F,
-    /// Overall size $\sigma$ of the object of the potential.
-    pub potential_strength: F,
-    /// Numerical bound $\beta$ of the interaction strength.
-    pub bound: F,
-    /// Defines a cutoff $\zeta$ after which the potential will be fixed to exactly zero.
-    pub cutoff: F,
-    /// Exponent $n$ of the potential
-    en: F,
-    /// Exponent $m$ of the potential
-    em: F,
-}
-
-impl<F, const D: usize, const N: usize, const M: usize>
-    Interaction<SVector<F, D>, SVector<F, D>, SVector<F, D>, F> for MiePotential<N, M, F>
-where
-    F: nalgebra::RealField + Copy,
-{
-    fn calculate_force_between(
-        &self,
-        own_pos: &SVector<F, D>,
-        _own_vel: &SVector<F, D>,
-        ext_pos: &SVector<F, D>,
-        _ext_vel: &SVector<F, D>,
-        ext_radius: &F,
-    ) -> Result<(SVector<F, D>, SVector<F, D>), CalcError> {
-        let z = own_pos - ext_pos;
-        let r = z.norm();
-        if r == F::zero() {
-            return Err(CalcError(format!(
-                "identical position for two objects. Cannot Calculate force in Mie potential"
-            )));
+macro_rules! implement_mie_potential(
+    ($name:ident, $float_type:ty) => {
+        /// Generalizeation of the [BoundLennardJones] potential.
+        ///
+        /// \\begin{align}
+        ///     U(r) &= C\epsilon\left[ \left(\frac{\sigma}{r}\right)^n -
+        ///         \left(\frac{\sigma}{r}\right)^m\right]\\\\
+        ///     C &= \frac{n}{n-m}\left(\frac{n}{m}\right)^{\frac{n}{n-m}}\\\\
+        ///     V(r) &= \min(U(r), \beta)\theta(r-\zeta)
+        /// \\end{align}
+        ///
+        /// This struct comes in a 64bit version [MiePotential] and a 32bit variant
+        /// [MiePotentialF32].
+        ///
[…]
