### 2024-12-02 0ec0d11
use field initialization instead of new function
