+++ cellular_raza-examples/cell_sorting/src/main-langevin.rs
@@ -120,7 +120,13 @@ fn main() -> Result<(), SimulationError> {
                 rng.gen_range(0.0..DOMAIN_SIZE),
             ];
             Cell {
-                mechanics: Langevin3D::new(pos, [0.0; 3], 10.0, CELL_DAMPING, CELL_KB_TEMP),
+                mechanics: Langevin3D {
+                    pos: pos.into(),
+                    vel: [0.0; 3].into(),
+                    mass: 10.0,
+                    damping: CELL_DAMPING,
+                    kb_temperature: CELL_KB_TEMP,
+                },
                 interaction: CellSpecificInteraction {
                     species: match n <= N_CELLS_BLUE {
                         true => Species::BlueCell,
