+++ cellular_raza-examples/sender-receiver/src/controller.rs
@@ -1,5 +1,5 @@
-use cellular_raza::concepts::reactions_old::*;
 use cellular_raza::concepts::domain_old::*;
+use cellular_raza::concepts::reactions_old::*;
 use nalgebra::DVector;
 use serde::{Deserialize, Serialize};
 
