### 2024-12-02 be953b7
do not use new function but field initialization
