### 2024-12-02 1a43dd1
remove legacy tests for quick_xml
