### 2025-01-08 d0bccb5
add example of how link to CUDA using nvcc within the cara backend
