+++ cellular_raza-core/src/backend/mod.rs
@@ -42,7 +42,5 @@ pub mod cpu_os_threads;
 
 pub mod chili;
 
-/// 🐯 (Placeholder) GPU-centered backend using [OpenCL](https://www.khronos.org/opencl/)
-///
-/// This backend is currently not developed and only here to serve as a placeholder.
-pub mod cara {}
+#[cfg(feature = "cara")]
+pub mod cara;
