+++ cellular_raza-core/Cargo.toml
@@ -16,6 +16,9 @@ homepage.workspace =  true
 rustdoc-args = [ "--html-in-header", ".docs-header.html" ]
 all-features = true
 
+[build-dependencies]
+cc = { version = "1.2.7", features = ["parallel"], optional=true}
+
 [dependencies]
 # Core
 crossbeam-channel = { version="0.5" }
@@ -62,6 +65,7 @@ tracing = ["dep:tracing", "dep:tracing-subscriber"]
 timestamp = ["dep:chrono"]
 gradients = ["cellular_raza-concepts/gradients"]
 pyo3 = ["dep:pyo3"]
+cara = ["dep:cc"]
 
 # [profile.release]
 # debug = 1
