+++ .github/workflows/test_nightly_macos-14.yml
@@ -1,10 +1,10 @@
 on: [push, pull_request]
 
-name: Test-Suite nightly macos-12
+name: Test-Suite nightly macos-14
 
 jobs:
-  CI-nightly-macos-12:
+  CI-nightly-macos-14:
     uses: ./.github/workflows/reuse.yml
     with:
       toolchain: nightly
-      os: macos-12
+      os: macos-14
