### 2025-01-08 63b648c
remove legacy tests
