+++ .github/workflows/test_stable.yml
@@ -1,10 +1,0 @@
-on: [push, pull_request]
-
-name: Test Stable
-
-jobs:
-  CI-stable-ubuntu-latest:
-    uses: ./.github/workflows/reuse.yml
-    with:
-      toolchain: stable
-      os: ubuntu-latest
