+++ .github/workflows/test_beta.yml
@@ -1,10 +1,0 @@
-on: [push, pull_request]
-
-name: Test Beta
-
-jobs:
-  CI-beta-ubuntu-latest:
-    uses: ./.github/workflows/reuse.yml
-    with:
-      toolchain: beta
-      os: ubuntu-latest
