### 2025-01-08 0967be8
remove macos-14 since it is not working with pyo3
