+++ README.md
@@ -19,7 +19,6 @@ A cellular Agent-Based modeling tool.
 |---|---|---|---|
 | `ubuntu-latest` | ![stable-ubuntu-latest](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_stable_ubuntu-latest.yml?style=flat-square&label=CI) |![beta-ubuntu-latest](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_beta_ubuntu-latest.yml?style=flat-square&label=CI) |![nightly-ubuntu-latest](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_nightly_ubuntu-latest.yml?style=flat-square&label=CI) |
 | `macos-13` | ![stable-macos-13](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_stable_macos-13.yml?style=flat-square&label=CI) |![beta-macos-13](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_beta_macos-13.yml?style=flat-square&label=CI) |![nightly-macos-13](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_nightly_macos-13.yml?style=flat-square&label=CI) |
-| `macos-14` | ![stable-macos-14](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_stable_macos-14.yml?style=flat-square&label=CI) |![beta-macos-14](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_beta_macos-14.yml?style=flat-square&label=CI) |![nightly-macos-14](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_nightly_macos-14.yml?style=flat-square&label=CI) |
 | `windows-latest` | ![stable-windows-latest](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_stable_windows-latest.yml?style=flat-square&label=CI) |![beta-windows-latest](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_beta_windows-latest.yml?style=flat-square&label=CI) |![nightly-windows-latest](https://img.shields.io/github/actions/workflow/status/jonaspleyer/cellular_raza/test_nightly_windows-latest.yml?style=flat-square&label=CI) |
 
 ## Documentation
