+++ .github/workflows/test_stable_macos-14.yml
@@ -1,10 +1,0 @@
-on: [push, pull_request]
-
-name: Test-Suite stable macos-14
-
-jobs:
-  CI-stable-macos-14:
-    uses: ./.github/workflows/reuse.yml
-    with:
-      toolchain: stable
-      os: macos-14
