### 2025-01-09 6e0fcc8
insert first basic layout for agent
