### 2025-01-16 #7520 @jonaspleyer
Hi there! Thanks for advancing this submission and apologies for the slow response on my end.
I have looked at the list of people and found some names which might be suitable.

https://reviewers.joss.theoj.org/reviewers/5922
https://reviewers.joss.theoj.org/reviewers/3276
https://reviewers.joss.theoj.org/reviewers/3961
https://reviewers.joss.theoj.org/reviewers/2247
