### 2025-01-19 #7520 @finsberg
> Hi [@Armavica](https://github.com/Armavica), [@finsberg](https://github.com/finsberg), [@johanneskoester](https://github.com/johanneskoester), and [@lazear](https://github.com/lazear) - any of you interessted in reviewing this article?

Hi @mahfuz05062, I would be happy to review this one👍
