### 2025-01-23 d4620d9
add code for switchting between building blocks
- this also prepares for more code which can be inserted into the
  homepage for explanations
- contains custom formatting to enforce maximum width via rustfmt
