+++ cellular_raza-examples/homepage-training/src/lib.rs
@@ -1,0 +1,1 @@
+mod building_blocks;
