### 2025-01-23 1650621
fixup! add code for switchting between building blocks
