### 2025-01-23 11945f2
update version of itertools crate
