### 2025-01-23 8c17448
modularize proc-macro even more to allow for calling of custom other functions than the one defined in chili backend
