### 2025-01-23 51f6cdb
remove unused command
