### 2025-01-23 41aa98b
add cara and elli backend features, dependencies and mod files
