+++ cellular_raza-core/src/backend/elli/mod.rs
@@ -1,0 +1,7 @@
+//! 🐺 (Placeholder) Cross-platform GPU-centered backend using
+//! [wgpu](https://docs.rs/wgpu/latest/wgpu/)
+
+#[test]
+fn test_wgpu() {
+    assert!(true);
+}
