+++ cellular_raza-core/src/backend/cara/cara.cu
@@ -2,7 +2,7 @@
 #include <cinttypes>
 #include <cstdio>
 
-typedef uint32_t UID;
+typedef uint64_t UID;
 
 struct Position {
     float p[3];
@@ -24,30 +24,85 @@ struct Agent {
 
 struct AgentContainer {
     Agent agent;
-    Force current_force;
 };
 
 struct Runner {
-    uint32_t n_agents;
+    /// Counter to assign unique identifiers to new agents
+    uint64_t uid_counter;
+
+    /// Number of agents currently stored
+    uint64_t n_agents;
+
+    /// Values of the Agents are fully updated.
+    bool agents_up_to_date;
+
     // This is a pointer to a single element which marks the first element of an array.
     // The length of this array is given by the previous value.
-    Agent *agents;
+    AgentContainer *agents;
+
+    /// Pointer to an array of floats which stores the positions of the agents
+    float *positions;
+
+    /// Pointer to an array of floats which stores the velocities of the agents
+    float *velocities;
+
+    /// Pointer to an array of floats which stores the forces of the agents
+    float *forces;
+
+    /// Function pointer which calculates the interaction force between two agents.
+    // Force (*fptr)(Position, Velocity, Position, Velocity);
 };
 
+__global__ void calculate_forces(int total, float *pos1, float *vel1, float *pos2, float *vel2,
+                                 float *forces) {
+    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    int j = blockIdx.y * blockDim.y + threadIdx.y;
+    if (i >= total || j >= total) {
+        return;
+    }
+
+    float p1, p2;
+    // float v1, v2;
+    p1 = pos1[i];
+    // v1 = vel1[i];
+    p2 = pos2[j];
+    // v2 = vel2[j];
+
+    forces[i] += 0.1 * (p1 - p2);
+    forces[j] += 0.1 * (p2 - p1);
+}
+
+/// This calculates with time complexity O(n^2) which is not ideal but the most simple thing that
+// one can write down so we will stick with it for now.
 extern "C" void update_positions(Runner &runner) {
-    for (uint32_t i = 0; i < runner.n_agents; i++) {
-        for (uint32_t j = 0; j < runner.n_agents; j++) {
+    for (uint64_t i = 0; i < runner.n_agents; i++) {
+        for (uint64_t j = 0; j < runner.n_agents; j++) {
+            float p1[3] = {runner.positions[i], runner.positions[i + 1], runner.positions[i + 2]};
+            /* float v1[3] = {runner.velocities[i], runner.velocities[i + 1],
+                           runner.velocities[i + 2]};*/
+            float p2[3] = {runner.positions[j], runner.positions[j + 1], runner.positions[j + 2]};
+            /* float v2[3] = {runner.velocities[j], runner.velocities[j + 1],
+                           runner.velocities[j + 2]};*/
+
+            float f[3];
+            for (uint8_t k = 0; k < 3; k++) {
+                // Calculate the force between agents
+                f[k] += 0.1 * (p1 - p2);
+
+                // Update forces
+                runner.forces[i + k] += f[k];
+                runner.forces[j + k] -= f[k];
+            }
         }
     }
 }
 
 extern "C" void print_positions(Runner &runner) {
-    printf("%p\n\n\n", runner.agents);
-    for (uint32_t i = 0; i < runner.n_agents; i++) {
-        Agent agent = runner.agents[i];
+    for (uint64_t i = 0; i < runner.n_agents; i++) {
+        Agent agent = runner.agents[i].agent;
[…]
