### 2025-01-23 ec19181
fixup! add cara and elli backend features, dependencies and mod files
