### 2025-01-23 0f0a393
update slogan
