+++ cellular_raza/Cargo.toml
@@ -4,7 +4,7 @@ name = "cellular_raza"
 version.workspace = true
 edition = "2021"
 license = "GPL-2.0"
-description = "Cellular simulation tool"
+description = "Cellular Agent-based Modeling from a Clean Slate"
 readme = "README.md"
 repository = "https://github.com/jonaspleyer/cellular_raza"
 keywords = ["cell", "cellular", "simulation"]
