### 2025-01-24 a5749b5
make compatible with new pyo3 version
- remove deprecated __reduce__ method from CellIdentifier
- add #[pyclass(eq, eq_int)] for enums
