+++ cellular_raza-core/src/storage/concepts.rs
@@ -106,7 +106,7 @@ impl Error for StorageError {}
 ///
 /// We currently support saving results in a [sled] database, or as a json file by using
 /// [serde_json].
-#[cfg_attr(feature = "pyo3", pyo3::pyclass)]
+#[cfg_attr(feature = "pyo3", pyo3::pyclass(eq, eq_int))]
 #[derive(Clone, Debug, Deserialize, PartialEq, Eq, Serialize)]
 pub enum StorageOption {
     /// Save results as [sled] database.
