+++ cellular_raza-core/src/backend/chili/mod.rs
@@ -395,23 +395,6 @@ impl CellIdentifier {
         self.hash(&mut hasher);
         hasher.finish()
     }
-
-    /// Used to pickle the object
-    #[deprecated(
-        note = "This method exists as part of experimentation and may change behaviour or\
-        be removed in the future"
-    )]
-    pub fn __reduce__<'a>(&'a self, py: pyo3::Python<'a>) -> pyo3::Bound<pyo3::types::PyTuple> {
-        use pyo3::prelude::*;
-        use pyo3::PyTypeInfo;
-        pyo3::types::PyTuple::new_bound(
-            py,
-            [
-                Self::type_object_bound(py).to_object(py),
-                pyo3::types::PyTuple::new_bound(py, [self.0 .0 as u64, self.1]).to_object(py),
-            ],
-        )
-    }
 }
 
 /// Contains structs to store aspects of the simulation and macros to construct them.
