### 2025-01-24 07c0c8e
enable pyo3 feature for testing
