### 2025-01-24 5a479cd
remove Box example; add example for when Pos, Vel, For are different in type
