### 2025-01-24 b84e69d
add test for previously added example
