### 2025-01-24 b8e59f3
remove unused, obosolete experimental cpu_rayon backend
