+++ cellular_raza-core/src/backend/cpu_rayon/engine.rs
@@ -1,830 +1,0 @@
-
-use cellular_raza_concepts::{CellAgent, CellAgentBox, CellularIdentifier};
-use cellular_raza_concepts::domain::Index;
-use cellular_raza_concepts::domain::{
-    Concentration, Controller, Domain, ExtracellularMechanics, Voxel,
-};
-use cellular_raza_concepts::{ControllerError, DrawingError, RequestError};
-use cellular_raza_concepts::{CellularReactions, InteractionExtracellularGradient};
-use cellular_raza_concepts::mechanics::{Force, Position, Velocity};
-use cellular_raza_concepts::{CreatePlottingRoot, PlotSelf};
-
-use super::errors::*;
-use crate::storage::{StorageInterface, StorageManager};
-
-use super::domain_decomposition::{
-    AuxiliaryCellPropertyStorage, DomainBox, MultiVoxelContainer, VoxelBox,
-};
-
-use super::config::{
-    ImageType, PlottingConfig, SimulationConfig, SimulationMetaParams, SimulationSetup, TimeSetup,
-    PROGRESS_BAR_STYLE,
-};
-
-use super::config::StorageConfig;
-
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
-use std::thread;
-
-use core::ops::{Add, AddAssign, Mul};
-
-use core::marker::PhantomData;
-
-use hurdles::Barrier;
-
-use serde::{Deserialize, Serialize};
-
-use plotters::{
-    coord::types::RangedCoordf64,
-    prelude::{BitMapBackend, Cartesian2d, DrawingArea},
-};
-
-use indicatif::{ProgressBar, ProgressStyle};
-
-#[derive(Clone, Serialize, Deserialize)]
-pub struct ControllerBox<Cont, Obs> {
-    pub controller: Cont,
-    pub measurements: std::collections::BTreeMap<u32, Obs>,
-}
-
-impl<Cont, Obs> ControllerBox<Cont, Obs> {
-    pub fn measure<'a, I, Cel>(&mut self, thread_index: u32, cells: I) -> Result<(), SimulationError>
-    where
-        Cel: 'a + Serialize + for<'b> Deserialize<'b>,
-        I: Iterator<Item = &'a CellAgentBox<Cel>> + Clone,
-        Cont: Controller<Cel, Obs>,
-    {
-        let obs = self.controller.measure(cells)?;
-        self.measurements.insert(thread_index, obs);
-        Ok(())
-    }
-
-    pub fn adjust<'a, Cel, J>(&mut self, cells: J) -> Result<(), ControllerError>
-    where
-        Cel: 'a + Serialize + for<'b> Deserialize<'b>,
-        J: Iterator<
-            Item = (
-                &'a mut CellAgentBox<Cel>,
-                &'a mut Vec<cellular_raza_concepts::CycleEvent>,
-            ),
-        >,
-        Cont: Controller<Cel, Obs>,
-    {
-        self.controller.adjust(self.measurements.values(), cells)
-    }
-}
-
-/// # Supervisor controlling simulation execution
-///
-pub struct SimulationSupervisor<MVC, Dom, Cel, Cont = (), Obs = ()>
-where
-    Cel: Serialize + for<'a> Deserialize<'a>,
-    Dom: Serialize + for<'a> Deserialize<'a>,
-    Cont: Serialize + for<'a> Deserialize<'a>,
-{
-    pub worker_threads: Vec<thread::JoinHandle<Result<MVC, SimulationError>>>,
-    pub multivoxelcontainers: Vec<MVC>,
-
-    pub time: TimeSetup,
[…]
