### 2025-01-24 c36163f
create new feature flag for cpu_os_threads backend; this is no default from now on
