+++ cellular_raza/Cargo.toml
@@ -44,4 +44,6 @@ tracing = ["cellular_raza-core/tracing"]
 timestamp = ["cellular_raza-core/timestamp"]
 gradients = ["cellular_raza-concepts/gradients", "cellular_raza-core/gradients", "cellular_raza-building-blocks/gradients"]
 pyo3 = ["cellular_raza-building-blocks/pyo3", "cellular_raza-core/pyo3"]
+
+cpu_os_threads = ["cellular_raza-core/cpu_os_threads"]
 cara = ["cellular_raza-core/cara"]
