+++ cellular_raza-core/Cargo.toml
@@ -26,7 +26,7 @@ hurdles = { version="1.0" }
 itertools = { workspace = true }
 num = { workspace = true }
 # ode_integrate = { version="0.0.2" }
-plotters = { workspace = true }
+plotters = { workspace = true, optional = true }
 rayon = { version="1.6" }
 kdam = { version="0.6", features = ["template", "rayon"] }
 
@@ -71,6 +71,7 @@ tracing = ["dep:tracing", "dep:tracing-subscriber"]
 timestamp = ["dep:chrono"]
 gradients = ["cellular_raza-concepts/gradients"]
 pyo3 = ["dep:pyo3"]
+cpu_os_threads = ["dep:plotters",]
 cara = ["dep:cc", "dep:cudarc"]
 elli = ["dep:wgpu"]
 
