### 2025-01-24 342a9e3
add feature flag for chili backend and make it default
