+++ cellular_raza-core/src/backend/mod.rs
@@ -42,6 +42,8 @@
 #[cfg_attr(docsrs, doc(cfg(feature = "cpu_os_threads")))]
 pub mod cpu_os_threads;
 
+#[cfg(feature = "chili")]
+#[cfg_attr(docsrs, doc(cfg(feature = "chili")))]
 pub mod chili;
 
 #[cfg(feature = "cara")]
