+++ cellular_raza-core/Cargo.toml
@@ -66,12 +66,13 @@ version = "0.1.6"
 tempdir.workspace = true
 
 [features]
-default = ["gradients", "timestamp"]
+default = ["gradients", "timestamp", "chili"]
 tracing = ["dep:tracing", "dep:tracing-subscriber"]
 timestamp = ["dep:chrono"]
 gradients = ["cellular_raza-concepts/gradients"]
 pyo3 = ["dep:pyo3"]
 cpu_os_threads = ["dep:plotters",]
+chili = []
 cara = ["dep:cc", "dep:cudarc"]
 elli = ["dep:wgpu"]
 
