+++ cellular_raza/Cargo.toml
@@ -48,3 +48,4 @@ pyo3 = ["cellular_raza-building-blocks/pyo3", "cellular_raza-core/pyo3"]
 cpu_os_threads = ["cellular_raza-core/cpu_os_threads"]
 chili = ["cellular_raza-core/chili"]
 cara = ["cellular_raza-core/cara"]
+elli = ["cellular_raza-core/elli"]
