+++ cellular_raza/src/prelude.rs
@@ -1,6 +1,22 @@
 pub use cellular_raza_building_blocks::*;
 pub use cellular_raza_concepts::*;
+
+#[cfg(feature = "cpu_os_threads")]
+#[cfg_attr(docsrs, doc(cfg(feature = "cpu_os_threads")))]
+pub use cellular_raza_core::backend::cpu_os_threads::*;
+
+#[cfg(feature = "chili")]
+#[cfg_attr(docsrs, doc(cfg(feature = "chili")))]
 pub use cellular_raza_core::backend::chili::*;
+
+#[cfg(feature = "cara")]
+#[cfg_attr(docsrs, doc(cfg(feature = "cara")))]
+pub use cellular_raza_core::backend::cara::*;
+
+#[cfg(feature = "elli")]
+#[cfg_attr(docsrs, doc(cfg(feature = "elli")))]
+pub use cellular_raza_core::backend::elli::*;
+
 pub use cellular_raza_core::storage::*;
 pub use cellular_raza_core::time::*;
 pub use cellular_raza_core::*;
