### 2025-01-24 635845a
update wgpu dependency
