+++ cellular_raza-core/Cargo.toml
@@ -52,7 +52,7 @@ uniquevec = { version = "0.1.0", features = ["serde"] }
 cudarc = { version = "0.13", optional = true, features = ["cuda-version-from-build-system"]}
 
 # Additional dependencies for elli backend
-wgpu = { version = "23.0", optional = true }
+wgpu = { version = "24.0", optional = true }
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
