### 2025-01-24 2e30472
remove gradients from default features
