+++ cellular_raza-core/Cargo.toml
@@ -66,7 +66,7 @@ version = "0.1.6"
 tempdir.workspace = true
 
 [features]
-default = ["gradients", "timestamp", "chili"]
+default = ["timestamp", "chili"]
 tracing = ["dep:tracing", "dep:tracing-subscriber"]
 timestamp = ["dep:chrono"]
 gradients = ["cellular_raza-concepts/gradients"]
