### 2025-01-24 5c0331c
inherit edition and license from workspace
