+++ cellular_raza/Cargo.toml
@@ -2,8 +2,8 @@
 authors = ["Jonas Pleyer <jonas.pleyer@fdm.uni-freiburg.de>"]
 name = "cellular_raza"
 version.workspace = true
-edition = "2021"
-license = "GPL-2.0"
+edition.workspace = true
+license.workspace = true
 description = "Cellular Agent-based Modeling from a Clean Slate"
 readme = "README.md"
 repository = "https://github.com/jonaspleyer/cellular_raza"
