+++ cellular_raza/Cargo.toml
@@ -39,7 +39,7 @@ nalgebra.workspace = true
 ode_integrate = "0.0.2"
 
 [features]
-default = ["gradients", "timestamp"]
+default = ["timestamp", "chili"]
 tracing = ["cellular_raza-core/tracing"]
 timestamp = ["cellular_raza-core/timestamp"]
 gradients = ["cellular_raza-concepts/gradients", "cellular_raza-core/gradients", "cellular_raza-building-blocks/gradients"]
