### 2025-01-24 7fb02de
restructure prelude in order to simplify usage of backends which are not chili
