### 2025-01-24 5483b09
adapt features and imports from examples
- this is only relevant for legacy examples using cpu_os_threads
