+++ cellular_raza-examples/cell_sorting/src/main-langevin.rs
@@ -1,8 +1,4 @@
-use cellular_raza::core::backend::chili::*;
-use cellular_raza::prelude::{
-    CalcError, CartesianCuboid3New, CellAgent, Interaction, Langevin3D, Mechanics, Position,
-    RngError, StorageBuilder, Velocity,
-};
+use cellular_raza::prelude::*;
 
 use nalgebra::Vector3;
 use rand::{Rng, SeedableRng};
