+++ cellular_raza-examples/ureter_signalling/src/main.rs
@@ -1,7 +1,5 @@
-use cellular_raza::building_blocks::cartesian_cuboid_n_old::*;
-use cellular_raza::building_blocks::*;
-use cellular_raza::core::backend::cpu_os_threads::*;
-use cellular_raza::core::storage::*;
+use cellular_raza::prelude::cartesian_cuboid_n_old::*;
+use cellular_raza::prelude::cpu_os_threads::*;
 
 use nalgebra::Vector2;
 
