+++ cellular_raza-examples/organoid_turing_growth/Cargo.toml
@@ -9,6 +9,6 @@ serde = { workspace = true, features=["rc"] }
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
 nalgebra = { version="0.33" }
-cellular_raza = { path="../../cellular_raza", features=["default"] }
+cellular_raza = { path="../../cellular_raza", default-features = false, features=["timestamp","cpu_os_threads"] }
 plotters = { workspace = true }
 num = { workspace = true }
