+++ cellular_raza-examples/sender-receiver/Cargo.toml
@@ -10,7 +10,7 @@ rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
 nalgebra = { version="0.33" }
 num = { workspace = true }
-cellular_raza = { path="../../cellular_raza", features=["default"] }
+cellular_raza = { path="../../cellular_raza", default-features = false, features=["gradients", "timestamp", "cpu_os_threads"] }
 plotters = { workspace = true, features = ["default", "ttf"] }
 ode_integrate = "0.0.2"
 rayon = "1.10.0"
