+++ cellular_raza-examples/bacteria_population/Cargo.toml
@@ -10,5 +10,5 @@ rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
 nalgebra = { version="0.33" }
 num = { workspace = true }
-cellular_raza = { path="../../cellular_raza", features=["default"] }
+cellular_raza = { path="../../cellular_raza", default-features = false, features=["timestamp", "cpu_os_threads"] }
 plotters = { workspace = true }
