### 2025-01-27 jonaspleyer/cellular_raza#2 @Armavica
Plotting of the cell_sorting example
Hi, as part of the review of your software/article that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/7723), I tried to run the first example myself.

I was eventually able to generate the video, but I had to adapt the file `cellular_raza-examples/cell_sorting/generate_movie.sh` that refers to a file `plot.py` which I believe should actually be `src/plot.py`?
