### 2025-01-27 jonaspleyer/cellular_raza#3 @Armavica
Documentation for running the examples
Hi, as part of the review that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/7723), I am trying to run the examples.

Even though I am relatively familiar with Rust and Cargo, it took me a little bit of time to understand how to run them. Have you considered adding some documentation (or READMEs) to explain the steps needed? Cloning the repository with submodules, moving to `cellular_raza/cellular_raza-examples/cell_sorting` (for example), running `cargo run --bin default --release`: this might be helpful for someone less familiar with git/Rust/Cargo.

Also, have you considered using Cargo's native support for examples? https://doc.rust-lang.org/cargo/reference/cargo-targets.html#examples
