### 2025-01-27 jonaspleyer/cellular_raza#2 @jonaspleyer
Thanks for reviewing cellular_raza. I hope that you saw the link to the hompage https://cellular-raza.com/ at which I give more details and explanations with in-depth guides on how to construct a first simulation from scratch.

You are referring to the file `cellular_raza-examples/cell_sorting/generate_movie.sh` which points to `plot.py` which itself was moved to `src/plot.py`. I will correct the path to `src/plot.py` and add an additional `README.md` to clear things up.
