### 2025-01-27 0b55a91
add readme describing the 3 different binaries which can be run
