### 2025-01-27 b0a96bf
add constructor methods for MiePotential and MorsePotential for pyo3 bindings
