+++ cellular_raza-examples/README.md
@@ -1,7 +1,9 @@
 # Examples
 
-These examples will be explained in more detail on the website [cellular_raza.com](https://cellular_raza.com).
-They do not only server as illustrative examples of what `cellular_raza` is capable of but also as a reliable method for continuous integration testing and compatibility.
+These examples will be explained in more detail on the website
+[cellular_raza.com](https://cellular_raza.com).
+They do not only server as illustrative examples of what `cellular_raza` is capable of but also as a
+reliable method for continuous integration testing and compatibility.
 
 # External
 ## Auophagy Project
