### 2025-01-27 cb67945
initialize variables; avoid tests failing
