### 2025-01-27 jonaspleyer/cellular_raza#3 @jonaspleyer
> Even though I am relatively familiar with Rust and Cargo, it took me a little bit of time to understand how to run them. Have you considered adding some documentation (or READMEs) to explain the steps needed?

This seems like a very good addition to the project. I will work on adding more `README.md` files as I have already done within the other issue for the `cell_sorting` example.

> Also, have you considered using Cargo's native support for examples? https://doc.rust-lang.org/cargo/reference/cargo-targets.html#examples

Most of the project is in line with general rust & cargo structure. But you are certainly right that adding examples explicitly would clear things up considerably.

Going further, it could be helpful to provide `requirements.txt` for the python plotting scripts.

- [ ] `README.md` files for examples
- [ ] explicit `[[examples]] ...` sections in `Cargo.toml`
- [ ] `requirements.txt` files for plotting scripts
