### 2025-01-27 jonaspleyer/cellular_raza#3 @jonaspleyer
That being said, I will probably still have some "old" examples lying around which I will keep in the main tree (not another repository) for (at least) the coming months to ensure any progress of the package conserves their functionality. However these examples might not be suited for end-users to directly follow and use. I will mention this in the `README.md` files which I will be adding.
