### 2025-01-27 34270c0
manually install libfontconfig files to avoid failing of CI
