+++ .github/workflows/reuse.yml
@@ -18,6 +18,9 @@ jobs:
           fetch-depth: 1
           submodules: 'recursive'
           lfs: 'false'
+      - name: Install libfontconfig1-dev
+        if: ${{ inputs.os == 'ubuntu-latest' }}
+        run: sudo apt-get install libfontconfig1-dev
 
       - name: Install toolchain
         uses: dtolnay/rust-toolchain@master
