### 2025-01-27 jonaspleyer/cellular_raza#2 @Armavica
Sounds good! Yes, I was indeed referring to this file. I did see the showcase on the website, but as this example is also provided in the repository that's how I tried to run it: it is what felt the most convenient to me on the moment. Either way is fine as long as it works!
Please don't include me on the changes, I am just reporting issues :)
