### 2025-01-27 jonaspleyer/cellular_raza#3 @Armavica
Sounds good! By the way, I should have mentioned it, but this is just a detail. The most important is that people can run your software, which is the case so far. What I am suggesting here is just what I think will make it easier for people to do so, but you may disagree on this point and leave out the `[[example]]`s if they would be difficult to add to your repository or if you think that the best way to run them is to follow the documentation of the showcases.
