### 2025-01-27 jonaspleyer/cellular_raza#3 @jonaspleyer
Thanks for clarifying that. I am happy to hear that. I will let you know how I decide. But overall it seems like a very good idea to me.
