### 2025-01-28 jonaspleyer/cellular_raza#4 @Armavica
Bacterial_branching example
Hi, as part of the review that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/7723), I am trying to run the examples shared in the article.

I first went to `cellular_raza-examples/bacterial_branching` and ran `cargo run --release`. Then, I ran `python src/plotting.py`.
The video generated is the following, which does not look like Figure 3 of the article.

https://github.com/user-attachments/assets/4b32b375-0e57-44d6-a415-2a69b5131844

I then tried to use the files listed on [this page](https://cellular-raza.com/showcase/bacterial-branching/) of the website. I used the following `Cargo.toml` file (copied from `bacterial_branching` example with just the name changed):

```toml
[package]
name = "cr_bacteria_branching"
version = "0.1.0"
edition = "2021"
authors = ["Jonas Pleyer <jonas.pleyer@fdm.uni-freiburg.de>"]

[dependencies]
serde = { workspace = true, features=["rc"] }
rand = { workspace = true, features=["small_rng"] }
rand_chacha = { workspace = true }
nalgebra = { version="0.33" }
num = { workspace = true }
cellular_raza = { path="../../cellular_raza", features=["default"] }
plotters = { workspace = true }
rayon = "1.10.0"
ndarray = { version = "0.16.0", features = ["blas", "serde", "serde-1"] }
```

 But upon running `cargo run --release`, I am getting the following error:

```
❯ cargo run --release                                                                           (cellular_raza)
   Compiling cr_bact_branching v0.1.0 (/home/virgile/src/cellular_raza/cellular_raza-examples/bact_branching)
error[E0432]: unresolved import `cellular_raza::prelude::cpu_os_threads`
 --> cellular_raza-examples/bact_branching/src/main.rs:3:29
  |
3 | use cellular_raza::prelude::cpu_os_threads::*;
  |                             ^^^^^^^^^^^^^^ could not find `cpu_os_threads` in `prelude`

For more information about this error, try `rustc --explain E0432`.
error: could not compile `cr_bact_branching` (bin "cr_bact_branching") due to 1 previous error
```

~~Is there something that I am doing wrong?~~

Edit: Ok, I figured it out: I had to enable the `cpu_os_threads` feature. I obtained the following final state:

![Image](https://github.com/user-attachments/assets/0b4db3c4-e7db-492e-a615-a1b62ef10f27)

It looks a bit more like what is shown in the article. I saw that you state [on this page](https://cellular-raza.com/showcase/bacterial-branching/) that you lost the parameters that generated the illustration picture. While this might be fine for the documentation, I think that it ought to be possible to recreate the figures that you use in the article. Have you considered using for Figure 3 an illustration coming from a known set of parameters?
