### 2025-01-29 jonaspleyer/cellular_raza#4 @jonaspleyer
> Edit: Ok, I figured it out: I had to enable the cpu_os_threads feature. I obtained the following final state:

I will include the Cargo.toml file on the website as well. You are right that without this file, the example is not complete.

> Have you considered using for Figure 3 an illustration coming from a known set of parameters?

I had considered that. One reason why I did not recreate the plot was that the total Runtime was very large. I will re-asses this issue and decide if to show a (possibly smaller) example with known parameters.

TODO
- [ ] Add `Cargo.toml` file in website
- [ ] Parameters and Reproducability of the Bacterial Branching Example
