### 2025-01-29 213a472
extend README.md file to give overview over all examples and how to execute them
