### 2025-01-29 037ce81
automatically import all examples as members
