+++ cellular_raza-core/Cargo.toml
@@ -28,7 +28,7 @@ num = { workspace = true }
 # ode_integrate = { version="0.0.2" }
 plotters = { workspace = true, optional = true }
 rayon = { version="1.6" }
-kdam = { version="0.6", features = ["template", "rayon"] }
+kdam = { workspace = true, features = ["template", "rayon"] }
 
 # Storage
 bincode = "1.3"
