+++ cellular_raza-benchmarks/Cargo.toml
@@ -22,7 +22,7 @@ serde = { workspace = true }
 num = { workspace = true }
 criterion = { version = "0.4", features = ["html_reports"] }
 clap = { version = "4.5.7", features = ["derive"] }
-kdam = "0.5.2"
+kdam.workspace = true
 serde_json = "1.0.118"
 glob = "0.3.1"
 tempdir.workspace = true
