+++ Cargo.toml
@@ -40,3 +40,4 @@ serde = "1.0"
 tempdir = "0.3"
 nalgebra = { version="<0.34, >=0.30", features=["serde-serialize", "rand"]}
 pyo3 = { version = "<0.24, >=0.22", features = ["serde", "py-clone"]}
+kdam = "0.6"
