### 2025-01-29 2763d0f
use workspace dependency for nalgebra
