+++ cellular_raza-examples/immersed_boundary/Cargo.toml
@@ -9,7 +9,7 @@ homepage.workspace =  true
 serde = { workspace = true }
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
-nalgebra = { version="0.33" }
+nalgebra = { workspace = true }
 cellular_raza = { path="../../cellular_raza", features=["default"] }
 num = { workspace = true }
 rayon = "1.8"
