+++ cellular_raza-examples/bacteria_population/Cargo.toml
@@ -8,7 +8,7 @@ authors = ["Jonas Pleyer <jonas.pleyer@fdm.uni-freiburg.de>"]
 serde = { workspace = true, features=["rc"] }
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
-nalgebra = { version="0.33" }
+nalgebra = { workspace = true }
 num = { workspace = true }
 cellular_raza = { path="../../cellular_raza", default-features = false, features=["timestamp", "cpu_os_threads"] }
 plotters = { workspace = true }
