+++ cellular_raza-benchmarks/Cargo.toml
@@ -17,7 +17,7 @@ homepage.workspace =  true
 [dependencies]
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true, features=["serde1"]}
-nalgebra = { version="0.33", features=["serde-serialize"]}
+nalgebra = { workspace = true, features=["serde-serialize"]}
 serde = { workspace = true }
 num = { workspace = true }
 criterion = { version = "0.4", features = ["html_reports"] }
