+++ cellular_raza-examples/pool_model_pyo3/Cargo.toml
@@ -13,7 +13,7 @@ pyo3 = { version = "0.22", features = ["extension-module"] }
 serde = { workspace = true, features=["rc"] }
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
-nalgebra = { version="0.33" }
+nalgebra = { workspace = true }
 num = { workspace = true }
 cellular_raza = { path="../../cellular_raza", default-features = false, features=["timestamp", "pyo3", "cpu_os_threads"] }
 plotters = { workspace = true }
