+++ cellular_raza-examples/diffusion/Cargo.toml
@@ -8,11 +8,11 @@ authors = ["Jonas Pleyer <jonas.pleyer@fdm.uni-freiburg.de>"]
 serde = { workspace = true }
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
-nalgebra = { version="0.33" }
+nalgebra = { workspace = true }
 cellular_raza = { path="../../cellular_raza", features=["default"] }
 num = { workspace = true }
 rayon = "1.8"
 kdam = { version="0.5", features = ["template"] }
-ndarray = { version = "0.15.6" }
+ndarray = { version = "0.16" }
 plotters = { workspace = true }
 tracing = "0.1.40"
