+++ cellular_raza-benchmarks/Cargo.toml
@@ -20,7 +20,7 @@ rand_chacha = { workspace = true, features=["serde1"]}
 nalgebra = { workspace = true, features=["serde-serialize"]}
 serde = { workspace = true }
 num = { workspace = true }
-criterion = { version = "0.4", features = ["html_reports"] }
+criterion = { version = "0.5", features = ["html_reports"] }
 clap = { version = "4.5.7", features = ["derive"] }
 kdam.workspace = true
 serde_json = "1.0.118"
