### 2025-01-29 6b525b2
add note for upgrading rand deps
