+++ Cargo.toml
@@ -34,6 +34,7 @@ plotters = { version="0.3", default-features = false, features=[
         "full_palette",
         "colormaps"
 ]}
+# We cannot upgrade these versions at the moment since nalgebra 0.33 still requires rand_distr 0.4
 rand = { version="0.8", features = ["small_rng"] }
 rand_chacha = { version="0.3.1", features=["serde1"] }
 serde = "1.0"
