### 2025-01-29 2760a53
make kdam workspace dependency
