+++ cellular_raza-examples/puzzle/Cargo.toml
@@ -12,7 +12,7 @@ nalgebra = { workspace = true }
 cellular_raza = { path="../../cellular_raza", features=["default"] }
 plotters = { workspace = true }
 num = { workspace = true }
-kdam = "0.5.2"
+kdam.workspace = true
 rayon = "1.10.0"
 itertools.workspace = true
 rand_distr = "0.4.3"
