### 2025-01-29 5ffc191
rename cell_sorting binaries
