+++ cellular_raza-examples/cell_sorting/Cargo.toml
@@ -13,13 +13,13 @@ cellular_raza = { path="../../cellular_raza" }
 num = { workspace = true }
 
 [[bin]]
-name = "default"
+name = "cr_cell_sorting_default"
 path = "src/main.rs"
 
 [[bin]]
-name = "langevin"
+name = "cr_cell_sorting_langevin"
 path = "src/main-langevin.rs"
 
 [[bin]]
-name = "brownian"
+name = "cr_cell_sorting_brownian"
 path = "src/main-brownian.rs"
