+++ cellular_raza-examples/puzzle/Cargo.toml
@@ -16,4 +16,4 @@ kdam.workspace = true
 rayon = "1.10.0"
 itertools.workspace = true
 rand_distr.workspace = true
-geo = { version = "0.28.0", features = ["earcutr"] }
+geo = { version = "0.29", features = ["earcutr"] }
