### 2025-01-29 jonaspleyer/cellular_raza#3 @jonaspleyer
Concerning the `[[example]] ...` sections:
I came across this thread https://users.rust-lang.org/t/how-to-include-example-projects-in-workspaces/122375/4 which also suggests to simply include examples as individual packages when incorporating them inside a workspace.

To me, it is rather important to have individual `Cargo.toml` files for every example since they should be as much self-contained as possible. I have previously seen in other projects that this was not the case and would like to avoid that.

But I have added a `README.md` file listing all examples and how to run them. See https://github.com/jonaspleyer/cellular_raza/tree/master/cellular_raza-examples
