### 2025-01-29 c223f04
correct names of binaries to run them
