+++ cellular_raza-examples/ureter_signalling/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cr_example-ureter-signalling"
+name = "cr_ureter_signalling"
 version = "0.0.3"
 edition = "2021"
 authors = ["Jonas Pleyer <jonas.pleyer@fdm.uni-freiburg.de>"]
