### 2025-01-29 c64b549
add README.md files for every example
