### 2025-01-29 f4e2b6d
correct bin names to execute
