+++ cellular_raza-examples/ureter_signalling/README.md
@@ -3,7 +3,7 @@
 To run the simulation from the base folder of the `cellular_raza` workspace use
 
 ```bash
-cargo run -r --bin ureter_signalling
+cargo run -r --bin cr_ureter_signalling
 ```
 
 or from the folder of this example `cellular_raza-examples/ureter_signalling`
