### 2025-01-30 jonaspleyer/cellular_raza#3 @Armavica
That's a very good point! I think that's great how it is now.
