### 2025-01-30 6ac8de5
add requirements.txt files for all subprojects that are using python plotting scripts
