+++ cellular_raza-examples/puzzle/requirements.txt
@@ -1,0 +1,2 @@
+matplotlib==3.10.0
+numpy==2.2.2
