+++ cellular_raza-examples/bacterial_branching/requirements.txt
@@ -1,0 +1,4 @@
+matplotlib==3.10.0
+numpy==2.2.2
+pandas==2.2.3
+tqdm==4.67.1
