+++ cellular_raza-examples/bacterial_rods/requirements.txt
@@ -1,0 +1,5 @@
+pyvista==0.43.8
+numpy==1.26.4
+argparse==1.4.0
+tqdm==4.66.4
+pandas==2.2.2
