+++ cellular_raza-examples/getting-started/requirements.txt
@@ -1,0 +1,3 @@
+matplotlib==3.10.0
+numpy==2.2.2
+tqdm==4.67.1
