+++ cellular_raza-examples/puzzle/plot.py
@@ -6,14 +6,17 @@ from glob import glob
 from pathlib import Path
 import os
 
+
 def get_last_save_path(opath: Path | None = None) -> Path:
     if opath is None:
         opath = Path("out/puzzles")
     results = sorted(glob(str(opath) + "/*"))
     return Path(results[-1])
 
+
 def get_iterations(opath: Path = get_last_save_path()) -> list[int]:
     return [int(os.path.basename(x)) for x in glob(str(opath) + "/cells/json/*")]
+
 
 def load_cells(iteration: int, opath: Path = get_last_save_path()) -> list[dict] | None:
     batches = glob(str(opath) + "/cells/json/{:020}/*.json".format(iteration))
@@ -28,9 +31,11 @@ def load_cells(iteration: int, opath: Path = get_last_save_path()) -> list[dict]
                 cells.append(d["element"][0]["cell"])
         return cells
 
+
 def extract_vertices(cells: list[dict]):
     for cell in cells:
         yield np.array(cell["mechanics"]["puzzle"]["vertices"]["Value"])
+
 
 def extract_triangulations(cells: list[dict]):
     for cell in cells:
@@ -39,20 +44,24 @@ def extract_triangulations(cells: list[dict]):
         for triangle in triangles:
             yield vertices[triangle]
 
+
 def get_domain_size(opath: Path = get_last_save_path()):
     d = glob(glob(str(opath) + "/subdomains/json/*")[0] + "/*")[0]
     f = open(d)
-    subdomain =  json.load(f)["element"]["subdomain"]
+    subdomain = json.load(f)["element"]["subdomain"]
     domain_min = subdomain["domain_min"]
     domain_max = subdomain["domain_max"]
     return domain_min, domain_max
+
 
 class Plotter:
     def __init__(self, opath: Path | str = get_last_save_path()):
         self.opath = Path(opath)
         self.domain_min, self.domain_max = get_domain_size(self.opath)
-        s = (self.domain_max[1] - self.domain_min[1]) / (self.domain_max[0] - self.domain_min[0])
-        self.fig, self.ax = plt.subplots(figsize=(12, s*12))
+        s = (self.domain_max[1] - self.domain_min[1]) / (
+            self.domain_max[0] - self.domain_min[0]
+        )
+        self.fig, self.ax = plt.subplots(figsize=(12, s * 12))
         self.ax.set_xlim((self.domain_min[0], self.domain_max[0]))
         self.ax.set_ylim((self.domain_min[1], self.domain_max[1]))
         self.fig.tight_layout()
@@ -61,12 +70,15 @@ class Plotter:
     def plot_cells_at_iteration(self, iteration: int, transparent: bool = True):
         cells = load_cells(iteration, self.opath)
         vertices = extract_vertices(cells)
-        pc1 = matplotlib.collections.PatchCollection([
-            matplotlib.patches.Polygon(
-                cell_vertices,
-                fill=True,
-                closed=True,
-            ) for cell_vertices in vertices],
+        pc1 = matplotlib.collections.PatchCollection(
+            [
+                matplotlib.patches.Polygon(
+                    cell_vertices,
+                    fill=True,
+                    closed=True,
+                )
+                for cell_vertices in vertices
+            ],
             alpha=0.7,
             facecolor="green",
             edgecolor="black",
@@ -74,12 +86,15 @@ class Plotter:
         pc1 = self.ax.add_collection(pc1)
         triangles = extract_triangulations(cells)
         pc2 = matplotlib.collections.PatchCollection(
-            [matplotlib.patches.Polygon(
-                triangle,
-                fill=False,
-                closed=True,
-            ) for triangle in triangles],
-            facecolors='none',
+            [
+                matplotlib.patches.Polygon(
[…]
