### 2025-01-31 8d7ebb4
add getters and setters for mie potential
