### 2025-02-01 jonaspleyer/cellular_raza#5 @finsberg
[JOSS REVIEW] Add installation instructions
The documentation for `cellular_raza` looks very thorough with many examples. However, I do miss a section about Installation instructions, which I think belongs in the README file.
While it might be obvious for a Rust developer how to do this using `crates.io` and the [Getting started guide](https://cellular-raza.com/guides/getting-started/), I think it is worth having a dedicated section about this and also provide instructions on how to install rust / cargo.

This issue is related to https://github.com/openjournals/joss-reviews/issues/7723
