### 2025-02-01 jonaspleyer/cellular_raza#5 @jonaspleyer
Hi @finsberg and thanks for the feedback on my paper.

> The documentation for cellular_raza looks very thorough with many examples.

Thanks this was quite a bit of work. 😃

> However, I do miss a section about Installation instructions, which I think belongs in the README file.
> While it might be obvious for a Rust developer how to do this using crates.io and the [Getting started guide](https://cellular-raza.com/guides/getting-started/), I think it is worth having a dedicated section about this and also provide instructions on how to install rust / cargo.

I think that there are 2 main statements here as far as I understood your point.

1. Installing `cargo` and `rustc` (possibly via the `rustup` toolchain)
I think that it is now wise to add specific instructions for how to install `cargo` and `rustc`. As with any major tool it is much better to directly refer to the website of the supplier (in this case probably https://www.rust-lang.org/tools/install. I can add a link and statement in my "getting-started" guide to cover this problem. Furthermore I would prefer to directly link to the "getting-started" guide from the main `README.md` file. This way my documentation is not split between: `README.md` files in my github repository and the documentation on the https://celular-raza.com website.

2. Installing `cellular_raza`
I am not quite sure how to understand this problem. Since `cellular_raza` is a library, installing it, only means to acquire the source code and then use said code within your project. There is no binary version to be installed in the system. This task is being done by the package manager `cargo` and is the recommended way of doing it. While it is also possible to only use `rustc` without relying on `cargo` it is not recommended and probably very cumbersome to set up. The command which adds `cellular_raza` to a new cargo project is simply
```bash
cargo add cellular_raza
```
as given in the "getting-started" guide. In this guide, we also add the dependencies `serde rand rand_chacha num`. When the `cargo build` is invoked, the source code is obtained, compiled and linked into a binary form.

- [ ] Add links to https://www.rust-lang.org/tools/install in "getting-started" guide
- [ ] Add link from README.md to "getting-started" guide ???

Would the TODOs above satisfy your concerns?
