### 2025-02-02 462e681
add link to getting-started guide in README.md for the crate and workspace
