### 2025-02-02 f8c930e
include logos for package release
