### 2025-02-02 jonaspleyer/cellular_raza#5 @jonaspleyer
Marking this as resolved due to:
1. [a21d8f8a2cc00edd681d61fa274ff4bdc2b2543b](https://github.com/jonaspleyer/cellular_raza-homepage/commit/a21d8f8a2cc00edd681d61fa274ff4bdc2b2543b)
2. f8c930eac8604abb35cf8ce1ea3375cf40ac14df
