### 2025-02-06 #7723 @jonaspleyer
Hi everyone. I have a small question: Since the review is open to the public, I was considering to post a small announcement on the [users.rust.lang](https://users.rust-lang.org/) forum and possibly in other channels. What is your and/or JOSS's stance on this?
