### 2025-02-08 jonaspleyer/cellular_raza#6 @finsberg
[JOSS REVIEW] Unable to run "Getting started" example
I am trying to run the [Getting started](https://cellular-raza.com/guides/getting-started/) example.

This works if I clone the repository and run the example from there. However, if I follow the steps in the documentation I get the following error when trying to run the example

```
cargo run -r
   Compiling autocfg v1.4.0
   Compiling proc-macro2 v1.0.93
   Compiling unicode-ident v1.0.16
   Compiling libc v0.2.169
   Compiling cfg-if v1.0.0
   Compiling serde v1.0.217
   Compiling libm v0.2.11
   Compiling byteorder v1.5.0
   Compiling rand_core v0.4.2
   Compiling crossbeam-utils v0.8.21
   Compiling simd-adler32 v0.3.7
   Compiling bytemuck v1.21.0
   Compiling autocfg v0.1.8
   Compiling crc32fast v1.4.2
   Compiling adler2 v2.0.0
   Compiling rand_core v0.3.1
   Compiling miniz_oxide v0.8.3
   Compiling fdeflate v0.3.7
   Compiling semver-parser v0.7.0
   Compiling num-traits v0.2.19
   Compiling bitflags v1.3.2
   Compiling color_quant v1.1.0
   Compiling semver v0.9.0
   Compiling flate2 v1.0.35
   Compiling rand_pcg v0.1.2
   Compiling crossbeam-epoch v0.9.18
   Compiling rand_chacha v0.1.1
   Compiling png v0.17.16
   Compiling maybe-uninit v2.0.0
   Compiling weezl v0.1.8
   Compiling core-foundation-sys v0.8.7
   Compiling quote v1.0.38
   Compiling rayon-core v1.12.1
   Compiling jpeg-decoder v0.3.1
   Compiling syn v2.0.98
   Compiling plotters-backend v0.3.7
   Compiling gif v0.12.0
   Compiling iana-time-zone v0.1.61
   Compiling crossbeam-deque v0.8.6
   Compiling getrandom v0.2.15
   Compiling num-integer v0.1.46
   Compiling image v0.24.9
   Compiling rustc_version v0.2.3
   Compiling num-bigint v0.4.6
   Compiling rand v0.6.5
   Compiling lock_api v0.4.12
   Compiling rustix v0.38.44
   Compiling either v1.13.0
   Compiling parking_lot_core v0.8.6
   Compiling paste v1.0.15
   Compiling parking_lot_core v0.4.0
   Compiling num-iter v0.1.45
   Compiling rand_os v0.1.3
   Compiling errno v0.3.10
   Compiling rand_jitter v0.1.4
   Compiling num-rational v0.4.2
   Compiling chrono v0.4.39
   Compiling plotters-svg v0.3.7
   Compiling plotters-bitmap v0.3.7
   Compiling matrixmultiply v0.3.9
   Compiling safe_arch v0.7.4
   Compiling rand_hc v0.1.0
   Compiling rand_isaac v0.1.1
   Compiling rand_xorshift v0.1.1
   Compiling instant v0.1.13
   Compiling getrandom v0.3.1
   Compiling scopeguard v1.2.0
   Compiling smallvec v1.13.2
   Compiling zerocopy v0.8.17
   Compiling typenum v1.17.0
   Compiling plotters v0.3.7
   Compiling wide v0.7.32
   Compiling rayon v1.10.0
   Compiling smallvec v0.6.14
   Compiling approx v0.5.1
   Compiling rawpointer v0.2.1
   Compiling memchr v2.7.4
   Compiling serde_json v1.0.138
   Compiling itertools v0.13.0
   Compiling serde_derive v1.0.217
   Compiling zerocopy-derive v0.7.35
   Compiling cellular_raza-concepts-derive v0.1.6
   Compiling nalgebra-macros v0.2.2
   Compiling zerocopy v0.7.35
   Compiling parking_lot v0.11.2
   Compiling fs2 v0.4.3
   Compiling fxhash v0.2.1
   Compiling ppv-lite86 v0.2.20
   Compiling base64 v0.21.7
   Compiling itoa v1.0.14
   Compiling formatx v0.2.3
   Compiling ryu v1.0.19
   Compiling log v0.4.25
   Compiling sled v0.34.7
   Compiling cellular_raza-core-proc-macro v0.1.6
   Compiling hurdles v1.0.1
   Compiling crossbeam-channel v0.5.14
   Compiling rand_core v0.9.0
   Compiling rand_chacha v0.9.0
   Compiling rand v0.9.0
   Compiling rand_core v0.6.4
   Compiling rand_chacha v0.3.1
   Compiling num-complex v0.4.6
   Compiling bitflags v2.8.0
   Compiling rand v0.8.5
   Compiling ron v0.8.1
   Compiling circ_buffer v0.1.9
   Compiling num v0.4.3
   Compiling cellular_raza-concepts v0.1.6
   Compiling simba v0.9.0
   Compiling quick-xml v0.36.2
   Compiling uniquevec v0.1.0
   Compiling bincode v1.3.3
   Compiling terminal_size v0.3.0
   Compiling kdam v0.5.2
   Compiling rand_distr v0.4.3
   Compiling cellular_raza-core v0.1.6
   Compiling nalgebra v0.33.2
   Compiling cellular_raza-building-blocks v0.1.6
   Compiling cellular_raza v0.1.6
   Compiling my-new-project v0.1.0 (/Users/finsberg/local/src/cellular_raza_sandbox/my-new-project)
error[E0053]: method `get_random_contribution` has an incompatible type for trait
 --> src/main.rs:7:10
  |
7 | #[derive(CellAgent, Clone, Deserialize, Serialize)]
  |          ^^^^^^^^^ expected `rand_chacha::chacha::ChaCha8Rng`, found `ChaCha8Rng`
  |
  = note: expected signature `fn(&Agent, &mut rand_chacha::chacha::ChaCha8Rng, _) -> Result<_, _>`
             found signature `fn(&Agent, &mut ChaCha8Rng, _) -> Result<_, _>`
  = note: this error originates in the derive macro `CellAgent` (in Nightly builds, run with -Z macro-backtrace for more info)

error[E0308]: mismatched types
   --> src/main.rs:7:10
    |
7   | #[derive(CellAgent, Clone, Deserialize, Serialize)]
    |          ^^^^^^^^^
    |          |
    |          expected `rand_chacha::chacha::ChaCha8Rng`, found `ChaCha8Rng`
    |          arguments to this function are incorrect
    |
    = note: `ChaCha8Rng` and `rand_chacha::chacha::ChaCha8Rng` have similar names, but are actually distinct types
note: `ChaCha8Rng` is defined in crate `rand_chacha`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/rand_chacha-0.9.0/src/chacha.rs:346:1
    |
346 | / chacha_impl!(
347 | |     ChaCha8Core,
348 | |     ChaCha8Rng,
349 | |     4,
350 | |     "ChaCha with 8 rounds",
351 | |     abstract8,
352 | | );
    | |_^
note: `rand_chacha::chacha::ChaCha8Rng` is defined in crate `rand_chacha`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/rand_chacha-0.3.1/src/chacha.rs:345:1
    |
345 | chacha_impl!(ChaCha8Core, ChaCha8Rng, 4, "ChaCha with 8 rounds", abstract8);
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: perhaps two different versions of crate `rand_chacha` are being used?
note: method defined here
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/cellular_raza-concepts-0.1.6/src/mechanics.rs:26:8
    |
26  |     fn get_random_contribution(
    |        ^^^^^^^^^^^^^^^^^^^^^^^
    = note: this error originates in the derive macro `CellAgent` which comes from the expansion of the macro `chacha_impl` (in Nightly builds, run with -Z macro-backtrace for more info)

warning: use of deprecated method `rand::Rng::gen_range`: Renamed to `random_range`
  --> src/main.rs:62:17
   |
62 |             rng.gen_range(0.0..parameters.domain_size),
   |                 ^^^^^^^^^
   |
   = note: `#[warn(deprecated)]` on by default

warning: use of deprecated method `rand::Rng::gen_range`: Renamed to `random_range`
  --> src/main.rs:63:17
   |
63 |             rng.gen_range(0.0..parameters.domain_size),
   |                 ^^^^^^^^^

error[E0308]: mismatched types
   --> src/main.rs:95:5
    |
95  | /     run_simulation!(
96  | |         domain: domain,
97  | |         agents: cells,
98  | |         settings: settings,
99  | |         aspects: [Mechanics, Interaction]
100 | |     )?;
    | |     ^
    | |     |
    | |_____expected `rand_chacha::chacha::ChaCha8Rng`, found `ChaCha8Rng`
    |       arguments to this function are incorrect
    |
    = note: `ChaCha8Rng` and `rand_chacha::chacha::ChaCha8Rng` have similar names, but are actually distinct types
note: `ChaCha8Rng` is defined in crate `rand_chacha`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/rand_chacha-0.9.0/src/chacha.rs:346:1
    |
346 | / chacha_impl!(
347 | |     ChaCha8Core,
348 | |     ChaCha8Rng,
349 | |     4,
350 | |     "ChaCha with 8 rounds",
351 | |     abstract8,
352 | | );
    | |_^
note: `rand_chacha::chacha::ChaCha8Rng` is defined in crate `rand_chacha`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/rand_chacha-0.3.1/src/chacha.rs:345:1
    |
345 | chacha_impl!(ChaCha8Core, ChaCha8Rng, 4, "ChaCha with 8 rounds", abstract8);
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: perhaps two different versions of crate `rand_chacha` are being used?
note: function defined here
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/cellular_raza-core-0.1.6/src/backend/chili/update_mechanics.rs:531:8
    |
531 | pub fn local_mechanics_update<
    |        ^^^^^^^^^^^^^^^^^^^^^^
    = note: this error originates in the macro `run_simulation` which comes from the expansion of the macro `chacha_impl` (in Nightly builds, run with -Z macro-backtrace for more info)

error[E0308]: mismatched types
   --> src/main.rs:95:5
    |
95  | /     run_simulation!(
96  | |         domain: domain,
97  | |         agents: cells,
98  | |         settings: settings,
99  | |         aspects: [Mechanics, Interaction]
100 | |     )?;
    | |     ^
    | |     |
    | |_____expected `rand_chacha::chacha::ChaCha8Rng`, found `ChaCha8Rng`
    |       arguments to this function are incorrect
    |
    = note: `ChaCha8Rng` and `rand_chacha::chacha::ChaCha8Rng` have similar names, but are actually distinct types
note: `ChaCha8Rng` is defined in crate `rand_chacha`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/rand_chacha-0.9.0/src/chacha.rs:346:1
    |
346 | / chacha_impl!(
347 | |     ChaCha8Core,
348 | |     ChaCha8Rng,
349 | |     4,
350 | |     "ChaCha with 8 rounds",
351 | |     abstract8,
352 | | );
    | |_^
note: `rand_chacha::chacha::ChaCha8Rng` is defined in crate `rand_chacha`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/rand_chacha-0.3.1/src/chacha.rs:345:1
    |
345 | chacha_impl!(ChaCha8Core, ChaCha8Rng, 4, "ChaCha with 8 rounds", abstract8);
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: perhaps two different versions of crate `rand_chacha` are being used?
note: function defined here
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/cellular_raza-core-0.1.6/src/backend/chili/update_mechanics.rs:563:8
    |
563 | pub fn local_interaction_react_to_neighbors<C, A, Pos, Vel, For, Inf, Float>(
    |        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: this error originates in the macro `run_simulation` which comes from the expansion of the macro `chacha_impl` (in Nightly builds, run with -Z macro-backtrace for more info)

error[E0631]: type mismatch in closure arguments
   --> src/main.rs:95:5
    |
95  | /     run_simulation!(
96  | |         domain: domain,
97  | |         agents: cells,
98  | |         settings: settings,
99  | |         aspects: [Mechanics, Interaction]
100 | |     )?;
    | |     ^
    | |     |
    | |_____expected due to this
    |       found signature defined here
    |
    = note: expected closure signature `for<'a, 'b, 'c> fn(&'a mut Agent, &'b mut _CrAuxStorage<_, _, _, _>, _, &'c mut rand_chacha::chacha::ChaCha8Rng) -> _`
               found closure signature `fn(&mut Agent, &mut _CrAuxStorage<_, _, _, _>, _, &mut ChaCha8Rng) -> _`
note: required by a bound in `SubDomainBox::<I, S, C, A, Com, Sy>::run_local_cell_funcs`
   --> /Users/finsberg/.cargo/registry/src/index.crates.io-6f17d22bba15001f/cellular_raza-core-0.1.6/src/backend/chili/datastructures.rs:307:14
    |
296 |     pub fn run_local_cell_funcs<Func, F>(
    |            -------------------- required by a bound in this associated function
...
307 |         ) -> Result<(), super::SimulationError>,
    |              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `SubDomainBox::<I, S, C, A, Com, Sy>::run_local_cell_funcs`
    = note: this error originates in the macro `run_simulation` (in Nightly builds, run with -Z macro-backtrace for more info)

Some errors have detailed explanations: E0053, E0308, E0631.
For more information about an error, try `rustc --explain E0053`.
warning: `my-new-project` (bin "my-new-project") generated 2 warnings
error: could not compile `my-new-project` (bin "my-new-project") due to 5 previous errors; 2 warnings emitted
```

Maybe there is just a mismatch between the version I get when I clone the repo and the version I get when adding it from `crates.io`.

Here is also some spec which might be useful
```
$ cargo --version
cargo 1.84.1 (66221abde 2024-11-19)
```
```
$ rustc --version
rustc 1.84.1 (e71f9a9a9 2025-01-27)
```
```
OS:  Apple Silicon M1 Max, MacOSX Sequioa Version 15.3
```
This issue is related to https://github.com/openjournals/joss-reviews/issues/7723
