### 2025-02-10 jonaspleyer/cellular_raza#6 @jonaspleyer
I have also opened a new PR (https://github.com/dimforge/nalgebra/pull/1484) to increase the version number of `rand` within `nalgebra`.
