### 2025-02-10 88ebf81
implement __getitem__ method for CellIdentifier
