### 2025-02-11 15f5738
Update ndarray requirement from 0.15.6 to 0.16.1
---
updated-dependencies:
- dependency-name: ndarray
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
