### 2025-02-11 jonaspleyer/cellular_raza#6 @jonaspleyer
*Update*
I also added a dependabot which automatically checks for newest versions of dependencies. It has already been used. See #8
