### 2025-02-14 3d14832
remove duplicate field in serialization
