### 2025-02-17 jonaspleyer/cellular_raza#9 @Armavica
Community guidelines
Hi, as part of my review of your software and documentation for JOSS (https://github.com/openjournals/joss-reviews/issues/7723) I have to look at the following point: "Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support".

I found invitations [for contributions](https://cellular-raza.com/internals/code-structure/coding-guidelines/), [for issues](https://cellular-raza.com/internals/code-structure/todo/) and [for feature requests](https://cellular-raza.com/internals/code-structure/roadmap/). I believe that these address points 1. and 2.
However, these pages are pretty deep into the documentation website (perhaps I missed them in more obvious places).

I don't think that there are strong JOSS rules concerning where these guidelines have to be placed, but I was wondering if you have considered **adding quick links for contributions, issues and support in more visible places**? Something that could be of interest is a `CONTRIBUTING.md` file: https://contributing.md/how-to-build-contributing-md/

Moreover, if I am interpreting the JOSS guideline correctly I believe that **there should be a place where people can feel free to ask questions and request support** in using the software (not only bug reports and feature requests). This could be the GitHub issue page, or this could also be a public forum or even an email address.
