### 2025-02-17 jonaspleyer/cellular_raza#10 @Armavica
References for software dependencies
Hi, as part of the review of your article that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/7723), I have to check that "Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?"

I think that it would be appropriate to cite the direct dependencies of `cellular_raza`. References and DOIs for software don't necessarily exist, but sometimes the desired format is provided by the developers (for example in a `CITATION.cff` or `CITATION.bib` file at the root of the repository). When nothing is provided, a link to the forge where the software is developed, or to its documentation, (or crate2bib ;) can be used.
