### 2025-02-18 2c9f83f
add new file for guidelines on contributing to this package
