### 2025-02-18 jonaspleyer/cellular_raza#9 @jonaspleyer
>  However, these pages are pretty deep into the documentation website (perhaps I missed them in more obvious places).

I have added a `CONTRIBUTING.md` file for which I used the template provided by Github.
https://github.com/jessesquires/.github/blob/main/CONTRIBUTING.md
See commit 2c9f83f47f1ac3d3a46c564ef407c06d21d96db0.

> Moreover, if I am interpreting the JOSS guideline correctly I believe that there should be a place where people can feel free to ask questions and request support in using the software (not only bug reports and feature requests). This could be the GitHub issue page, or this could also be a public forum or even an email address.

I have activated the discussions forum provided by Github. This should hopefully solve this problem.
