### 2025-02-18 jonaspleyer/cellular_raza#9 @Armavica
That's great! Although the `CONTRIBUTING.md` file is great for developers, I think that for regular users who might be less used to github practices and just read the documentation website, it could be good to mention the discussion forum there as well. But feel free to do this or not - as I understand, your software is more of a library targeted to developers than a self-standing software aimed at the general public, so that might be less relevant in your case. I'm closing this issue because I think this point has been addressed satisfactorily.
