### 2025-02-18 jonaspleyer/cellular_raza#10 @jonaspleyer
As far as I can understand, this would be related to the paper rather than the software. Am I correct?

I personally think that this bullet point does not point towards a complete list of dependencies. For example the publication of [`faer`](https://github.com/sarah-quinones/faer-rs) which is being wrapped up at the moment does also not follow this approach.

The paper does not contain any of the [dependencies](https://github.com/sarah-quinones/faer-rs/blob/main/faer/Cargo.toml) of `faer`.

Also the other bullet point from the [Documentation](https://github.com/openjournals/joss-reviews/issues/7723#issuecomment-2614624857) section concerning installation should cover listing of dependencies imo.

>  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

# Dependencies
All dependencies of `cellular_raza` are given by the following list.
The number of dependencies is 23. This would mean, that I would need to add 23 additional citations to the paper which might be removed in the future if the software changes and the dependency is not required anymore.

```
nalgebra v0.33.2
num v0.4.3
ode_integrate v0.0.2
rand v0.8.5
rand_chacha v0.3.1
serde v1.0.217
tempdir v0.3.7
clap v4.5.27
criterion v0.5.1
glob v0.3.2
kdam v0.6.2
serde_json v1.0.138
proc-macro2 v1.0.93
quote v1.0.38
syn v2.0.96
itertools v0.14.0
rayon v1.10.0
pyo3 v0.22.6
ndarray v0.16.1
plotters v0.3.7
rand_distr v0.4.3
tracing v0.1.41
geo v0.29.3
```

I checked these repositories and found explicit citation entries for the following crates:
## clap
The first is a CLI parsing library used for benchmarks and not included in the distributed library.
```
@software{Knapp_clap_2025,
    author = {Knapp, Kevin B. and {The Clap Community}},
    license = {["Apache-2.0", "MIT"]},
    month = feb,
    title = {{clap}},
    url = {https://github.com/clap-rs/clap},
    version = {4.5.30},
    year = {2025}
}
```

## pyo3
The `pyo3` dependency handles generation of python bindings and is an optional dependency but frequently used. I will definitely add this citation since I am also referring to it in the text.
```
@software{PyO3_Project_and_Contributors_PyO3,
    author = {{PyO3 Project and Contributors}},
    license = {["Apache-2.0", "MIT"]},
    title = {{PyO3}}
}
```

# Dependency Graphs
The dependency graphs for dependencies of depth 1 look like this at the moment:
## cellular_raza-core
```
cellular_raza-core v0.1.6
├── bincode v1.3.3
├── cellular_raza-concepts v0.1.6
├── cellular_raza-core-proc-macro v0.1.6 (proc-macro)
├── chrono v0.4.39
├── circ_buffer v0.1.9
├── crossbeam-channel v0.5.14
├── hurdles v1.0.1
├── itertools v0.14.0
├── kdam v0.6.2
├── num v0.4.3
├── quick-xml v0.37.2
├── rand v0.8.5
├── rand_chacha v0.3.1
├── rayon v1.10.0
├── ron v0.8.1
├── serde v1.0.217
├── serde_json v1.0.138
├── sled v0.34.7
└── uniquevec v0.1.0
```

## cellular_raza-core-proc-macro
```
cellular_raza-core-proc-macro v0.1.6 (proc-macro)
├── itertools v0.14.0
├── proc-macro2 v1.0.93
├── quote v1.0.38
└── syn v2.0.96
```

## cellular_raza-concepts
```
cellular_raza-concepts v0.1.6
├── cellular_raza-concepts-derive v0.1.6 (proc-macro)
├── num v0.4.3
├── plotters v0.3.7
├── rand_chacha v0.3.1
└── serde v1.0.217
```

## cellular_raza-concepts-derive
```
cellular_raza-concepts-derive v0.1.6 (proc-macro)
├── proc-macro2 v1.0.93
├── quote v1.0.38
└── syn v2.0.96
```

## cellular_raza-building-blocks
```
cellular_raza-building-blocks v0.1.6
├── cellular_raza-concepts v0.1.6
├── itertools v0.14.0
├── nalgebra v0.33.2
├── num v0.4.3
├── plotters v0.3.7
├── rand v0.8.5
├── rand_chacha v0.3.1
├── rand_distr v0.4.3
├── rayon v1.10.0
└── serde v1.0.217
```
