### 2025-02-18 c0d227e
add citations for pyo3 and maturin crates and mention documentation website in text
