+++ paper.bib
@@ -356,3 +356,19 @@
     month=jan,
     language={en},
 }
+
+@software{PyO3_Project_and_Contributors_PyO3,
+    author = {{PyO3 Project and Contributors}},
+    license = {["Apache-2.0", "MIT"]},
+    title = {{PyO3}},
+}
+
+@software{maturin2025,
+    author = {konsti},
+    title = {{maturin}: Build and publish crates with pyo3, cffi and uniffi
+             bindings as well as rust binaries as python packages},
+    url = {https://github.com/pyo3/maturin},
+    date = {2025-02-05},
+    version = {1.8.2},
+    license = {MIT OR Apache-2.0},
+}
