+++ paper.bib
@@ -1,293 +1,329 @@
 @article{Pleyer2023,
-  title = {Agent-based models in cellular systems},
-  volume = {10},
-  ISSN = {2296-424X},
-  url = {http://dx.doi.org/10.3389/fphy.2022.968409},
-  DOI = {10.3389/fphy.2022.968409},
-  journal = {Frontiers in Physics},
-  publisher = {Frontiers Media SA},
-  author = {Pleyer,  Jonas and Fleck,  Christian},
-  year = {2023},
-  month = jan 
+    title = {Agent-based models in cellular systems},
+    volume = {10},
+    ISSN = {2296-424X},
+    url = {http://dx.doi.org/10.3389/fphy.2022.968409},
+    DOI = {10.3389/fphy.2022.968409},
+    journal = {Frontiers in Physics},
+    publisher = {Frontiers Media SA},
+    author = {Pleyer, Jonas and Fleck, Christian},
+    year = {2023},
+    month = jan,
 }
 
 @article{Abar2017,
-  title = {Agent Based Modelling and Simulation tools: A review of the state-of-art software},
-  volume = {24},
-  ISSN = {1574-0137},
-  url = {http://dx.doi.org/10.1016/j.cosrev.2017.03.001},
-  DOI = {10.1016/j.cosrev.2017.03.001},
-  journal = {Computer Science Review},
-  publisher = {Elsevier BV},
-  author = {Abar,  Sameera and Theodoropoulos,  Georgios K. and Lemarinier,  Pierre and O’Hare,  Gregory M.P.},
-  year = {2017},
-  month = may,
-  pages = {13–33}
+    title = {Agent Based Modelling and Simulation tools: A review of the
+             state-of-art software},
+    volume = {24},
+    ISSN = {1574-0137},
+    url = {http://dx.doi.org/10.1016/j.cosrev.2017.03.001},
+    DOI = {10.1016/j.cosrev.2017.03.001},
+    journal = {Computer Science Review},
+    publisher = {Elsevier BV},
+    author = {Abar, Sameera and Theodoropoulos, Georgios K. and Lemarinier,
+              Pierre and O’Hare, Gregory M.P.},
+    year = {2017},
+    month = may,
+    pages = {13–33},
 }
 
 @article{Rodgers1985,
-  title = {Improvements in multiprocessor system design},
-  volume = {13},
-  ISSN = {0163-5964},
-  url = {http://dx.doi.org/10.1145/327070.327215},
-  DOI = {10.1145/327070.327215},
-  number = {3},
-  journal = {ACM SIGARCH Computer Architecture News},
-  publisher = {Association for Computing Machinery (ACM)},
-  author = {Rodgers,  David P.},
-  year = {1985},
-  month = jun,
-  pages = {225–231}
+    title = {Improvements in multiprocessor system design},
+    volume = {13},
+    ISSN = {0163-5964},
+    url = {http://dx.doi.org/10.1145/327070.327215},
+    DOI = {10.1145/327070.327215},
+    number = {3},
+    journal = {ACM SIGARCH Computer Architecture News},
+    publisher = {Association for Computing Machinery (ACM)},
+    author = {Rodgers, David P.},
+    year = {1985},
+    month = jun,
+    pages = {225–231},
 }
 
 @article{kawasakiModelingSpatioTemporalPatterns1997,
-  title = {Modeling {{Spatio-Temporal Patterns Generated byBacillus}} Subtilis},
-  author = {Kawasaki, K. and Mochizuki, A. and Matsushita, M. and Umeda, T. and Shigesada, N.},
-  year = {1997},
-  month = sep,
-  doi = {10.1006/jtbi.1997.0462},
[…]
