+++ cellular_raza/Cargo.toml
@@ -35,7 +35,7 @@ num.workspace = true
 rand_chacha.workspace = true
 serde.workspace = true
 rand.workspace = true
-tempdir.workspace = true
+tempfile.workspace = true
 nalgebra.workspace = true
 ode_integrate = "0.0.2"
 
