+++ cellular_raza-benchmarks/Cargo.toml
@@ -25,7 +25,7 @@ clap = { version = "4.5.7", features = ["derive"] }
 kdam.workspace = true
 serde_json = "1.0.118"
 glob = "0.3.1"
-tempdir.workspace = true
+tempfile.workspace = true
 
 [dependencies.cellular_raza]
 path = "../cellular_raza"
