+++ Cargo.toml
@@ -39,7 +39,7 @@ rand = { version="0.8", features = ["small_rng"] }
 rand_distr = "0.4"
 rand_chacha = { version="0.3.1", features=["serde1"] }
 serde = "1.0"
-tempdir = "0.3"
+tempfile = "3"
 nalgebra = { version="<0.34, >=0.30", features=["serde-serialize", "rand"]}
 pyo3 = { version = "<0.24, >=0.22", features = ["serde", "py-clone"]}
 kdam = "0.6"
