### 2025-02-18 jonaspleyer/cellular_raza#10 @Armavica
Yes, this concerns the paper. Ok, then I'll ask the editor to clarify what this means.
