### 2025-02-18 #7723 @Armavica
@mahfuz05062 I have a question about the following item of the checklist: "_References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?_"

Does this mean that the software (direct? transitive?) dependencies should be cited in the paper?
