### 2025-02-18 99370c4
Update rand_distr requirement from 0.4 to 0.5
Updates the requirements on [rand_distr](https://github.com/rust-random/rand) to permit the latest version.
- [Release notes](https://github.com/rust-random/rand/releases)
- [Changelog](https://github.com/rust-random/rand/blob/master/CHANGELOG.md)
- [Commits](https://github.com/rust-random/rand/compare/rand_distr-0.4.3...0.5.0)

---
updated-dependencies:
- dependency-name: rand_distr
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
