### 2025-02-19 jonaspleyer/cellular_raza#12 @finsberg
[JOSS REVIEW] Improve "Statement of need"
I belive the "Statement of need" section in the paper could be improved by described what types of problem you would like to solve (i.e what types of question can be answered with the software). Also a short paragraph about what an agent-based model is would be good.

This is related to https://github.com/openjournals/joss-reviews/issues/7723
