### 2025-02-19 jonaspleyer/cellular_raza#12 @jonaspleyer
Hi @finsberg and thanks for the comment. My main problem when creating the manuscript was that the given total length of the paper was exceeding the threshold of the 1000 word limit initially given by JOSS (see https://github.com/openjournals/joss-reviews/issues/7520#issuecomment-2497204680).

Adding additional paragraphs would either mean that I will have to reduce content elsewhere or again exceed the word limit. What are your thoughts on this?
