### 2025-02-20 jonaspleyer/cellular_raza#12 @finsberg
My opinion is that what you currently write in the _Statement of need_ is better suited, and in some sense already covered, in the _State of the field_ section. I belive the _Statement of need_ should cover __what is the problem you are trying to solve__, rather than __how does your software compare with other software__. I also think that if you are limited on space in the paper then removing one example and expand on the statement of need would be better, since you have already covered the examples very nicely in your documenation
