### 2025-02-23 ddf10a5
bump to 0.2.0
- correct include statements for cellular_raza base crate
- include cellular_raza base crate in workspace
- update information in CITATION.cff
