+++ cellular_raza-building-blocks/Cargo.toml
@@ -33,7 +33,7 @@ serde_test = "1.0.177"
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.1.6"
+version = "0.2.0"
 
 [features]
 default = ["gradients"]
