+++ cellular_raza-benchmarks/Cargo.toml
@@ -29,7 +29,7 @@ tempfile.workspace = true
 
 [dependencies.cellular_raza]
 path = "../cellular_raza"
-version = "0.1.6"
+version = "0.2.0"
 
 [features]
 default = ["gradients"]
