+++ cellular_raza/Cargo.toml
@@ -11,7 +11,11 @@ keywords = ["cell", "cellular", "simulation"]
 categories = ["science", "mathematics", "simulation"]
 autoexamples = false
 homepage.workspace =  true
-include = ["logos/"]
+include = [
+    "logos/",
+    "src/",
+    "tests/",
+]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [package.metadata.docs.rs]
@@ -20,15 +24,15 @@ all-features = true
 
 [dependencies.cellular_raza-core]
 path = "../cellular_raza-core"
-version = "0.1.6"
+version = "0.2.0"
 
 [dependencies.cellular_raza-building-blocks]
 path = "../cellular_raza-building-blocks"
-version = "0.1.6"
+version = "0.2.0"
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.1.6"
+version = "0.2.0"
 
 [dev-dependencies]
 num.workspace = true
