+++ .github/workflows/cargo-docs-rs.yml
@@ -1,0 +1,14 @@
+name: test docs
+on: [push, pull_request]
+
+jobs:
+  doc:
+    name: Documentation
+    runs-on: ubuntu-latest
+    env:
+      RUSTDOCFLAGS: -Dwarnings
+    steps:
+      - uses: actions/checkout@v4
+      - uses: dtolnay/rust-toolchain@nightly
+      - uses: dtolnay/install@cargo-docs-rs
+      - run: cargo docs-rs
