### 2025-02-23 1a4698e
fix build script for docsrs
