+++ .github/workflows/cargo-docs-rs.yml
@@ -11,4 +11,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: dtolnay/rust-toolchain@nightly
       - uses: dtolnay/install@cargo-docs-rs
-      - run: cargo docs-rs
+      - run: cargo docs-rs -p cellular_raza-concepts
+      - run: cargo docs-rs -p cellular_raza-building_blocks
+      - run: cargo docs-rs -p cellular_raza-core
+      - run: cargo docs-rs -p cellular_raza
