### 2025-02-24 c52f187
produce release candidate for version 0.2.1
- include missing header file for cellular_raza documentation
- Add info for logo file
