+++ cellular_raza/logos/Infos.txt
@@ -1,0 +1,2 @@
+=== Font ===
+Liberation Sans (Bold)
