+++ cellular_raza/Cargo.toml
@@ -15,6 +15,7 @@ include = [
     "logos/",
     "src/",
     "tests/",
+    ".docs-header.html"
 ]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -24,15 +25,15 @@ all-features = true
 
 [dependencies.cellular_raza-core]
 path = "../cellular_raza-core"
-version = "0.2.0"
+version = "0.2.1-rc.1"
 
 [dependencies.cellular_raza-building-blocks]
 path = "../cellular_raza-building-blocks"
-version = "0.2.0"
+version = "0.2.1-rc.1"
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.2.0"
+version = "0.2.1-rc.1"
 
 [dev-dependencies]
 num.workspace = true
