### 2025-02-24 5e4381a
add documentation of Domain derive macro
