### 2025-02-24 0a88db7
do not deny when private docs are missing
