### 2025-02-24 e26884b
correct path name of building blocks crate
