+++ .github/workflows/cargo-docs-rs.yml
@@ -16,6 +16,6 @@ jobs:
       - uses: dtolnay/rust-toolchain@nightly
       - uses: dtolnay/install@cargo-docs-rs
       - run: cargo docs-rs -p cellular_raza-concepts
-      - run: cargo docs-rs -p cellular_raza-building_blocks
+      - run: cargo docs-rs -p cellular_raza-building-blocks
       - run: cargo docs-rs -p cellular_raza-core
       - run: cargo docs-rs -p cellular_raza
