### 2025-02-24 d63c11f
try with changing directory before invoking cargo-docs-rs
