### 2025-02-24 eee6f18
introduce reusable workflow
- package each crate before testing docs
- create workflow for each (major) crate
