### 2025-02-24 d675bec
change target-dir and simply iterate over all candidates
