+++ .github/workflows/cargo-docs-rs-reuse.yml
@@ -22,7 +22,7 @@ jobs:
       - uses: dtolnay/rust-toolchain@nightly
       - uses: dtolnay/install@cargo-docs-rs
       - run: |
-          cargo package -p --all-features ${{ inputs.crate }}
+          cargo package --all-features -p ${{ inputs.crate }}
           crate_path=target/package/${{ inputs.crate }}
           for c in $( ls $crate_path-*/ ); do
             echo Generating Documentation for $c
