### 2025-02-24 a5bde44
restructure build file to work for docsrs
