### 2025-02-24 707bd3d
fixup! restructure build file to work for docsrs
