### 2025-02-24 dc43a02
try enabling via environment variable
