### 2025-02-24 6c4a68c
fixup! try enabling via environment variable
