### 2025-02-24 d696171
fixup! fixup! try enabling via environment variable
