+++ .github/workflows/cargo-docs-rs.yml
@@ -1,28 +1,0 @@
-name: test docs
-on: [push, pull_request]
-
-jobs:
-  doctest-cellular_raza-concepts-derive:
-    uses: ./.github/workflows/cargo-docs-rs-reuse.yml
-    with:
-      crate: cellular_raza-concepts-derive
-  doctest-cellular_raza-concepts:
-    uses: ./.github/workflows/cargo-docs-rs-reuse.yml
-    with:
-      crate: cellular_raza-concepts
-  doctest-cellular_raza-core-proc-macro:
-    uses: ./.github/workflows/cargo-docs-rs-reuse.yml
-    with:
-      crate: cellular_raza-core-proc-macro
-  doctest-cellular_raza-core:
-    uses: ./.github/workflows/cargo-docs-rs-reuse.yml
-    with:
-      crate: cellular_raza-core
-  doctest-cellular_raza-building-blocks:
-    uses: ./.github/workflows/cargo-docs-rs-reuse.yml
-    with:
-      crate: cellular_raza-building-blocks
-  doctest-cellular_raza:
-    uses: ./.github/workflows/cargo-docs-rs-reuse.yml
-    with:
-      crate: cellular_raza
