+++ .github/workflows/cargo-docs-rs-reuse.yml
@@ -1,31 +1,0 @@
-name: cargo-docs-rs
-
-on:
-  workflow_call:
-    inputs:
-      crate:
-        required: true
-        type: string
-
-jobs:
-  doc:
-    name: Create Docs
-    runs-on: ubuntu-latest
-    env:
-      DOCS_RS: 1
-      RUSTDOCFLAGS: -Dwarnings
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 1
-          submodules: 'recursive'
-          lfs: 'false'
-      - uses: dtolnay/rust-toolchain@nightly
-      - uses: dtolnay/install@cargo-docs-rs
-      - run: |
-          cargo package --all-features -p ${{ inputs.crate }}
-          crate_path=target/package/${{ inputs.crate }}
-          for c in $( ls $crate_path-*/ ); do
-            echo Generating Documentation for $c
-            cd target/package/$c && cargo docs-rs; cd ../../../
-          done
