+++ CITATION.cff
@@ -20,5 +20,5 @@ repository-code: 'https://github.com/jonaspleyer/cellular_raza'
 url: 'https://cellular-raza.com'
 license: GPL-2.0
 commit: 0.0.5
-version: 0.2.1-rc.1
+version: 0.2.1-rc.2
 date-released: '2025-02-23'
