+++ cellular_raza-concepts/Cargo.toml
@@ -24,7 +24,7 @@ plotters = { workspace = true }
 
 [dependencies.cellular_raza-concepts-derive]
 path = "../cellular_raza-concepts-derive"
-version = "0.2.1-rc.1"
+version = "0.2.1-rc.2"
 
 [dev-dependencies]
 rand = { workspace = true }
