### 2025-02-25 9d42c7b
add creature comfort scripts
