+++ cellular_raza-core/Cargo.toml
@@ -56,11 +56,11 @@ wgpu = { version = "24.0", optional = true }
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.2.1-rc.3"
+version = "0.2.1"
 
 [dependencies.cellular_raza-core-proc-macro]
 path = "../cellular_raza-core-proc-macro"
-version = "0.2.1-rc.3"
+version = "0.2.1"
 
 [dev-dependencies]
 tempfile.workspace = true
