+++ cellular_raza-benchmarks/Cargo.toml
@@ -29,7 +29,7 @@ tempfile.workspace = true
 
 [dependencies.cellular_raza]
 path = "../cellular_raza"
-version = "0.2.1-rc.3"
+version = "0.2.1"
 
 [features]
 default = ["gradients"]
