+++ cellular_raza-building-blocks/Cargo.toml
@@ -33,7 +33,7 @@ serde_test = "1.0.177"
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.2.1-rc.3"
+version = "0.2.1"
 
 [features]
 default = ["gradients"]
