### 2025-03-04 #7723 @mahfuz05062
> [@mahfuz05062](https://github.com/mahfuz05062) May I ask if you have any advice on this? [#7723 (comment)](https://github.com/openjournals/joss-reviews/issues/7723#issuecomment-2665911545)

Sorry for the delayed replay! Do you mean software package dependencies? If so, I don't see any need unless this new package heavily relies/builds upon a previous package.
