### 2025-03-05 863f735
implement ordering for CellIdentifier
