### 2025-03-05 #7723 @jonaspleyer
> Sorry for the delayed replay! Do you mean software package dependencies? If so, I don't see any need unless this new package heavily relies/builds upon a previous package.

Thanks for the clarification. I have cited two dependencies (`pyo3` and `maturin`) which have been mentioned by name in the text. Otherwise, the dependencies are documented in the `Cargo.toml` file which is used by the `cargo` package manager for organizing said dependencies.
