### 2025-03-05 7bf50fc
extend statement of need; move some information to "State of Field"
