### 2025-03-05 jonaspleyer/cellular_raza#10 @jonaspleyer
Closing this issue with https://github.com/openjournals/joss-reviews/issues/7723#issuecomment-2696421218
