### 2025-03-05 jonaspleyer/cellular_raza#12 @jonaspleyer
I have rewritten the two sections and included more details about the nature of ABMs and generated a new pdf.
https://github.com/openjournals/joss-reviews/issues/7723#issuecomment-2700434414
