### 2025-03-05 #7723 @finsberg
@mahfuz05062 I am done with my review of `cellular_raza` and can recommend it for publication in JOSS. @jonaspleyer has developed a novel package for agent-based modeling of cells. The package is easy to install and has an extensive documentation with well written tutorials. @jonaspleyer has also been very responsive in addressing the issues we have raised.
