### 2025-03-05 df216e9
reorder acronyms
