### 2025-03-05 #7723 @mahfuz05062
> [@mahfuz05062](https://github.com/mahfuz05062) I am done with my review of `cellular_raza` and can recommend it for publication in JOSS. [@jonaspleyer](https://github.com/jonaspleyer) has developed a novel package for agent-based modeling of cells. The package is easy to install and has an extensive documentation with well written tutorials. [@jonaspleyer](https://github.com/jonaspleyer) has also been very responsive in addressing the issues we have raised.

Thanks @finsberg! Appreciate you completing your review in a timely manner. Can you also check the two remaining items of your list?
