### 2025-03-10 #7723 @jonaspleyer
> [@Armavica](https://github.com/Armavica) I see you are almost done with your review. Please let me know if you have any questions. Thanks!

It's me who is holding up. I am almost done with the last issue raised so far. I hope that I can finish this within the next days.
https://github.com/jonaspleyer/cellular_raza/issues/4
