### 2025-03-10 0ea256d
Update bincode requirement from 1.3 to 2.0
Updates the requirements on [bincode](https://github.com/bincode-org/bincode) to permit the latest version.
- [Release notes](https://github.com/bincode-org/bincode/releases)
- [Commits](https://github.com/bincode-org/bincode/compare/v1.3.0...v2.0.0)

---
updated-dependencies:
- dependency-name: bincode
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
