+++ cellular_raza-core/Cargo.toml
@@ -31,7 +31,7 @@ rayon = { version="1.6" }
 kdam = { workspace = true, features = ["template", "rayon"] }
 
 # Storage
-bincode = "1.3"
+bincode = "2.0"
 serde = { workspace = true }
 quick-xml = { version="0.37", features=["serialize"]}
 serde_json = { version="1.0" }
