+++ cellular_raza-core/src/backend/cpu_os_threads/domain_decomposition.rs
@@ -126,16 +126,16 @@ pub struct VoxelBox<
 }
 
 impl<
-        Ind,
-        Vox,
-        Cel,
-        Pos,
-        Vel,
-        For,
-        ConcVecExtracellular,
-        ConcBoundaryExtracellular,
-        ConcVecIntracellular,
-    > Id
+    Ind,
+    Vox,
+    Cel,
+    Pos,
+    Vel,
+    For,
+    ConcVecExtracellular,
+    ConcBoundaryExtracellular,
+    ConcVecIntracellular,
+> Id
     for VoxelBox<
         Ind,
         Pos,
@@ -195,16 +195,16 @@ where
 }
 
 impl<
-        Ind,
-        Vox,
-        Cel,
-        Pos,
-        Vel,
-        For,
-        ConcVecExtracellular,
-        ConcBoundaryExtracellular,
-        ConcVecIntracellular,
-    >
+    Ind,
+    Vox,
+    Cel,
+    Pos,
+    Vel,
+    For,
+    ConcVecExtracellular,
+    ConcBoundaryExtracellular,
+    ConcVecIntracellular,
+>
     VoxelBox<
         Ind,
         Pos,
@@ -260,16 +260,16 @@ where
 }
 
 impl<
-        Ind,
-        Vox,
-        Cel,
-        Pos,
-        Vel,
-        For,
-        ConcVecExtracellular,
-        ConcBoundaryExtracellular,
-        ConcVecIntracellular,
-    >
+    Ind,
+    Vox,
+    Cel,
+    Pos,
+    Vel,
+    For,
+    ConcVecExtracellular,
+    ConcBoundaryExtracellular,
+    ConcVecIntracellular,
+>
     VoxelBox<
         Ind,
         Pos,
@@ -579,18 +579,18 @@ pub struct MultiVoxelContainer<
 }
 
 impl<
-        Ind,
-        Pos,
-        Vel,
-        For,
-        Inf,
-        ConcVecExtracellular,
-        ConcBoundaryExtracellular,
-        ConcVecIntracellular,
-        Vox,
-        Dom,
-        Cel,
-    >
+    Ind,
+    Pos,
+    Vel,
+    For,
+    Inf,
+    ConcVecExtracellular,
+    ConcBoundaryExtracellular,
+    ConcVecIntracellular,
+    Vox,
+    Dom,
+    Cel,
+>
     MultiVoxelContainer<
         Ind,
         Pos,
@@ -633,13 +633,13 @@ where
         Cel: Agent<Pos, Vel, For, Inf>
             + InteractionExtracellularGradient<Cel, ConcGradientExtracellular>,
         Vox: ExtracellularMechanics<
-            Ind,
-            Pos,
-            ConcVecExtracellular,
-            ConcGradientExtracellular,
-            ConcTotalExtracellular,
-            ConcBoundaryExtracellular,
-        >,
+                Ind,
+                Pos,
+                ConcVecExtracellular,
+                ConcGradientExtracellular,
+                ConcTotalExtracellular,
+                ConcBoundaryExtracellular,
+            >,
     {
         self.voxels
             .iter_mut()
@@ -777,13 +777,13 @@ where
     ) -> Result<(), SimulationError>
     where
         Vox: ExtracellularMechanics<
-            Ind,
-            Pos,
-            ConcVecExtracellular,
-            ConcGradientExtracellular,
-            ConcTotalExtracellular,
-            ConcBoundaryExtracellular,
-        >,
+                Ind,
+                Pos,
+                ConcVecExtracellular,
+                ConcGradientExtracellular,
+                ConcTotalExtracellular,
+                ConcBoundaryExtracellular,
+            >,
     {
         let indices_iterator = self
[…]
