### 2025-03-11 199570f
if index is not found return error instead
