### 2025-03-11 5281e08
use clap to parse command line arguments instead of hard-coding them
