+++ cellular_raza-examples/bacterial_branching/src/main.rs
@@ -2,8 +2,8 @@ use backend::chili::Settings;
 use cellular_raza::core::backend::chili;
 use cellular_raza::prelude::*;
 
+use clap::{Args, Parser};
 use nalgebra::Vector2;
-
 use num::Zero;
 
 use rand::{Rng, SeedableRng};
@@ -11,96 +11,168 @@ use rand_chacha::ChaCha8Rng;
 
 use serde::{Deserialize, Serialize};
 
-// Number of cells to put into simulation in the Beginning
-pub const N_BACTERIA_INITIAL: u32 = 3;
 
-// Mechanical parameters
-pub const BACTERIA_MECHANICS_RADIUS: f32 = 6.0;
-pub const BACTERIA_MECHANICS_EXPONENT: f32 = 0.2;
-pub const BACTERIA_MECHANICS_POTENTIAL_STRENGTH: f32 = 4.0;
-pub const BACTERIA_MECHANICS_VELOCITY_REDUCTION: f32 = 1.0;
 
-// Reaction parameters of the cell
-pub const BACTERIA_FOOD_INITIAL_CONCENTRATION_LOW: f32 = 0.0;
-pub const BACTERIA_FOOD_INITIAL_CONCENTRATION_HIGH: f32 = 0.0;
-pub const BACTERIA_FOOD_TURNOVER_RATE: f32 = 0.0;
-pub const BACTERIA_FOOD_UPTAKE_RATE: f32 = 0.5;
+#[derive(Clone, Args, Debug)]
+#[group()]
+#[clap(next_help_heading = Some("Bacteria"))]
+struct BacterialParameters {
+    /// Number of cells to put into simulation in the Beginning
+    #[arg(short, long, default_value_t = 3)]
+    n_bacteria_initial: u32,
 
-// Parameters for cell cycle
-pub const BACTERIA_CYCLE_GROWTH_RATE: f32 = 5.2;
+    /// Mechanical parameters
+    #[arg(short, long, default_value_t = 6.0)]
+    radius: f32,
+    #[arg(long, default_value_t = 0.2)]
+    exponent: f32,
+    #[arg(long, default_value_t = 4.0)]
+    potential_strength: f32,
+    #[arg(long, default_value_t = 1.0)]
+    damping_constant: f32,
 
-// Parameters for domain
-pub const DOMAIN_SIZE: f32 = 5_000.0;
-pub const DOMAIN_MIDDLE: Vector2<f32> = nalgebra::vector![DOMAIN_SIZE / 2.0, DOMAIN_SIZE / 2.0];
+    /// Parameters for cell cycle
+    #[arg(short, long, default_value_t = 0.8)]
+    uptake_rate: f32,
+    #[arg(short, long, default_value_t = 3.0)]
+    growth_rate: f32,
+}
 
-// Where will the cells be placed initially
-pub const STARTING_DOMAIN_X_LOW: f32 = DOMAIN_SIZE / 2.0 - 50.0;
-pub const STARTING_DOMAIN_X_HIGH: f32 = DOMAIN_SIZE / 2.0 + 50.0;
-pub const STARTING_DOMAIN_Y_LOW: f32 = DOMAIN_SIZE / 2.0 - 50.0;
-pub const STARTING_DOMAIN_Y_HIGH: f32 = DOMAIN_SIZE / 2.0 + 50.0;
+#[derive(Clone, Args, Debug)]
+#[group()]
+#[clap(next_help_heading = Some("Domain"))]
+struct DomainParameters {
+    /// Parameters for domain
+    #[arg(short, long, default_value_t = 300.0)]
+    domain_size: f32,
+    /// Discretization used to model the diffusion process
+    #[arg(long, default_value_t = 20.0)]
[…]
