+++ cellular_raza-examples/bacterial_branching/src/subdomain.rs
@@ -1,0 +1,329 @@
+use crate::{ReactionVector, N_REACTIONS};
+use cellular_raza::prelude::*;
+use nalgebra::SVector;
+
+use serde::{Deserialize, Serialize};
+
+pub struct MyDomain {
+    #[DomainRngSeed]
+    #[DomainPartialDerive]
+    #[SortCells]
+    pub domain: CartesianCuboid<f32, 2>,
+    pub reactions_dx: f32,
+    pub diffusion_constant: f32,
+    pub initial_value: ReactionVector,
+}
+
+#[derive(SubDomain, Clone, Serialize)]
+pub struct MySubDomain {
+    #[Base]
+    #[SortCells]
+    #[Mechanics]
+    pub subdomain: CartesianSubDomain<f32, 2>,
+    pub reactions_min: nalgebra::Vector2<f32>,
+    pub reactions_max: nalgebra::Vector2<f32>,
+    pub reactions_dx: nalgebra::Vector2<f32>,
+    pub extracellular: ndarray::Array3<f32>,
+    pub diffusion_constant: f32,
+    increments: [ndarray::Array3<f32>; 3],
+    increments_start: usize,
+    helper: ndarray::Array3<f32>,
+}
+
+impl MySubDomain {
+    fn merge_values(
+        &mut self,
+        neighbor: Result<NeighborValue, CalcError>,
+    ) -> Result<(), CalcError> {
+        let neighbor = neighbor?;
+        let neighbor_slice = neighbor.extracellular_slice;
+        self.helper
+            .slice_mut(ndarray::s![
+                neighbor.index_min[0]..neighbor.index_max[0],
+                neighbor.index_min[1]..neighbor.index_max[1],
+                ..
+            ])
+            .assign(&neighbor_slice);
+        Ok(())
+    }
+
+    pub fn get_extracellular_index_raw(
+        reactions_min: &nalgebra::Vector2<f32>,
+        reactions_max: &nalgebra::Vector2<f32>,
+        reactions_dx: &nalgebra::Vector2<f32>,
+        pos: &nalgebra::Vector2<f32>,
+    ) -> Result<nalgebra::Vector2<usize>, CalcError> {
+        if pos[0] < reactions_min[0]
+            || pos[0] > reactions_max[0]
+            || pos[1] < reactions_min[1]
+            || pos[1] > reactions_max[1]
+        {
+            return Err(CalcError(format!(
+                "position {:?} is not contained in domain with boundaries {:?} {:?}",
+                pos, reactions_min, reactions_max
+            )));
+        }
+        let index = (pos - reactions_min)
+            .component_div(&reactions_dx)
+            .map(|i| i as usize);
+        Ok(index)
+    }
+
+    pub fn get_extracellular_index(
+        &self,
+        pos: &nalgebra::Vector2<f32>,
+    ) -> Result<nalgebra::Vector2<usize>, CalcError> {
+        Self::get_extracellular_index_raw(
+            &self.reactions_min,
+            &self.reactions_max,
+            &self.reactions_dx,
+            &pos,
+        )
+    }
+}
+
+pub struct NeighborValue {
+    index_min: nalgebra::Vector2<usize>,
+    index_max: nalgebra::Vector2<usize>,
+    extracellular_slice: ndarray::Array3<f32>,
+}
+
+pub struct BorderInfo {
+    min_sent: nalgebra::Vector2<f32>,
+    max_sent: nalgebra::Vector2<f32>,
+}
+
+impl SubDomainReactions<nalgebra::SVector<f32, 2>, ReactionVector, f32> for MySubDomain {
+    type NeighborValue = Result<NeighborValue, CalcError>;
[…]
