### 2025-03-11 a587c87
formatting of plotting script
