### 2025-03-11 0db3c12
switch from custom interaction potential to morsepotential
