+++ cellular_raza-examples/bacterial_branching/src/bacteria_properties.rs
@@ -1,47 +1,19 @@
-use core::f32;
+use core::f32::consts::{PI, SQRT_2};
 
 use cellular_raza::prelude::*;
-use nalgebra::Vector2;
 use serde::{Deserialize, Serialize};
 
-pub const N_REACTIONS: usize = 1;
-pub type ReactionVector = nalgebra::SVector<f32, N_REACTIONS>;
+use crate::ReactionVector;
 
-#[derive(Serialize, Deserialize, Clone, core::fmt::Debug)]
-pub struct MyInteraction {
-    pub potential_strength: f32,
-    pub exponent: f32,
-    pub cell_radius: f32,
-}
-
-impl Interaction<Vector2<f32>, Vector2<f32>, Vector2<f32>, f32> for MyInteraction {
-    fn calculate_force_between(
-        &self,
-        own_pos: &Vector2<f32>,
-        _own_vel: &Vector2<f32>,
-        ext_pos: &Vector2<f32>,
-        _ext_vel: &Vector2<f32>,
-        ext_radius: &f32,
-    ) -> Result<(Vector2<f32>, Vector2<f32>), CalcError> {
-        let z = own_pos - ext_pos;
-        let r = z.norm();
-        if r == 0.0 {
-            return Ok((num::Zero::zero(), num::Zero::zero()));
-        }
-
-        let sigma = r / (self.cell_radius + ext_radius);
-        let strength = if sigma <= 1.0 {
-            self.potential_strength
-        } else {
-            0.0
-        };
-        let force = strength * z.normalize() / sigma.powf(self.exponent);
-        Ok((force, -force))
-    }
-
-    fn get_interaction_information(&self) -> f32 {
-        self.cell_radius
-    }
+#[derive(Clone, Serialize, Deserialize, CellAgent)]
+pub struct MyAgent {
+    #[Mechanics]
+    pub mechanics: NewtonDamped2DF32,
+    #[Interaction]
+    pub interaction: MorsePotentialF32,
+    pub uptake_rate: f32,
+    pub division_radius: f32,
+    pub growth_rate: f32,
 }
 
 impl Cycle<MyAgent, f32> for MyAgent {
@@ -51,7 +23,7 @@ impl Cycle<MyAgent, f32> for MyAgent {
         cell: &mut MyAgent,
     ) -> Option<CycleEvent> {
         // If the cell is not at the maximum size let it grow
-        if cell.interaction.cell_radius > cell.division_radius {
+        if cell.interaction.radius > cell.division_radius {
             return Some(CycleEvent::Division);
         }
         None
@@ -64,17 +36,17 @@ impl Cycle<MyAgent, f32> for MyAgent {
         // Clone existing cell
         let mut c2 = c1.clone();
 
-        let r = c1.interaction.cell_radius;
+        let r = c1.interaction.radius;
 
         // Make both cells smaller
         // Also keep old cell larger
-        c1.interaction.cell_radius /= std::f32::consts::SQRT_2;
-        c2.interaction.cell_radius /= std::f32::consts::SQRT_2;
+        c1.interaction.radius /= SQRT_2;
+        c2.interaction.radius /= SQRT_2;
 
         // Generate cellular splitting direction randomly
         use rand::Rng;
-        let angle_1 = rng.gen_range(0.0..2.0 * std::f32::consts::PI);
-        let dir_vec = nalgebra::Rotation2::new(angle_1) * nalgebra::Vector2::from([1.0, 0.0]);
+        let alpha = rng.gen_range(0.0..2.0 * PI);
+        let dir_vec = nalgebra::Vector2::from([alpha.cos(), alpha.sin()]);
 
         // Define new positions for cells
[…]
