### 2025-03-11 f864805
correctly implement diffusion calculation
