+++ cellular_raza-examples/bacterial_branching/src/subdomain.rs
@@ -1,22 +1,24 @@
-use crate::{ReactionVector, N_REACTIONS};
+pub const N_REACTIONS: usize = 1;
+pub type ReactionVector = nalgebra::SVector<f32, N_REACTIONS>;
+
 use cellular_raza::prelude::*;
-use nalgebra::SVector;
+use serde::{Deserialize, Serialize};
 
-use serde::Serialize;
-
-#[derive(Domain)]
-pub struct MyDomain {
+#[derive(Domain, Clone)]
+pub struct CartesianDiffusion2D {
     #[DomainRngSeed]
     #[DomainPartialDerive]
     #[SortCells]
     pub domain: CartesianCuboid<f32, 2>,
-    pub reactions_dx: f32,
+    /// The discretization must be a multiple of the voxel size.
+    /// This quantity will be used as an initial estimate and rounded to the nearest candidate.
+    pub reactions_dx: nalgebra::Vector2<f32>,
     pub diffusion_constant: f32,
     pub initial_value: ReactionVector,
 }
 
-#[derive(SubDomain, Clone, Serialize)]
-pub struct MySubDomain {
+#[derive(Deserialize, SubDomain, Clone, Serialize, Debug)]
+pub struct CartesianDiffusion2DSubDomain {
     #[Base]
     #[SortCells]
     #[Mechanics]
@@ -24,98 +26,122 @@ pub struct MySubDomain {
     pub reactions_min: nalgebra::Vector2<f32>,
     pub reactions_max: nalgebra::Vector2<f32>,
     pub reactions_dx: nalgebra::Vector2<f32>,
+    pub index_min: nalgebra::Vector2<usize>,
+    pub index_max: nalgebra::Vector2<usize>,
     pub extracellular: ndarray::Array3<f32>,
+    pub ownership_array: ndarray::Array2<bool>,
     pub diffusion_constant: f32,
     increments: [ndarray::Array3<f32>; 3],
     increments_start: usize,
     helper: ndarray::Array3<f32>,
 }
 
-impl MySubDomain {
-    fn merge_values(
-        &mut self,
-        neighbor: Result<NeighborValue, CalcError>,
-    ) -> Result<(), CalcError> {
-        let neighbor = neighbor?;
-        let neighbor_slice = neighbor.extracellular_slice;
-        self.helper
-            .slice_mut(ndarray::s![
-                neighbor.index_min[0]..neighbor.index_max[0],
-                neighbor.index_min[1]..neighbor.index_max[1],
-                ..
-            ])
-            .assign(&neighbor_slice);
-        Ok(())
-    }
+pub struct BorderInfo {
+    min_sent: nalgebra::Vector2<usize>,
+    max_sent: nalgebra::Vector2<usize>,
+}
 
-    pub fn get_extracellular_index_raw(
-        reactions_min: &nalgebra::Vector2<f32>,
-        reactions_max: &nalgebra::Vector2<f32>,
-        reactions_dx: &nalgebra::Vector2<f32>,
-        pos: &nalgebra::Vector2<f32>,
-    ) -> Result<nalgebra::Vector2<usize>, CalcError> {
-        if pos[0] < reactions_min[0]
-            || pos[0] > reactions_max[0]
-            || pos[1] < reactions_min[1]
-            || pos[1] > reactions_max[1]
-        {
-            return Err(CalcError(format!(
-                "position {:?} is not contained in domain with boundaries {:?} {:?}",
-                pos, reactions_min, reactions_max
-            )));
-        }
-        let index = (pos - reactions_min)
-            .component_div(&reactions_dx)
-            .map(|i| i as usize);
-        Ok(index)
-    }
+impl CartesianDiffusion2DSubDomain {
[…]
