### 2025-03-11 9ca1f6b
adjust arguments to new inputs; also simplify some descriptions and adjust default settings
