### 2025-03-11 abda7de
print warnings for different cases where input values can be problematic
