### 2025-03-11 1d29cf9
adjust plotting to new name of variables
