### 2025-03-11 670b442
combine values in array instead of plotting them individually
