### 2025-03-12 0e03813
set new parameters for simple simulation
