### 2025-03-12 a857232
format warning messages
