### 2025-03-12 01e3a97
correctly specify intracellular bounds
