### 2025-03-12 c1c04a0
add ability to specify figsize
