### 2025-03-12 bcb6155
do not use different color for edgecolors
