### 2025-03-12 54b3420
use config for (de)serializing with bincode
