+++ cellular_raza-examples/bacterial_branching/src/plotting.py
@@ -10,6 +10,7 @@ import tqdm
 import multiprocessing as mp
 import itertools
 
+
 def get_last_output_path(search_dir: Path | None = Path("out")) -> Path:
     """
     Parameters
@@ -24,20 +25,24 @@ def get_last_output_path(search_dir: Path | None = Path("out")) -> Path:
     """
     return Path(sorted(list(glob(str(search_dir) + "/*")))[-1])
 
+
 def _get_all_iteration_files(output_path: Path = get_last_output_path()) -> list[Path]:
     return [Path(p) for p in sorted(glob(str(output_path) + "/cells/json/*"))]
 
+
 def _iteration_to_file(iteration: int, output_path: Path, cs: str = "cells") -> Path:
     return output_path / "{}/json/{:020}".format(cs, iteration)
+
 
 def get_all_iterations(output_path: Path = get_last_output_path()) -> list[int]:
     iterations_files = _get_all_iteration_files(output_path)
     return [int(os.path.basename(it)) for it in iterations_files]
 
+
 def load_cells_at_iteration(
-        iteration: int,
-        output_path: Path,
-    ):
+    iteration: int,
+    output_path: Path,
+):
     iteration_file = _iteration_to_file(iteration, output_path, "cells")
     data = []
     for filename in glob(str(iteration_file) + "/*"):
@@ -52,10 +57,10 @@ def load_cells_at_iteration(
         df[key] = df[key].apply(lambda x: np.array(x, dtype=float))
     return df
 
+
 def load_subdomains_at_iteration(
-        iteration: int,
-        output_path: Path = get_last_output_path()
-    ) -> pd.DataFrame:
+    iteration: int, output_path: Path = get_last_output_path()
+) -> pd.DataFrame:
     iteration_file = _iteration_to_file(iteration, output_path, "subdomains")
     data = []
     for filename in glob(str(iteration_file) + "/*"):
@@ -74,63 +79,77 @@ def load_subdomains_at_iteration(
         "reactions_max",
         "reactions_dx",
         "extracellular.data",
+        "ownership_array.data",
     ]:
         df[key] = df[key].apply(lambda x: np.array(x, dtype=float))
     return df
 
+
 def plot_iteration(
-        iteration: int,
-        intra_bounds: tuple[float, float],
-        extra_bounds: tuple[float, float],
-        output_path: Path = get_last_output_path(),
-        save_figure: bool = True,
-    ) -> matplotlib.figure.Figure | None:
+    iteration: int,
+    intra_bounds: tuple[float, float],
+    extra_bounds: tuple[float, float],
+    output_path: Path = get_last_output_path(),
+    save_figure: bool = True,
+    figsize: int = 32,
+) -> matplotlib.figure.Figure | None:
     dfc = load_cells_at_iteration(iteration, output_path)
     dfs = load_subdomains_at_iteration(iteration, output_path)
 
     # Set size of the image
     domain_min = dfs["subdomain.domain_min"][0]
     domain_max = dfs["subdomain.domain_max"][0]
-    fig, ax = plt.subplots(figsize=(16,16))
+    fig, ax = plt.subplots(figsize=(figsize, figsize))
     ax.set_xlim([domain_min[0], domain_max[0]])
     ax.set_ylim([domain_min[1], domain_max[1]])
 
     # Plot background
+    max_size = np.max([dfsi["index_max"] for _, dfsi in dfs.iterrows()], axis=0)
+    all_values = np.zeros(max_size)
     for n_sub, dfsi in dfs.iterrows():
-        smin = dfsi["reactions_min"]
-        smax = dfsi["reactions_max"]
[…]
