+++ cellular_raza-core/src/storage/sled_database.rs
@@ -15,6 +15,7 @@ pub struct SledStorageInterface<Id, Element, const TEMP: bool = false> {
     // buffer: StorageBuffer<Id, Element>,
     id_phantom: PhantomData<Id>,
     element_phantom: PhantomData<Element>,
+    bincode_config: bincode::config::Configuration,
 }
 
 impl<Id, Element, const TEMP: bool> SledStorageInterface<Id, Element, TEMP> {
@@ -24,8 +25,11 @@ impl<Id, Element, const TEMP: bool> SledStorageInterface<Id, Element, TEMP> {
     }
 
     /// Transform the key given by the tree to the corresponding iteartion u64 value
-    fn key_to_iteration(key: &sled::IVec) -> Result<u64, StorageError> {
-        let iteration = bincode::deserialize::<u64>(key)?;
+    fn key_to_iteration(
+        key: &sled::IVec,
+        bincode_config: bincode::config::Configuration,
+    ) -> Result<u64, StorageError> {
+        let iteration = bincode::serde::decode_from_slice(key, bincode_config)?.0;
         Ok(iteration)
     }
 
@@ -62,11 +66,13 @@ impl<Id, Element, const TEMP: bool> StorageInterfaceOpen
             .use_compression(false);
 
         let db = config.open()?;
+        let bincode_config = bincode::config::standard();
 
         Ok(SledStorageInterface {
             db,
             id_phantom: PhantomData,
             element_phantom: PhantomData,
+            bincode_config,
         })
     }
 }
@@ -87,8 +93,9 @@ impl<Id, Element, const TEMP: bool> StorageInterfaceStore<Id, Element>
         let tree = self.open_or_create_tree(iteration)?;
 
         // Serialize the identifier and the element
-        let identifier_serialized = bincode::serialize(&identifier)?;
-        let element_serialized = bincode::serialize(&element)?;
+        let identifier_serialized =
+            bincode::serde::encode_to_vec(&identifier, self.bincode_config)?;
+        let element_serialized = bincode::serde::encode_to_vec(&element, self.bincode_config)?;
         match tree.insert(identifier_serialized, element_serialized)? {
             None => Ok(()),
             Some(_) => Err(StorageError::InitError(format!(
@@ -112,8 +119,9 @@ impl<Id, Element, const TEMP: bool> StorageInterfaceStore<Id, Element>
         let tree = self.open_or_create_tree(iteration)?;
         let mut batch = sled::Batch::default();
         for (identifier, element) in identifiers_elements.into_iter() {
-            let identifier_serialized = bincode::serialize(&identifier)?;
-            let element_serialized = bincode::serialize(&element)?;
+            let identifier_serialized =
+                bincode::serde::encode_to_vec(&identifier, self.bincode_config)?;
+            let element_serialized = bincode::serde::encode_to_vec(&element, self.bincode_config)?;
             batch.insert(identifier_serialized, element_serialized)
         }
         tree.apply_batch(batch)?;
@@ -137,10 +145,11 @@ impl<Id, Element, const TEMP: bool> StorageInterfaceLoad<Id, Element>
             Some(tree) => tree,
             None => return Ok(None),
         };
-        let identifier_serialized = bincode::serialize(identifier)?;
+        let identifier_serialized = bincode::serde::encode_to_vec(identifier, self.bincode_config)?;
         match tree.get(&identifier_serialized)? {
             Some(element_serialized) => {
-                let element: Element = bincode::deserialize(&element_serialized)?;
+                let element: Element =
+                    bincode::serde::decode_from_slice(&element_serialized, self.bincode_config)?.0;
                 Ok(Some(element))
             }
             None => Ok(None),
[…]
