### 2025-03-12 68acc12
manually implement Debug trait for Sled database struct
