### 2025-03-12 147cbbc
catch (De)serialize errors from bincode
- remove box around type
