### 2025-03-12 9914c92
fixup! Update bincode requirement from 1.3 to 2.0
