+++ cellular_raza-core/Cargo.toml
@@ -31,7 +31,7 @@ rayon = { version="1.6" }
 kdam = { workspace = true, features = ["template", "rayon"] }
 
 # Storage
-bincode = "2.0"
+bincode = { version = "2.0", features = ["serde"] }
 serde = { workspace = true }
 quick-xml = { version="0.37", features=["serialize"]}
 serde_json = { version="1.0" }
