### 2025-03-12 jonaspleyer/cellular_raza#4 @jonaspleyer
I have completely reworked this example such that it can reliably produce the desired output. However, it will take me 1-2 more days to document it and update the website. Sorry for the delay.
