### 2025-03-12 e81a54c
pick new default parameters for simulation
