+++ Cargo.toml
@@ -15,7 +15,7 @@ resolver = "2"
 [workspace.package]
 version = "0.2.1"
 authors = ["Jonas Pleyer <jonas.dev@pleyer.org>"]
-edition = "2021"
+edition = "2024"
 license = "GPL-2.0"
 repository = "https://github.com/jonaspleyer/cellular_raza"
 keywords = ["cell", "cellular", "simulation"]
