### 2025-03-12 dd528e4
bump to 0.2.2
