+++ cellular_raza/Cargo.toml
@@ -25,15 +25,15 @@ all-features = true
 
 [dependencies.cellular_raza-core]
 path = "../cellular_raza-core"
-version = "0.2.1"
+version = "0.2.2"
 
 [dependencies.cellular_raza-building-blocks]
 path = "../cellular_raza-building-blocks"
-version = "0.2.1"
+version = "0.2.2"
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.2.1"
+version = "0.2.2"
 
 [dev-dependencies]
 num.workspace = true
