### 2025-03-12 5e13f54
specify that functions of this extern block are unsafe
