### 2025-03-12 jonaspleyer/cellular_raza#4 @jonaspleyer
@Armavica I have updated the documentation at
https://cellular-raza.com/showcase/bacterial-branching/
I am now using the new `chili` backend instead of the old `cpu_os_threads` backend. Furthermore, one can now use a CLI tool to specify parameters of the simulation. I have generated new images and a new movie which are now reproducible with the CLI arguments specified in the documentation.

Note that the default parameters for domain and maximum time have been chosen smaller than the one in the image to provide a quicker running example. However, the parameters for the image are documented on the homepage.

Le me know what you think about this.
