+++ Cargo.toml
@@ -13,7 +13,7 @@ members = [
 resolver = "2"
 
 [workspace.package]
-version = "0.2.2"
+version = "0.2.3"
 authors = ["Jonas Pleyer <jonas.dev@pleyer.org>"]
 edition = "2024"
 license = "GPL-2.0"
