### 2025-03-15 #7723 @jonaspleyer
@mahfuz05062 I would like to modify the version number from `0.1.6` to `0.2.3`. The issues raised and other developments have justified this version increase.
I have already adjusted the date in the paper, referring to the date of commit [66bdd09bb8d358875d57a2c1406479b0d7d51c1e](https://github.com/jonaspleyer/cellular_raza/commit/66bdd09bb8d358875d57a2c1406479b0d7d51c1e)
