+++ cellular_raza-building-blocks/Cargo.toml
@@ -20,8 +20,8 @@ all-features = true
 plotters = { workspace = true }
 num = { workspace = true }
 rand = { workspace = true, features=["small_rng"] }
-rand_distr = { workspace = true, features = ["serde1"]}
-rand_chacha = { workspace = true, features=["serde1"]}
+rand_distr = { workspace = true, features = ["serde"]}
+rand_chacha = { workspace = true, features=["serde"]}
 nalgebra.workspace = true
 serde = { workspace = true }
 itertools = { workspace = true }
