+++ cellular_raza-examples/pool_model_pyo3/Cargo.toml
@@ -9,7 +9,7 @@ name = "cr_pool_model_pyo3"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = { version = "0.22", features = ["extension-module"] }
+pyo3 = { workspace = true, features = ["extension-module"] }
 serde = { workspace = true, features=["rc"] }
 rand = { workspace = true, features=["small_rng"] }
 rand_chacha = { workspace = true }
