### 2025-03-18 de5ec3d
use new api with random_... instead of old gen_ specified by rand
