### 2025-03-18 6577baf
update dependencies to fit to workspace
