+++ cellular_raza-examples/bacterial_rods/Cargo.toml
@@ -5,12 +5,11 @@ edition = "2021"
 authors = ["Jonas Pleyer <jonas.pleyer@fdm.uni-freiburg.de>"]
 
 [dependencies]
-serde = { workspace = true }
+serde.workspace = true
 rand = { workspace = true, features=["small_rng"] }
-rand_chacha = { workspace = true }
+rand_chacha.workspace = true
+rand_distr.workspace = true
 nalgebra = { workspace = true }
 cellular_raza = { path="../../cellular_raza", features=["tracing"] }
 num = { workspace = true }
 itertools.workspace = true
-rand_distr = "0.4.3"
-
