### 2025-03-19 #7723 @jonaspleyer
I have included the updated figure which has been generated in issue https://github.com/jonaspleyer/cellular_raza/issues/4 in the paper.
