### 2025-03-24 a17a1b8
correct type conversions
