+++ Cargo.toml
@@ -14,6 +14,7 @@ resolver = "2"
 
 [patch.crates-io]
 nalgebra = { git = "https://github.com/dimforge/nalgebra", ref="0e8aed8809999313f0d964341e7dd83b6f1f742d" }
+ron = { git = "https://github.com/ron-rs/ron/" }
 
 [workspace.package]
 version = "0.2.3"
