+++ cellular_raza-core/Cargo.toml
@@ -35,7 +35,7 @@ bincode = { version = "2.0", features = ["serde"] }
 serde = { workspace = true }
 quick-xml = { version="0.37", features=["serialize"]}
 serde_json = { version="1.0" }
-ron = "0.8"
+ron = "0.9"
 sled = { version="0.34" }
 chrono = { version = "0.4.31", optional = true }
 tracing = { version = "0.1.40", optional = true }
