### 2025-03-24 6723a9a
simplify and reduce boilerplate
