### 2025-03-26 #7723 @mahfuz05062
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/14075810722?check_suite_focus=true).

Hi @jonaspleyer , can you check why the generate pdf is giving error. It says an image is missing. Thanks!
