### 2025-03-26 #7723 @jonaspleyer
> Hi [@jonaspleyer](https://github.com/jonaspleyer) , can you check why the generate pdf is giving error. It says an image is missing. Thanks!

Yes the image was indeed missing. I fixed it.
