### 2025-03-27 b07fa9c
test out spatial decomposition algorithm
