### 2025-03-27 #7723 @jonaspleyer
- Our Institute has very recently decided to rename itself. I have updated the name accordingly.
  - The relrease which should be used is `0.2.3`. This includes all changes made during the review process.
  - The release has been published on [crates.io](https://crates.io/crates/cellular_raza)
  - Version `0.2.3` has been uploaded to Zenodo https://zenodo.org/records/15095966 and is available under the GPLv2.0 License

I hope that this addresses all requirements. Let me if there is still anything else missing.
