### 2025-04-02 2829e6e
include macos-ci in README.md
