### 2025-04-02 403ba79
fixup! include macos-ci in README.md
