### 2025-04-08 b5f1ee0
add new function to clone new StorageInstance from existing one
