### 2025-04-08 5f4f186
add new function to clone new StorageInstance from existing one
