+++ cellular_raza-core/src/storage/serde_json.rs
@@ -63,4 +63,13 @@ impl<Id, Element> StorageInterfaceOpen for JsonStorageInterface<Id, Element> {
             phantom_element: PhantomData,
         })
     }
+
+    fn clone_to_new_instance(&self, storage_instance: u64) -> Self {
+        Self {
+            path: self.path.clone(),
+            storage_instance,
+            phantom_id: PhantomData,
+            phantom_element: PhantomData,
+        }
+    }
 }
