### 2025-04-08 83f9e49
use Arc<Mute<T>> guard to share data in memory storage
