### 2025-04-08 af5b45f
add new error for poisoned lock
