### 2025-04-08 1108c6e
use clone_to_new_instance method; reorder statements to make sense when parallelizing
