### 2025-04-08 6d61654
introduce gravity which is only relevant in 3D
