### 2025-04-08 202fb0b
set gravity to 0 in example
