+++ Cargo.toml
@@ -14,7 +14,6 @@ resolver = "2"
 
 [patch.crates-io]
 nalgebra = { git = "https://github.com/dimforge/nalgebra" }
-ron = { git = "https://github.com/ron-rs/ron/" }
 
 [workspace.package]
 version = "0.2.3"
