### 2025-04-08 d5ff9aa
activate extension module for pool_model_pyo3
