### 2025-04-08 1ecf795
do not test template-pyo3 for macos
