+++ .github/workflows/reuse.yml
@@ -33,5 +33,9 @@ jobs:
       - name: DocTest
         run: cargo test --doc
 
+      - name: Test MacOs
+        if: ${{ contains(inputs.os, 'macos') }}
+        run: cargo test --workspace --exclude cellular_raza_template_pyo3
       - name: Test
+        if: ${{ !contains(inputs.os, 'macos') }}
         run: cargo test
