### 2025-04-08 5fe476e
test two ubuntu versions instead of only latest
