+++ .github/workflows/test_beta_ubuntu-22.04.yml
@@ -1,0 +1,10 @@
+on: [push, pull_request]
+
+name: Test-Suite beta ubuntu-22.04
+
+jobs:
+  CI-beta-ubuntu-22.04:
+    uses: ./.github/workflows/reuse.yml
+    with:
+      toolchain: beta
+      os: ubuntu-22.04
