+++ .github/workflows/test_nightly_ubuntu-24.04.yml
@@ -1,10 +1,10 @@
 on: [push, pull_request]
 
-name: Test-Suite nightly ubuntu-latest
+name: Test-Suite nightly ubuntu-24.04
 
 jobs:
-  CI-nightly-ubuntu-latest:
+  CI-nightly-ubuntu-24.04:
     uses: ./.github/workflows/reuse.yml
     with:
       toolchain: nightly
-      os: ubuntu-latest
+      os: ubuntu-24.04
