+++ .github/workflows/test_stable_ubuntu-24.04.yml
@@ -1,0 +1,10 @@
+on: [push, pull_request]
+
+name: Test-Suite stable ubuntu-24.04
+
+jobs:
+  CI-stable-ubuntu-24.04:
+    uses: ./.github/workflows/reuse.yml
+    with:
+      toolchain: stable
+      os: ubuntu-24.04
