+++ .github/workflows/test_stable_ubuntu-22.04.yml
@@ -1,10 +1,10 @@
 on: [push, pull_request]
 
-name: Test-Suite stable ubuntu-latest
+name: Test-Suite stable ubuntu-22.04
 
 jobs:
-  CI-stable-ubuntu-latest:
+  CI-stable-ubuntu-22.04:
     uses: ./.github/workflows/reuse.yml
     with:
       toolchain: stable
-      os: ubuntu-latest
+      os: ubuntu-22.04
