+++ .github/workflows/test_nightly_ubuntu-22.04.yml
@@ -1,0 +1,10 @@
+on: [push, pull_request]
+
+name: Test-Suite nightly ubuntu-22.04
+
+jobs:
+  CI-nightly-ubuntu-22.04:
+    uses: ./.github/workflows/reuse.yml
+    with:
+      toolchain: nightly
+      os: ubuntu-22.04
