### 2025-04-08 fa2a885
adjust links for ubuntu
