+++ .github/workflows/test_beta_ubuntu-24.04.yml
@@ -3,7 +3,7 @@ on: [push, pull_request]
 name: Test-Suite beta ubuntu-24.04
 
 jobs:
-  CI-beta-ubuntu-24.04:
+  CI-beta-ubuntu-24_04:
     uses: ./.github/workflows/reuse.yml
     with:
       toolchain: beta
