+++ .github/workflows/test_nightly_ubuntu-22.04.yml
@@ -3,7 +3,7 @@ on: [push, pull_request]
 name: Test-Suite nightly ubuntu-22.04
 
 jobs:
-  CI-nightly-ubuntu-22.04:
+  CI-nightly-ubuntu-22_04:
     uses: ./.github/workflows/reuse.yml
     with:
       toolchain: nightly
