### 2025-04-08 ef780c8
also exclude pool model for testing in macos
