+++ .github/workflows/reuse.yml
@@ -35,7 +35,7 @@ jobs:
 
       - name: Test MacOs
         if: ${{ contains(inputs.os, 'macos') }}
-        run: cargo test --workspace --exclude cellular_raza-template-pyo3
+        run: cargo test --workspace --exclude cellular_raza-template-pyo3 --exclude cr_pool_model_pyo3
       - name: Test
         if: ${{ !contains(inputs.os, 'macos') }}
         run: cargo test
