### 2025-04-08 7306ba8
try patched pathfinder crate
