+++ Cargo.toml
@@ -14,6 +14,7 @@ resolver = "2"
 
 [patch.crates-io]
 nalgebra = { git = "https://github.com/dimforge/nalgebra" }
+pathfinder_simd = { git = "https://github.com/theoparis/pathfinder.git" }
 
 [workspace.package]
 version = "0.2.3"
