+++ Cargo.toml
@@ -14,7 +14,7 @@ resolver = "2"
 
 [patch.crates-io]
 nalgebra = { git = "https://github.com/dimforge/nalgebra" }
-pathfinder_simd = { git = "https://github.com/theoparis/pathfinder.git" }
+pathfinder_simd = { git = "https://github.com/servo/pathfinder.git" }
 
 [workspace.package]
 version = "0.2.3"
