### 2025-04-08 827e195
bumpto new cudarc version; also fix some allocation issues
