+++ cellular_raza-core/src/backend/cara/mod.rs
@@ -13,6 +13,14 @@ struct Runner {
     forces: *mut f32,
 }
 
+impl Drop for Runner {
+    fn drop(&mut self) {
+        unsafe {
+            drop_runner(self);
+        }
+    }
+}
+
 type UID = u64;
 
 #[repr(C)]
@@ -34,7 +42,6 @@ struct Force {
 }
 
 #[repr(C)]
-#[derive(Clone, Copy, Debug)]
 struct Agent {
     id: UID,
     position: Position,
@@ -42,12 +49,13 @@ struct Agent {
 }
 
 #[repr(C)]
-#[derive(Clone, Copy, Debug)]
 struct AgentContainer {
     agent: Agent,
 }
 
 unsafe extern "C" {
+    fn new_runner(agents: *const Agent, n_agents: usize) -> *mut Runner;
+    fn drop_runner(runner: *mut Runner);
     fn do_compute();
     fn update_positions(runner: &mut Runner);
     fn print_positions(runner: &Runner);
@@ -56,17 +64,16 @@ unsafe extern "C" {
 /// Empty main routine to showcase how to include cuda files
 pub fn run() {
     let n_agents = 10;
-    let mut agents: Vec<_> = (0..n_agents)
-        .map(|n| AgentContainer {
-            agent: Agent {
-                id: n,
-                position: Position { p: [n as f32; 3] },
-                velocity: Velocity { v: [0.; 3] },
-            },
+    let agents: Vec<_> = (0..n_agents)
+        .map(|n| Agent {
+            id: n,
+            position: Position { p: [n as f32; 3] },
+            velocity: Velocity { v: [0.; 3] },
         })
         .collect();
 
-    let mut runner = Runner {
+    let runner = unsafe { new_runner(agents.as_ptr(), agents.len()) };
+    /* let mut runner = Runner {
         uid_counter: 0,
         n_agents,
         agents_up_to_date: false,
@@ -75,11 +82,11 @@ pub fn run() {
         positions: &mut 0.0,
         velocities: &mut 0.0,
         forces: &mut 0.0,
-    };
+    };*/
 
     unsafe {
-        update_positions(&mut runner);
-        print_positions(&runner);
+        update_positions(&mut *runner);
+        print_positions(&*runner);
         do_compute();
     }
 }
@@ -91,22 +98,33 @@ fn test_main() {
 
 #[test]
 fn test_cudarc() -> Result<(), Box<dyn std::error::Error>> {
-    use cudarc::driver::LaunchAsync;
+    use cudarc::driver::*;
     const SIN_KERNEL: &str = include_str!("kernel.cu");
-    let dev = cudarc::driver::CudaDevice::new(0)?;
+    // Get a stream for GPU 0
+    let ctx = cudarc::driver::CudaContext::new(0)?;
+    let stream = ctx.default_stream();
 
-    // allocate buffers
-    let inp = dev.htod_copy(vec![1.0f32; 100])?;
-    let mut out = dev.alloc_zeros::<f32>(100)?;
+    // copy a rust slice to the device
+    let inp = stream.memcpy_stod(&[1.0f32; 100])?;
+
+    // or allocate directly
+    let mut out = stream.alloc_zeros::<f32>(100)?;
+
     let ptx = cudarc::nvrtc::compile_ptx(SIN_KERNEL)?;
-    dev.load_ptx(ptx, "my_module", &["sin_kernel"])?;
-    let sin_kernel = dev.get_func("my_module", "sin_kernel").unwrap();
-    let cfg = cudarc::driver::LaunchConfig::for_num_elems(100);
[…]
