+++ cellular_raza-core/src/backend/cara/cara.cu
@@ -72,6 +72,35 @@ __global__ void calculate_forces(int total, float *pos1, float *vel1, float *pos
     forces[j] += 0.1 * (p2 - p1);
 }
 
+extern "C" Runner *new_runner(Agent *agents, uint64_t n_agents) {
+
+    // float positions[3 * n_agents] = malloc(3 * n_agents * sizeof(float));
+    float *positions, *velocities, *forces;
+    positions  = (float *) malloc(n_agents * sizeof(float));
+    velocities = (float *) malloc(n_agents * sizeof(float));
+    forces     = (float *) malloc(n_agents * sizeof(float));
+
+    AgentContainer *agent_containers;
+    agent_containers = (AgentContainer *) malloc(n_agents * sizeof(AgentContainer));
+    for (uint64_t k = 0; k < n_agents; k++) {
+        agent_containers[k] = AgentContainer{
+            .agent = agents[k],
+        };
+    }
+    struct Runner runner{
+        n_agents, n_agents, true, agent_containers, positions, velocities, forces,
+    };
+    struct Runner *runner_ptr = &runner;
+    return runner_ptr;
+}
+
+extern "C" void drop_runner(Runner *runner) {
+    free(runner->agents);
+    free(runner->positions);
+    free(runner->velocities);
+    free(runner->forces);
+}
+
 /// This calculates with time complexity O(n^2) which is not ideal but the most simple thing that
 // one can write down so we will stick with it for now.
 extern "C" void update_positions(Runner &runner) {
@@ -84,17 +113,18 @@ extern "C" void update_positions(Runner &runner) {
             /* float v2[3] = {runner.velocities[j], runner.velocities[j + 1],
                            runner.velocities[j + 2]};*/
 
-            float f[3];
+            float f[3] = {0.0};
             for (uint8_t k = 0; k < 3; k++) {
                 // Calculate the force between agents
                 f[k] += 0.1 * (p1 - p2);
 
                 // Update forces
-                runner.forces[i + k] += f[k];
-                runner.forces[j + k] -= f[k];
+                // runner.forces[i + k] += f[k];
+                // runner.forces[j + k] -= f[k];
             }
         }
     }
+    return;
 }
 
 extern "C" void print_positions(Runner &runner) {
@@ -102,8 +132,8 @@ extern "C" void print_positions(Runner &runner) {
         Agent agent = runner.agents[i].agent;
         Position p  = agent.position;
         Velocity v  = agent.velocity;
-        printf("%lu p=[%f, %f, %f] v=[%f, %f, %f]\n", agent.id, p.p[0], p.p[1], p.p[2], v.v[0],
-               v.v[1], v.v[2]);
+        // printf("%lu p=[%f, %f, %f] v=[%f, %f, %f]\n", agent.id, p.p[0], p.p[1], p.p[2], v.v[0],
+        //        v.v[1], v.v[2]);
     }
 }
 
@@ -115,7 +145,7 @@ __global__ void saxpy(int n, float a, float *x, float *y) {
 }
 
 /// This is a comment
-extern "C" void do_compute(void) {
+extern "C" void do_compute() {
     int N = 1 << 20;
     float *x, *y, *d_x, *d_y;
     x = (float *) malloc(N * sizeof(float));
@@ -140,8 +170,8 @@ extern "C" void do_compute(void) {
     float maxError = 0.0f;
     for (int i = 0; i < N; i++) {
         maxError = abs(y[i] - 4.0f);
-        printf("%f", maxError);
-        assert(maxError < 0.001);
+        // printf("%f\n", maxError);
+        // assert(maxError < 0.001);
     }
 
     cudaFree(d_x);
