### 2025-04-08 ce41a8d
cleanup of cara sample code
