### 2025-04-14 c8314e4
Update wgpu requirement from 24.0 to 25.0
Updates the requirements on [wgpu](https://github.com/gfx-rs/wgpu) to permit the latest version.
- [Release notes](https://github.com/gfx-rs/wgpu/releases)
- [Changelog](https://github.com/gfx-rs/wgpu/blob/trunk/CHANGELOG.md)
- [Commits](https://github.com/gfx-rs/wgpu/compare/wgpu-v24.0.1...v25.0.0)

---
updated-dependencies:
- dependency-name: wgpu
  dependency-version: 25.0.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
