### 2025-04-14 a2682a9
Update cudarc requirement from 0.14 to 0.15
Updates the requirements on [cudarc](https://github.com/coreylowman/cudarc) to permit the latest version.
- [Release notes](https://github.com/coreylowman/cudarc/releases)
- [Commits](https://github.com/coreylowman/cudarc/compare/v0.14.0...v0.15.1)

---
updated-dependencies:
- dependency-name: cudarc
  dependency-version: 0.15.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
