+++ cellular_raza-core/Cargo.toml
@@ -49,7 +49,7 @@ circ_buffer = { version = "0.1.8", features = ["serde"] }
 uniquevec = { version = "0.1.0", features = ["serde"] }
 
 # Additional dependencies for cara backend
-cudarc = { version = "0.14", optional = true, features = ["cuda-12080"] }
+cudarc = { version = "0.15", optional = true, features = ["cuda-12080"] }
 
 # Additional dependencies for elli backend
 wgpu = { version = "24.0", optional = true }
