### 2025-04-15 7236589
implement neighbor interactions for RodMechanics
