### 2025-04-16 51cb0a3
include surface friction in bacterial rods domain
