+++ cellular_raza-building-blocks/src/cell_building_blocks/bacterial_rods.rs
@@ -315,6 +315,10 @@ pub struct CartesianCuboidRods<F, const D: usize> {
     /// Gravitational force which is only relevant for 3D simulations and always acts with constant
     /// force downwards (negative z-direction).
     pub gravity: F,
+    /// Computes friction at all surfaces of the box
+    pub surface_friction: F,
+    /// The distance for which the friction will be applied
+    pub surface_friction_distance: F,
 }
 
 impl<C, F, const D: usize> SortCells<C> for CartesianCuboidRods<F, D>
@@ -368,6 +372,8 @@ where
                     CartesianSubDomainRods::<F, D> {
                         subdomain,
                         gravity: self.gravity,
+                        surface_friction: self.surface_friction,
+                        surface_friction_distance: self.surface_friction,
                     },
                     voxels,
                 )
@@ -383,6 +389,10 @@ pub struct CartesianSubDomainRods<F, const D: usize> {
     pub subdomain: CartesianSubDomain<F, D>,
     /// See [CartesianCuboidRods]
     pub gravity: F,
+    /// Computes friction at all surfaces of the box
+    pub surface_friction: F,
+    /// The distance for which the friction will be applied
+    pub surface_friction_distance: F,
 }
 
 impl<F>
@@ -397,14 +407,29 @@ where
     fn calculate_custom_force(
         &self,
         pos: &Matrix<F, Dyn, Const<3>, VecStorage<F, Dyn, Const<3>>>,
-        _: &Matrix<F, Dyn, Const<3>, VecStorage<F, Dyn, Const<3>>>,
+        vel: &Matrix<F, Dyn, Const<3>, VecStorage<F, Dyn, Const<3>>>,
     ) -> Result<
         Matrix<F, Dyn, Const<3>, VecStorage<F, Dyn, Const<3>>>,
         cellular_raza_concepts::CalcError,
     > {
-        Ok(nalgebra::MatrixXx3::from_fn(pos.nrows(), |_, m| {
+        use core::ops::AddAssign;
+        let mut force = nalgebra::MatrixXx3::from_fn(pos.nrows(), |_, m| {
             if m == 2 { -self.gravity } else { F::zero() }
-        }))
+        });
+        for (i, (p, v)) in pos.row_iter().zip(vel.row_iter()).enumerate() {
+            let d1 = (p.transpose() - self.subdomain.domain_min)
+                .map(|x| <F as num::Float>::abs(x) <= self.surface_friction_distance);
+            let d2 = (p.transpose() - self.subdomain.domain_max)
+                .map(|x| <F as num::Float>::abs(x) <= self.surface_friction_distance);
+            let q = v.norm();
+            if q != F::zero() && d1.zip_map(&d2, |x, y| x || y).into_iter().any(|x| *x) {
+                let dir = v / q;
+                force
+                    .row_mut(i)
+                    .add_assign(-dir * self.gravity * self.surface_friction);
+            }
+        }
+        Ok(force)
     }
 }
 
@@ -475,6 +500,8 @@ where
 {
     subdomain: CartesianSubDomain<F, D2>,
     gravity: F,
+    surface_friction: F,
+    surface_friction_distance: F,
 }
 
 impl<F, const D: usize> From<__CartesianSubDomainRodsSerde<F, D>> for CartesianSubDomainRods<F, D>
@@ -485,6 +512,8 @@ where
         CartesianSubDomainRods {
             subdomain: s.subdomain,
             gravity: s.gravity,
+            surface_friction: s.surface_friction,
+            surface_friction_distance: s.surface_friction_distance,
         }
     }
 }
