### 2025-04-16 6deff84
disable surface friction for example
