### 2025-04-21 #7723 @mahfuz05062
> [@mahfuz05062](https://github.com/mahfuz05062) Do you need anything else from my side? Did the above points fulfill the tasks you posted?

Hi @jonaspleyer, sorry, for the late reply as I was out for travel last week. I have gone through the article and have some comments that I would like you to address. These are mainly grammatical/typo etc. so should be straightforward. Please let me know if you have any questions. Thanks!

- In statement of need, "These models live on the esoscopic scale where the underlying complexity of the problem can often neither be fully attributed to intracellular or extracellular processes." - Do you mean 'nor' here? I think having neither just before 'intracellular' would make it more clear.
- In line 25, you mention "While these tools", however, you haven't described any tools in the first line. I would suggest to clarify what tool you refer to here.
- ", if desired" seems unnecessary in line 50.
- I see extra spaces before "." on line 51 and line 54. Please remove those spaces.
- Line 57 - should be from building to running to visualizing for consistency.
- Line 59 - examples "of"
- Line 63 - is there any reason to capitalize 'I' in Interaction?
- This is for the future, but I think the shades of blue and green are sometimes difficult to separate. Use of another contrasting color for dormant may make it better.
