### 2025-04-21 #7723 @jonaspleyer
> In statement of need, "These models live on the esoscopic scale where the underlying complexity of the problem can often neither be fully attributed to intracellular or extracellular processes." - Do you mean 'nor' here? I think having neither just before 'intracellular' would make it more clear.

I rewrote the sentence such that it now reads:
"These models live on the mesoscopic scale where the underlying complexity of the problem can often not be fully attributed to neither intracellular nor extracellular processes."

---

> In line 25, you mention "While these tools", however, you haven't described any tools in the first line. I would suggest to clarify what tool you refer to here.

This sentence only makes sense partially. It is supposed to refer to the tools which are used within the citations. I have inserted another short explanatory sentence which references two reviews which discuss the role of these models.

"Agent-based models have become popular in cellular biology [@Mogilner2016; @Cess2022; @Delile2017; @Delile_Herrmann_Peyrieras_Doursat_2017]. Many tools have emerged which are capable of describing cellular systems in various details [@Pleyer2023; @Abar2017]."

---

> ", if desired" seems unnecessary in line 50.

I have removed it.

---

> I see extra spaces before "." on line 51 and line 54. Please remove those spaces.

I have removed the spaces.

---

> Line 57 - should be from building to running to visualizing for consistency.

I have rewritten it to:
"from building to running and visualizing."
which is slightly different to what you suggested.

---

> Line 63 - is there any reason to capitalize 'I' in Interaction?

Fixed it.

---

> This is for the future, but I think the shades of blue and green are sometimes difficult to separate. Use of another contrasting color for dormant may make it better.

Thanks for the suggestion. I will keep this in mind for any coming work.
