### 2025-04-24 ebab7f6
switch sections; remove "State of Field" heading
