### 2025-04-25 6727dd5
use workspace ndarray
