+++ Cargo.toml
@@ -45,5 +45,6 @@ rand_chacha = { version="0.9", features=["serde"] }
 serde = "1.0"
 tempfile = "3"
 nalgebra = { version="<0.34, >=0.30", features=["serde-serialize", "rand"]}
+ndarray = { version = "0.16", features = ["blas", "serde", "serde-1"] }
 pyo3 = { version = "0.24", features = ["serde", "py-clone"]}
 kdam = "0.6"
