### 2025-04-25 13686dc
allow dynamically-sized vectors for reactions
