### 2025-04-25 eea08de
move subdomain to building blocks
