### 2025-04-25 c54d237
fixup! use workspace ndarray
