### 2025-04-25 e2bda08
Remove implementations for single dimension case
