+++ cellular_raza-building-blocks/src/domains/cartesian_cuboid_n.rs
@@ -1,13 +1,6 @@
 // Imports from this crate
 use cellular_raza_concepts::*;
 
-#[cfg(feature = "pyo3")]
-use pyo3::prelude::*;
-
-// Imports from std and core
-use core::cmp::{max, min};
-use std::usize;
-
 // Imports from other crates
 use itertools::Itertools;
 use nalgebra::SVector;
@@ -805,521 +798,6 @@ impl<F, const D: usize> SubDomain for CartesianSubDomain<F, D> {
             .collect()
     }
 }
-
-macro_rules! implement_cartesian_cuboid_domain {
-    (
-        $d: literal,
-        $domain_name: ident,
-        $subdomain_name: ident,
-        $voxel_name: ident,
-        $float_type: ty,
-        $($k: expr),+
-    ) => {
-        #[derive(Clone, Debug, Deserialize, Serialize)]
-        #[cfg_attr(feature = "pyo3", pyclass)]
-        #[cfg_attr(feature = "pyo3", pyo3(get_all, set_all))]
-        /// Cartesian cuboid in
-        #[doc = concat!(" `", stringify!($d), "D`")]
-        /// with float type
-        #[doc = concat!(" `", stringify!($float_type), "`")]
-        pub struct $domain_name {
-            /// Lower boundary of domain
-            pub min: [$float_type; $d],
-            /// Upper boundary of domain
-            pub max: [$float_type; $d],
-            /// Number of voxels in domain along axes
-            pub n_voxels: [i64; $d],
-            /// Length of individual voxels in domain
-            pub dx_voxels: [$float_type; $d],
-            /// Initial seed from which to generate seeds for voxels
-            pub rng_seed: u64,
-        }
-
-        impl $domain_name {
-            fn check_min_max(min: [$float_type; $d], max: [$float_type; $d]) -> Result<(), CalcError> {
-                for i in 0..$d {
-                    match max[i] > min[i] {
-                        false => Err(CalcError(format!(
-                            "Min {:?} must be smaller than Max {:?} for domain boundaries!",
-                            min,
-                            max
-                        ))),
-                        true => Ok(()),
-                    }?;
-                }
-                Ok(())
-            }
-
-            fn check_positive<F>(interaction_ranges: [F; $d]) -> Result<(), CalcError>
-            where
-                F: PartialOrd + num::Zero + core::fmt::Debug,
-            {
-                for i in 0..$d {
-                    match interaction_ranges[i] > F::zero() {
-                        false => Err(CalcError(format!("Interaction range must be positive and non-negative! Got value {:?}", interaction_ranges[i]))),
-                        true => Ok(())
-                    }?;
-                }
-                Ok(())
-            }
-
-            /// Construct the domain from given lower/upper boundaries and maximum
-            /// length of interaction ranges along axes.
-            pub fn from_boundaries_and_interaction_ranges(
-                min: [$float_type; $d],
-                max: [$float_type; $d],
-                interaction_ranges: [$float_type; $d]
-            ) -> Result<$domain_name, CalcError>
-            {
-                Self::check_min_max(min, max)?;
-                Self::check_positive(interaction_ranges)?;
-                let mut n_voxels = [0; $d];
-                let mut dx_voxels = [0.0; $d];
-                for i in 0..$d {
-                    n_voxels[i] = ((max[i] - min[i]) / interaction_ranges[i] * 0.5).ceil() as i64;
-                    dx_voxels[i] = (max[i]-min[i])/n_voxels[i] as $float_type;
-                }
-                Ok(Self {
-                    min,
-                    max,
-                    n_voxels,
-                    dx_voxels,
-                    rng_seed: 0,
-                })
-            }
-
[…]
