### 2025-04-25 86e0d86
add docstrings
