+++ cellular_raza-building-blocks/src/domains/cartesian_2d_diffusion.rs
@@ -1,10 +1,17 @@
 type ReactionVector = nalgebra::DVector<f32>;
 
-use cellular_raza::prelude::*;
+use cellular_raza_concepts::*;
 use serde::{Deserialize, Serialize};
 
+use super::{CartesianCuboid, CartesianSubDomain};
+
+/// Domain based on [CartesianCuboid] in 2 dimensions with extracellular diffusion.
+///
+/// We solve the equations
+///
 #[derive(Domain, Clone)]
 pub struct CartesianDiffusion2D {
+    /// See [CartesianCuboid]
     #[DomainRngSeed]
     #[DomainPartialDerive]
     #[SortCells]
@@ -12,29 +19,36 @@ pub struct CartesianDiffusion2D {
     /// The discretization must be a multiple of the voxel size.
     /// This quantity will be used as an initial estimate and rounded to the nearest candidate.
     pub reactions_dx: nalgebra::Vector2<f32>,
+    /// Diffusion constant
     pub diffusion_constant: f32,
+    /// Initial values which are uniform over the simulation domain
     pub initial_value: ReactionVector,
 }
 
+/// Subdomain for the [CartesianDiffusion2D] domain.
 #[derive(Deserialize, SubDomain, Clone, Serialize, Debug)]
 pub struct CartesianDiffusion2DSubDomain {
+    /// See [CartesianSubDomain]
     #[Base]
     #[SortCells]
     #[Mechanics]
-    pub subdomain: CartesianSubDomain<f32, 2>,
-    pub reactions_min: nalgebra::Vector2<f32>,
-    pub reactions_max: nalgebra::Vector2<f32>,
-    pub reactions_dx: nalgebra::Vector2<f32>,
-    pub index_min: nalgebra::Vector2<usize>,
-    pub index_max: nalgebra::Vector2<usize>,
-    pub extracellular: ndarray::Array3<f32>,
-    pub ownership_array: ndarray::Array2<bool>,
+    subdomain: CartesianSubDomain<f32, 2>,
+    reactions_min: nalgebra::Vector2<f32>,
+    reactions_max: nalgebra::Vector2<f32>,
+    reactions_dx: nalgebra::Vector2<f32>,
+    index_min: nalgebra::Vector2<usize>,
+    index_max: nalgebra::Vector2<usize>,
+    extracellular: ndarray::Array3<f32>,
+    ownership_array: ndarray::Array2<bool>,
+    /// Diffusion constant
     pub diffusion_constant: f32,
     increments: [ndarray::Array3<f32>; 3],
     increments_start: usize,
     helper: ndarray::Array3<f32>,
 }
 
+/// This type is not intendend for direct usage
+#[doc(hidden)]
 pub struct BorderInfo {
     min_sent: nalgebra::Vector2<usize>,
     max_sent: nalgebra::Vector2<usize>,
@@ -101,12 +115,14 @@ impl CartesianDiffusion2DSubDomain {
         )
         .for_each(|mut w, v, t| {
             if let (false, Some(vi)) = (*t, v) {
-                w.assign(&vi);
+                w.assign(vi);
             }
         });
     }
 }
 
+/// This type is not intendend for direct usage
+#[doc(hidden)]
 pub struct NeighborValue {
     min: nalgebra::Vector2<usize>,
     max: nalgebra::Vector2<usize>,
@@ -119,11 +135,7 @@ impl SubDomainReactions<nalgebra::SVector<f32, 2>, ReactionVector, f32>
     type BorderInfo = BorderInfo;
     type NeighborValue = NeighborValue;
 
-    fn treat_increments<I, J>(
-        &mut self,
-        neighbors: I,
-        sources: J,
[…]
