+++ cellular_raza/tests/brownian_diffusion_constant_approx.rs
@@ -123,11 +123,11 @@ fn analyze_positions_brownian<const D: usize>(
 }
 
 macro_rules! test_brownian {
-    ($parameters:ident, $domain_name:ident, $particle_name:ident, $d:literal) => {{
+    ($parameters:ident, $particle_name:ident, $d:literal) => {{
         let domain_size = $parameters.domain_size;
         assert!(domain_size > 0.0);
         let mut domain =
-            $domain_name::from_boundaries_and_n_voxels([0.0; $d], [domain_size; $d], [3; $d])?;
+            CartesianCuboid::from_boundaries_and_n_voxels([0.0; $d], [domain_size; $d], [3; $d])?;
         domain.rng_seed = $parameters.random_seed;
 
         let initial_position = nalgebra::SVector::from([domain_size / 2.0; $d]);
@@ -173,7 +173,7 @@ fn diffusion_constant_brownian_3d_1() -> Result<(), Box<dyn std::error::Error>> 
     parameters.storage_name = "out/brownian_3d_1".into();
     parameters.diffusion_constant = 1.0;
     parameters.random_seed = 1;
-    test_brownian!(parameters, CartesianCuboid3New, Brownian3D, 3)?;
+    test_brownian!(parameters, Brownian3D, 3)?;
     Ok(())
 }
 
@@ -183,7 +183,7 @@ fn diffusion_constant_brownian_3d_2() -> Result<(), Box<dyn std::error::Error>> 
     parameters.diffusion_constant = 0.5;
     parameters.storage_name = "out/brownian_3d_2".into();
     parameters.random_seed = 2;
-    test_brownian!(parameters, CartesianCuboid3New, Brownian3D, 3)?;
+    test_brownian!(parameters, Brownian3D, 3)?;
     Ok(())
 }
 
@@ -193,7 +193,7 @@ fn diffusion_constant_brownian_3d_3() -> Result<(), Box<dyn std::error::Error>> 
     parameters.diffusion_constant = 0.25;
     parameters.storage_name = "out/brownian_3d_3".into();
     parameters.random_seed = 3;
-    test_brownian!(parameters, CartesianCuboid3New, Brownian3D, 3)?;
+    test_brownian!(parameters, Brownian3D, 3)?;
     Ok(())
 }
 
@@ -203,7 +203,7 @@ fn diffusion_constant_brownian_2d_1() -> Result<(), Box<dyn std::error::Error>> 
     parameters.storage_name = "out/brownian_2d_1".into();
     parameters.diffusion_constant = 1.0;
     parameters.random_seed = 4;
-    test_brownian!(parameters, CartesianCuboid2New, Brownian2D, 2)?;
+    test_brownian!(parameters, Brownian2D, 2)?;
     Ok(())
 }
 
@@ -213,7 +213,7 @@ fn diffusion_constant_brownian_2d_2() -> Result<(), Box<dyn std::error::Error>> 
     parameters.diffusion_constant = 0.5;
     parameters.storage_name = "out/brownian_2d_2".into();
     parameters.random_seed = 5;
-    test_brownian!(parameters, CartesianCuboid2New, Brownian2D, 2)?;
+    test_brownian!(parameters, Brownian2D, 2)?;
     Ok(())
 }
 
@@ -223,7 +223,7 @@ fn diffusion_constant_brownian_2d_3() -> Result<(), Box<dyn std::error::Error>> 
     parameters.diffusion_constant = 0.25;
     parameters.storage_name = "out/brownian_2d_3".into();
     parameters.random_seed = 6;
[…]
