### 2025-04-25 bde1c34
migrate to generic CartesianCuboid
