### 2025-04-25 713050e
remove boilerplate for serialization and introduce manual trait bounds instead
