+++ cellular_raza-building-blocks/src/domains/cartesian_cuboid_n.rs
@@ -5,7 +5,7 @@ use cellular_raza_concepts::*;
 use itertools::Itertools;
 use nalgebra::SVector;
 
-use serde::{Deserialize, Deserializer, Serialize, ser::SerializeStruct};
+use serde::{Deserialize, Serialize};
 
 /// Helper function to calculate the decomposition of a large number N into n as evenly-sizedchunks
 /// chunks as possible
@@ -403,7 +403,16 @@ fn generate_subdomains() {
 }
 
 /// Subdomain corresponding to the [CartesianCuboid] struct.
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(bound = "
+F: 'static
+    + PartialEq
+    + Clone
+    + core::fmt::Debug
+    + Serialize
+    + for<'a> Deserialize<'a>,
+[usize; D]: Serialize + for<'a> Deserialize<'a>,
+")]
 pub struct CartesianSubDomain<F, const D: usize> {
     min: SVector<F, D>,
     max: SVector<F, D>,
@@ -412,151 +421,6 @@ pub struct CartesianSubDomain<F, const D: usize> {
     pub(crate) domain_min: SVector<F, D>,
     pub(crate) domain_max: SVector<F, D>,
     domain_n_voxels: SVector<usize, D>,
-}
-
-#[derive(Deserialize)]
-#[serde(rename(serialize = "CartesianSubDomain", deserialize = "CartesianSubDomain",))]
-struct __CartesianSubDomainSerde<F: 'static + Clone + core::fmt::Debug + PartialEq, const D: usize>
-{
-    min: SVector<F, D>,
-    max: SVector<F, D>,
-    dx: SVector<F, D>,
-    voxels: Vec<SVector<usize, D>>,
-    domain_min: SVector<F, D>,
-    domain_max: SVector<F, D>,
-    domain_n_voxels: SVector<usize, D>,
-}
-
-impl<F, const D: usize> From<__CartesianSubDomainSerde<F, D>> for CartesianSubDomain<F, D>
-where
-    F: 'static + Clone + core::fmt::Debug + PartialEq,
-{
-    fn from(s: __CartesianSubDomainSerde<F, D>) -> Self {
-        CartesianSubDomain {
-            min: s.min,
-            max: s.max,
-            dx: s.dx,
-            voxels: s.voxels.into_iter().map(<[usize; D]>::from).collect(),
-            domain_min: s.domain_min,
-            domain_max: s.domain_max,
-            domain_n_voxels: s.domain_n_voxels,
-        }
-    }
-}
-
-impl<F, const D: usize> Serialize for CartesianSubDomain<F, D>
-where
-    F: nalgebra::Scalar + Serialize,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let mut state = serializer.serialize_struct("CartesianSubDomain", 7)?;
-        state.serialize_field("min", &self.min)?;
-        state.serialize_field("max", &self.max)?;
-        state.serialize_field("dx", &self.dx)?;
-        let voxels = self
-            .voxels
-            .iter()
-            .map(|ind| ind.into_iter().collect::<Vec<_>>())
-            .collect::<Vec<_>>();
-        state.serialize_field("voxels", &voxels)?;
-        state.serialize_field("domain_min", &self.domain_min)?;
-        state.serialize_field("domain_max", &self.domain_max)?;
-        state.serialize_field("domain_n_voxels", &self.domain_n_voxels)?;
-        state.end()
-    }
-}
-
[…]
