### 2025-04-25 b3ac8cf
rewrite more cleanly
