### 2025-04-25 469911b
implement AbsDiffEq and RelativeEq for some types
