+++ Cargo.toml
@@ -48,3 +48,5 @@ nalgebra = { version="<0.34, >=0.30", features=["serde-serialize", "rand"]}
 ndarray = { version = "0.16", features = ["blas", "serde", "serde-1"] }
 pyo3 = { version = "0.24", features = ["serde", "py-clone"]}
 kdam = "0.6"
+approxim = "0.6.5"
+approx = { package = "approxim", version = "0.6.5" }
