### 2025-04-25 86ec8e6
make Diffusion Domain available generically for f32 and f64
