+++ cellular_raza-building-blocks/src/domains/cartesian_2d_diffusion.rs
@@ -1,4 +1,4 @@
-type ReactionVector = nalgebra::DVector<f32>;
+type ReactionVector<F> = nalgebra::DVector<F>;
 
 use cellular_raza_concepts::*;
 use serde::{Deserialize, Serialize};
@@ -10,41 +10,48 @@ use super::{CartesianCuboid, CartesianSubDomain};
 /// We solve the equations
 ///
 #[derive(Domain, Clone)]
-pub struct CartesianDiffusion2D {
+pub struct CartesianDiffusion2D<F> {
     /// See [CartesianCuboid]
     #[DomainRngSeed]
     #[DomainPartialDerive]
     #[SortCells]
-    pub domain: CartesianCuboid<f32, 2>,
+    pub domain: CartesianCuboid<F, 2>,
     /// The discretization must be a multiple of the voxel size.
     /// This quantity will be used as an initial estimate and rounded to the nearest candidate.
-    pub reactions_dx: nalgebra::Vector2<f32>,
+    pub reactions_dx: nalgebra::Vector2<F>,
     /// Diffusion constant
-    pub diffusion_constant: f32,
+    pub diffusion_constant: F,
     /// Initial values which are uniform over the simulation domain
-    pub initial_value: ReactionVector,
+    pub initial_value: ReactionVector<F>,
 }
 
 /// Subdomain for the [CartesianDiffusion2D] domain.
-#[derive(Deserialize, SubDomain, Clone, Serialize, Debug)]
-pub struct CartesianDiffusion2DSubDomain {
+#[derive(SubDomain, Clone, Debug, Serialize, Deserialize)]
+#[serde(bound = "F:
+    'static
+    + PartialEq
+    + Clone
+    + core::fmt::Debug
+    + Serialize
+    + for<'a> Deserialize<'a>")]
+pub struct CartesianDiffusion2DSubDomain<F> {
     /// See [CartesianSubDomain]
     #[Base]
     #[SortCells]
     #[Mechanics]
-    subdomain: CartesianSubDomain<f32, 2>,
-    reactions_min: nalgebra::Vector2<f32>,
-    reactions_max: nalgebra::Vector2<f32>,
-    reactions_dx: nalgebra::Vector2<f32>,
+    subdomain: CartesianSubDomain<F, 2>,
+    reactions_min: nalgebra::Vector2<F>,
+    reactions_max: nalgebra::Vector2<F>,
+    reactions_dx: nalgebra::Vector2<F>,
     index_min: nalgebra::Vector2<usize>,
     index_max: nalgebra::Vector2<usize>,
-    extracellular: ndarray::Array3<f32>,
+    extracellular: ndarray::Array3<F>,
     ownership_array: ndarray::Array2<bool>,
     /// Diffusion constant
-    pub diffusion_constant: f32,
-    increments: [ndarray::Array3<f32>; 3],
+    pub diffusion_constant: F,
+    increments: [ndarray::Array3<F>; 3],
     increments_start: usize,
-    helper: ndarray::Array3<f32>,
+    helper: ndarray::Array3<F>,
 }
 
 /// This type is not intendend for direct usage
@@ -54,8 +61,11 @@ pub struct BorderInfo {
     max_sent: nalgebra::Vector2<usize>,
 }
 
-impl CartesianDiffusion2DSubDomain {
-    fn assign_neighbor(&mut self, neighbor: NeighborValue) {
+impl<F> CartesianDiffusion2DSubDomain<F>
+where
+    F: nalgebra::RealField + Clone,
+{
+    fn assign_neighbor(&mut self, neighbor: NeighborValue<F>) {
         use ndarray::*;
         let NeighborValue { min, max, values } = neighbor;
 
[…]
