### 2025-04-25 c0c6fac
comment out reactions max since it is not needed at this time
