### 2025-04-25 dd76efe
remove obsolete module
