+++ cellular_raza-building-blocks/src/domains/mod.rs
@@ -1,10 +1,6 @@
 mod cartesian_2d_diffusion;
 mod cartesian_cuboid_n;
 
-/// Contains deprecated cartesian cuboid implementations for an older vertex model
-// TODO #[allow(deprecated)]
-pub mod cartesian_cuboid_2_vertex_old;
-
 /// Contains deprecated cartesian cuboid implementations
 // TODO #[allow(deprecated)]
 pub mod cartesian_cuboid_n_old;
