+++ cellular_raza-building-blocks/src/domains/cartesian_cuboid_2_vertex_old.rs
@@ -1,590 +1,0 @@
-// Imports from this crate
-use cellular_raza_concepts::domain_old::*;
-use cellular_raza_concepts::reactions_old::Volume;
-use cellular_raza_concepts::{
-    BoundaryError, CalcError, CreatePlottingRoot, DrawingError, IndexError, RequestError,
-};
-
-use super::cartesian_cuboid_n::get_decomp_res;
-
-// Imports from std and core
-use core::cmp::{max, min};
-use nalgebra::SVector;
-
-// Imports from other crates
-use itertools::Itertools;
-
-use serde::{Deserialize, Serialize};
-
-use plotters::backend::BitMapBackend;
-use plotters::coord::cartesian::Cartesian2d;
-use plotters::coord::types::RangedCoordf64;
-use plotters::prelude::DrawingArea;
-
-/// Cuboid Domain with coordinates specialized for vertex systems in 2 dimensions
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct CartesianCuboid2Vertex {
-    min: [f64; 2],
-    max: [f64; 2],
-    n_vox: [i64; 2],
-    voxel_sizes: [f64; 2],
-}
-
-impl CartesianCuboid2Vertex {
-    fn check_min_max(min: [f64; 2], max: [f64; 2]) -> Result<(), CalcError> {
-        for i in 0..2 {
-            match max[i] > min[i] {
-                false => Err(CalcError(format!(
-                    "Min {:?} must be smaller than Max {:?} for domain boundaries!",
-                    min, max
-                ))),
-                true => Ok(()),
-            }?;
-        }
-        Ok(())
-    }
-
-    fn check_positive<F>(interaction_ranges: [F; 2]) -> Result<(), CalcError>
-    where
-        F: PartialOrd + num::Zero + core::fmt::Debug,
-    {
-        for i in 0..2 {
-            match interaction_ranges[i] > F::zero() {
-                false => Err(CalcError(format!(
-                    "Interaction range must be positive and non-negative! Got value {:?}",
-                    interaction_ranges[i]
-                ))),
-                true => Ok(()),
-            }?;
-        }
-        Ok(())
-    }
-
-    // TODO write this nicely!
-    /// Builds a new [CartesianCuboid2Vertex]
-    /// from given boundaries and maximum interaction ranges of the containing cells.
-    pub fn from_boundaries_and_interaction_ranges(
-        min: [f64; 2],
-        max: [f64; 2],
-        interaction_ranges: [f64; 2],
-    ) -> Result<CartesianCuboid2Vertex, CalcError> {
-        CartesianCuboid2Vertex::check_min_max(min, max)?;
-        CartesianCuboid2Vertex::check_positive(interaction_ranges)?;
-        let mut n_vox = [0; 2];
-        let mut voxel_sizes = [0.0; 2];
-        for i in 0..2 {
-            n_vox[i] = ((max[i] - min[i]) / interaction_ranges[i] * 0.5).ceil() as i64;
-            voxel_sizes[i] = (max[i] - min[i]) / n_vox[i] as f64;
-        }
-        Ok(CartesianCuboid2Vertex {
-            min,
-            max,
-            n_vox,
-            voxel_sizes,
-        })
-    }
-
-    /// Builds a new [CartesianCuboid2Vertex] from given boundaries and the number of voxels per dimension specified
-    pub fn from_boundaries_and_n_voxels(
-        min: [f64; 2],
-        max: [f64; 2],
-        n_vox: [usize; 2],
-    ) -> Result<CartesianCuboid2Vertex, CalcError> {
[…]
