### 2025-04-25 88f7945
introduce new feature which guards plotters usage
