+++ cellular_raza-building-blocks/Cargo.toml
@@ -17,7 +17,7 @@ rustdoc-args = ["--cfg", "doc_cfg", "--html-in-header", ".docs-header.html" ]
 all-features = true
 
 [dependencies]
-plotters = { workspace = true }
+plotters = { workspace = true, optional = true }
 num = { workspace = true }
 rand = { workspace = true, features=["small_rng"] }
 rand_distr = { workspace = true, features = ["serde"]}
@@ -42,3 +42,4 @@ default = ["gradients"]
 pyo3 = ["dep:pyo3"]
 gradients = ["cellular_raza-concepts/gradients"]
 approx = ["dep:approx"]
+plotters = ["dep:plotters"]
