+++ cellular_raza/Cargo.toml
@@ -51,8 +51,9 @@ timestamp = ["cellular_raza-core/timestamp"]
 gradients = ["cellular_raza-concepts/gradients", "cellular_raza-core/gradients", "cellular_raza-building-blocks/gradients"]
 pyo3 = ["cellular_raza-building-blocks/pyo3", "cellular_raza-core/pyo3"]
 approx = ["cellular_raza-building-blocks/approx"]
+plotters = ["cellular_raza-building-blocks/plotters", "cellular_raza-core/plotters"]
 
-cpu_os_threads = ["cellular_raza-core/cpu_os_threads"]
+cpu_os_threads = ["cellular_raza-core/cpu_os_threads", "plotters"]
 chili = ["cellular_raza-core/chili"]
 cara = ["cellular_raza-core/cara"]
 elli = ["cellular_raza-core/elli"]
