+++ cellular_raza-core/Cargo.toml
@@ -71,10 +71,11 @@ tracing = ["dep:tracing", "dep:tracing-subscriber"]
 timestamp = ["dep:chrono"]
 gradients = ["cellular_raza-concepts/gradients"]
 pyo3 = ["dep:pyo3"]
-cpu_os_threads = ["dep:plotters",]
+cpu_os_threads = ["plotters"]
 chili = []
 cara = ["dep:cc", "dep:cudarc"]
 elli = ["dep:wgpu"]
+plotters = ["dep:plotters", "cellular_raza-concepts/plotters"]
 
 # [profile.release]
 # debug = 1
