### 2025-04-25 e48be23
correctly guard plotters functionality
