### 2025-04-25 8365d42
fixup! correctly guard plotters functionality
