+++ cellular_raza-building-blocks/src/domains/mod.rs
@@ -3,6 +3,8 @@ mod cartesian_cuboid_n;
 
 /// Contains deprecated cartesian cuboid implementations
 // TODO #[allow(deprecated)]
+#[cfg(feature = "plotters")]
+#[cfg_attr(docsrs, doc(cfg(feature = "plotters")))]
 pub mod cartesian_cuboid_n_old;
 
 pub use cartesian_cuboid_n::*;
