+++ cellular_raza-core/Cargo.toml
@@ -52,7 +52,7 @@ uniquevec = { version = "0.1.0", features = ["serde"] }
 cudarc = { version = "0.16", optional = true, features = ["cuda-12080"] }
 
 # Additional dependencies for elli backend
-wgpu = { version = "25.0", optional = true }
+opencl3 = { version = "0.12.0", optional = true }
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
@@ -74,7 +74,7 @@ pyo3 = ["dep:pyo3"]
 cpu_os_threads = ["plotters"]
 chili = []
 cara = ["dep:cc", "dep:cudarc"]
-elli = ["dep:wgpu"]
+elli = ["dep:opencl3"]
 plotters = ["dep:plotters", "cellular_raza-concepts/plotters"]
 
 # [profile.release]
