### 2025-05-01 e3a3ef9
merge first example for saxpy kernel with additional iterations
