+++ cellular_raza-core/src/backend/elli/mod.rs
@@ -1,7 +1,166 @@
 //! 🐺 (Placeholder) Cross-platform GPU-centered backend using
-//! [wgpu](https://docs.rs/wgpu/latest/wgpu/)
+//! [opencl](https://docs.rs/opencl3/latest/opencl3/)
+
+struct OpenCLComm {
+    device: opencl3::device::Device,
+    context: opencl3::context::Context,
+    queue: opencl3::command_queue::CommandQueue,
+}
+
+impl OpenCLComm {
+    fn open_connection() -> opencl3::Result<Self> {
+        // TODO we should not be selecting the first device by default
+        let device_id = *opencl3::device::get_all_devices(opencl3::device::CL_DEVICE_TYPE_GPU)?
+            .first()
+            .ok_or(opencl3::error_codes::CL_DEVICE_NOT_AVAILABLE)?;
+        let device = opencl3::device::Device::new(device_id);
+
+        // Create a Context on an OpenCL device
+        let context = opencl3::context::Context::from_device(&device)?;
+
+        // Create a command_queue on the Context's device
+        let queue = opencl3::command_queue::CommandQueue::create_default(
+            &context,
+            opencl3::command_queue::CL_QUEUE_PROFILING_ENABLE,
+        )?;
+
+        Ok(Self {
+            device,
+            context,
+            queue,
+        })
+    }
+
+    fn build_program(
+        &self,
+        source: &str,
+        options: Option<&str>,
+    ) -> opencl3::Result<opencl3::program::Program> {
+        let options = options.unwrap_or("");
+
+        // Build the OpenCL program source and create the kernel.
+        Ok(
+            opencl3::program::Program::create_and_build_from_source(&self.context, source, options)
+                .map_err(|_| opencl3::error_codes::CL_BUILD_PROGRAM_FAILURE)?,
+        )
+    }
+
+    fn create_kernel(
+        &self,
+        program: &opencl3::program::Program,
+        kernel_name: &str,
+    ) -> opencl3::Result<opencl3::kernel::Kernel> {
+        opencl3::kernel::Kernel::create(&program, kernel_name)
+    }
+}
+
+fn my_kernel_saxpy_float(z: &mut [f32], x: &[f32], y: &[f32], a: f32) {
+    z.iter_mut()
+        .zip(x.iter().zip(y.iter()))
+        .for_each(|(z, (x, y))| {
+            *z = a * x + y;
+        })
+}
+
+fn run_main(n_agents: usize) -> opencl3::Result<()> {
+    let comm = OpenCLComm::open_connection()?;
+    let program = comm.build_program(include_str!("saxpy.cl"), None)?;
+    let kernel = comm.create_kernel(&program, "saxpy_float")?;
+
+    /////////////////////////////////////////////////////////////////////
+    // Compute data
+    use opencl3::kernel::ExecuteKernel;
+    use opencl3::memory::Buffer;
+    use opencl3::memory::{CL_MEM_READ_ONLY, CL_MEM_WRITE_ONLY};
+    use opencl3::types::cl_float;
+    use opencl3::types::{CL_BLOCKING, CL_NON_BLOCKING};
+
+    let array_size = n_agents * 8 * 3;
+    // The input data
+    let val1s = vec![0.0; array_size];
+    let val2s = vec![1.0; array_size];
+
+    let t4 = std::time::Instant::now();
+    // Create OpenCL device buffers
+    let mut x = unsafe {
+        Buffer::<cl_float>::create(
+            &comm.context,
+            CL_MEM_READ_ONLY,
+            array_size,
+            core::ptr::null_mut(),
+        )?
[…]
