+++ Cargo.toml
@@ -12,9 +12,6 @@ members = [
 ]
 resolver = "2"
 
-[patch.crates-io]
-nalgebra = { git = "https://github.com/dimforge/nalgebra" }
-pathfinder_simd = { git = "https://github.com/servo/pathfinder.git" }
 
 [workspace.package]
 version = "0.2.3"
@@ -44,7 +41,7 @@ rand_distr = "0.5"
 rand_chacha = { version="0.9", features=["serde"] }
 serde = "1.0"
 tempfile = "3"
-nalgebra = { version="<0.34, >=0.30", features=["serde-serialize", "rand"]}
+nalgebra = { package="nalgebra-randv09", version="0.33", features=["serde-serialize", "rand"]}
 ndarray = { version = "0.16", features = ["blas", "serde", "serde-1"] }
 pyo3 = { version = "0.24", features = ["serde", "py-clone"]}
 kdam = "0.6"
