### 2025-05-01 77e590e
bump to 0.2.4
