+++ cellular_raza-core/Cargo.toml
@@ -56,11 +56,11 @@ opencl3 = { version = "0.12.0", optional = true }
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.2.3"
+version = "0.2.4"
 
 [dependencies.cellular_raza-core-proc-macro]
 path = "../cellular_raza-core-proc-macro"
-version = "0.2.3"
+version = "0.2.4"
 
 [dev-dependencies]
 tempfile.workspace = true
