+++ cellular_raza-building-blocks/Cargo.toml
@@ -35,7 +35,7 @@ serde_test = "1.0.177"
 
 [dependencies.cellular_raza-concepts]
 path = "../cellular_raza-concepts"
-version = "0.2.3"
+version = "0.2.4"
 
 [features]
 default = ["gradients"]
