### 2025-05-01 55f3976
switch back to nalgebra version and use patch from github
