### 2025-05-01 6ce00c4
use github patch for pathfinder_simd
