### 2025-05-02 aa3ac8f
add benchmark for interactions of particles with almost no effect
