### 2025-05-03 07a4653
add documentation for update_mechanics functions
