### 2025-05-03 84556ff
split benchmark into preparation code and execution
