### 2025-05-06 30628fa
explicitly disable incremental compilation
