+++ cellular_raza-building-blocks/Cargo.toml
@@ -42,4 +42,4 @@ default = ["gradients"]
 pyo3 = ["dep:pyo3"]
 gradients = ["cellular_raza-concepts/gradients"]
 approx = ["dep:approx"]
-plotters = ["dep:plotters"]
+plotters = ["dep:plotters", "cellular_raza-concepts/plotters"]
