### 2025-05-06 e4ac88e
deactivate extension-module
