### 2025-05-06 28d90ea
experimental: include all packages in test for macos
