### 2025-05-06 1944522
implement error conversion for pyo3::PyErr
