### 2025-05-06 de554c9
fixup! implement error conversion for pyo3::PyErr
