### 2025-05-08 09e426a
implement AbsDiffEq for RodInteraction
