### 2025-05-13 b00a75a
add stronger incentive for inlining
