+++ cellular_raza-concepts-derive/src/cell_agent.rs
@@ -328,7 +328,7 @@ impl AgentImplementer {
                 #[automatically_derived]
                 impl #impl_generics Cycle<#tokens>
                 for #struct_name #struct_ty_generics #where_clause {
-                    #[inline]
+                    #[inline(always)]
                     fn update_cycle(
                         rng: &mut rand_chacha::ChaCha8Rng,
                         dt: &#float_type,
@@ -337,7 +337,7 @@ impl AgentImplementer {
                         <#field_type as Cycle<#tokens>>::update_cycle(rng, dt, cell)
                     }
 
-                    #[inline]
+                    #[inline(always)]
                     fn divide(
                         rng: &mut rand_chacha::ChaCha8Rng,
                         cell: &mut Self
@@ -345,6 +345,7 @@ impl AgentImplementer {
                         <#field_type as Cycle<#tokens>>::divide(rng, cell)
                     }
 
+                    #[inline(always)]
                     fn update_conditional_phased_death(
                         rng: &mut rand_chacha::ChaCha8Rng,
                         dt: &#float_type,
@@ -392,7 +393,7 @@ impl AgentImplementer {
                 impl #impl_generics Mechanics<#tokens> for #struct_name #struct_ty_generics
                     #where_clause
                 {
-                    #[inline]
+                    #[inline(always)]
                     fn calculate_increment(&self, force: #force)
                         -> Result<(#position, #velocity), CalcError> {
                         <#field_type as Mechanics<#tokens>>::calculate_increment(
@@ -400,7 +401,7 @@ impl AgentImplementer {
                             force
                         )
                     }
-                    #[inline]
+                    #[inline(always)]
                     fn get_random_contribution(
                         &self,
                         rng: &mut rand_chacha::ChaCha8Rng,
@@ -440,12 +441,12 @@ impl AgentImplementer {
                 impl #impl_generics Position<#tokens> for #struct_name #struct_ty_generics
                     #where_clause
                 {
-                    #[inline]
+                    #[inline(always)]
                     fn pos(&self) -> #position {
                         <#field_type as Position<#tokens>>::pos(&self.#field_name)
                     }
 
-                    #[inline]
+                    #[inline(always)]
                     fn set_pos(&mut self, pos: &#position) {
                         <#field_type as Position<#tokens>>::set_pos(&mut self.#field_name, pos)
                     }
@@ -477,12 +478,12 @@ impl AgentImplementer {
                 impl #impl_generics Velocity<#tokens> for #struct_name #struct_ty_generics
                     #where_clause
                 {
-                    #[inline]
+                    #[inline(always)]
                     fn velocity(&self) -> #velocity {
                         <#field_type as Velocity<#tokens>>::velocity(&self.#field_name)
                     }
 
-                    #[inline]
+                    #[inline(always)]
                     fn set_velocity(&mut self, velocity: &#velocity) {
                         <#field_type as Velocity<#tokens>>::set_velocity(&mut self.#field_name, velocity)
                     }
@@ -520,7 +521,7 @@ impl AgentImplementer {
                 #[automatically_derived]
                 impl #impl_generics Interaction<#tokens>
                     for #struct_name #struct_ty_generics #where_clause {
-                    #[inline]
+                    #[inline(always)]
                     fn get_interaction_information(&self) -> #information {
                         <#field_type as Interaction<#tokens
                         >>::get_interaction_information(
@@ -528,7 +529,7 @@ impl AgentImplementer {
                         )
                     }
 
-                    #[inline]
+                    #[inline(always)]
                     fn calculate_force_between(
                         &self,
                         own_pos: &#position,
@@ -547,7 +548,7 @@ impl AgentImplementer {
                         )
                     }
 
-                    #[inline]
+                    #[inline(always)]
                     fn is_neighbor(
[…]
