+++ cellular_raza-concepts-derive/src/domain.rs
@@ -268,157 +268,157 @@ impl DomainImplementer {
             let impl_generics = generics.split_for_impl().0;
 
             quote::quote!(
-            #[automatically_derived]
-            impl #impl_generics Domain<#tokens> for #struct_name #struct_ty_generics
-            #where_clause
-                Self: DomainRngSeed,
-                Self: DomainCreateSubDomains<#subdomain>,
-                Self: SortCells<
-                    #cell,
-                    VoxelIndex = <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex
-                >,
-                #subdomain: SubDomain<
-                    VoxelIndex = <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex
-                >,
-                <Self as DomainCreateSubDomains<#subdomain>>::SubDomainIndex: Clone
-                    + core::hash::Hash + Eq + Ord,
-                <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex: Clone
-                    + core::hash::Hash + Eq + Ord,
-                #cell_iterator: IntoIterator<Item = #cell>,
-            {
-                type SubDomainIndex = <Self as DomainCreateSubDomains<#subdomain>>::SubDomainIndex;
-                type VoxelIndex = <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex;
+                #[automatically_derived]
+                impl #impl_generics Domain<#tokens> for #struct_name #struct_ty_generics
+                #where_clause
+                    Self: DomainRngSeed,
+                    Self: DomainCreateSubDomains<#subdomain>,
+                    Self: SortCells<
+                        #cell,
+                        VoxelIndex = <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex
+                    >,
+                    #subdomain: SubDomain<
+                        VoxelIndex = <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex
+                    >,
+                    <Self as DomainCreateSubDomains<#subdomain>>::SubDomainIndex: Clone
+                        + core::hash::Hash + Eq + Ord,
+                    <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex: Clone
+                        + core::hash::Hash + Eq + Ord,
+                    #cell_iterator: IntoIterator<Item = #cell>,
+                {
+                    type SubDomainIndex = <Self as DomainCreateSubDomains<#subdomain>>::SubDomainIndex;
+                    type VoxelIndex = <Self as DomainCreateSubDomains<#subdomain>>::VoxelIndex;
 
-                fn decompose(
-                    self,
-                    n_subdomains: core::num::NonZeroUsize,
-                    cells: #cell_iterator,
-                ) -> Result<
-                    DecomposedDomain<Self::SubDomainIndex, #subdomain, #cell>,
-                    DecomposeError
-                > {
-                    // Get all subdomains
-                    let subdomains: Vec<_> = self.create_subdomains(n_subdomains)?
-                        .into_iter()
-                        .collect();
-
-                    // Build a map from a voxel_index to the subdomain_index in which it is
-                    let mut voxel_index_to_subdomain_index:
-                        ::std::collections::BTreeMap<Self::VoxelIndex, Self::SubDomainIndex> =
-                        subdomains
-                        .iter()
-                        .map(|(subdomain_index, subdomain, _)| subdomain
-                            .get_all_indices()
+                    fn decompose(
+                        self,
+                        n_subdomains: core::num::NonZeroUsize,
+                        cells: #cell_iterator,
+                    ) -> Result<
+                        DecomposedDomain<Self::SubDomainIndex, #subdomain, #cell>,
+                        DecomposeError
+                    > {
+                        // Get all subdomains
+                        let subdomains: Vec<_> = self.create_subdomains(n_subdomains)?
[…]
