### 2025-05-13 05ad6fa
simplify name of benchmark
