### 2025-05-13 ffd375e
add new benchmark for testing in one subdomain with many cells
