### 2025-05-13 ff134a2
reorder statements for small performance boost (ca. 4%)
