### 2025-05-13 73d5de2
correctly initialize cells; use new method in benchmark
