+++ cellular_raza-benchmarks/Cargo.toml
@@ -26,6 +26,7 @@ kdam.workspace = true
 serde_json = "1.0.118"
 glob = "0.3.1"
 tempfile.workspace = true
+itertools.workspace = true
 
 [dependencies.cellular_raza]
 path = "../cellular_raza"
