### 2025-05-13 7536032
rewrite benchmark with new function
