### 2025-05-14 833f5d5
split interaction trait into 3 parts:
1. InteractionInformation: exposes information
2. Interaction: For Calculating Interaction Forces
3. NeighborInteraction: Allows reacting to neighbor count
