### 2025-05-14 24261e8
implement new traits for modularcell
