### 2025-05-14 67d0cdc
implement derivation of new concepts
