+++ cellular_raza-concepts-derive/src/cell_agent.rs
@@ -77,6 +77,10 @@ enum CellAspect {
     Velocity,
     Cycle,
     Interaction,
+    InteractionRaw,
+    InteractionInformation,
+    NeighborInteraction,
+    NeighborInteractionRaw,
     Intracellular,
     Reactions,
     ReactionsRaw,
@@ -100,6 +104,10 @@ impl CellAspect {
                 "Velocity" => Some(CellAspect::Velocity),
                 "Cycle" => Some(CellAspect::Cycle),
                 "Interaction" => Some(CellAspect::Interaction),
+                "InteractionRaw" => Some(CellAspect::InteractionRaw),
+                "InteractionInformation" => Some(CellAspect::InteractionInformation),
+                "NeighborInteraction" => Some(CellAspect::NeighborInteraction),
+                "NeighborInteractionRaw" => Some(CellAspect::NeighborInteractionRaw),
                 "Intracellular" => Some(CellAspect::Intracellular),
                 "Reactions" => Some(CellAspect::Reactions),
                 "ReactionsRaw" => Some(CellAspect::ReactionsRaw),
@@ -128,8 +136,7 @@ impl CellAspectField {
         let aspects = field
             .attrs
             .iter()
-            .map(CellAspect::from_attribute)
-            .filter_map(|s| s)
+            .filter_map(CellAspect::from_attribute)
             .collect::<Vec<_>>();
         Self { aspects, field }
     }
@@ -142,12 +149,12 @@ impl CellAspectField {
             syn::Fields::Named(fields_named) => Ok(fields_named
                 .named
                 .into_iter()
-                .map(|field| CellAspectField::from_field(field))
+                .map(CellAspectField::from_field)
                 .collect::<Vec<_>>()),
             syn::Fields::Unnamed(fields_unnamed) => Ok(fields_unnamed
                 .unnamed
                 .into_iter()
-                .map(|field| CellAspectField::from_field(field))
+                .map(CellAspectField::from_field)
                 .collect::<Vec<_>>()),
             syn::Fields::Unit => Err(syn::Error::new(span, "Cannot derive from unit struct")),
         }
@@ -184,7 +191,9 @@ pub struct AgentImplementer {
     mechanics_raw: Option<FieldInfo>,
     position: Option<FieldInfo>,
     velocity: Option<FieldInfo>,
-    interaction: Option<FieldInfo>,
+    interaction_raw: Option<FieldInfo>,
+    interaction_information: Option<FieldInfo>,
+    neighbor_interaction_raw: Option<FieldInfo>,
     intracellular: Option<FieldInfo>,
     reactions_raw: Option<FieldInfo>,
     reactions_extra_raw: Option<FieldInfo>,
@@ -200,6 +209,9 @@ impl From<AgentParser> for AgentImplementer {
         let mut position = None;
         let mut velocity = None;
         let mut interaction = None;
+        let mut interaction_raw = None;
+        let mut interaction_information = None;
+        let mut neighbor_interaction_raw = None;
         let mut intracellular = None;
         let mut reactions_raw = None;
         let mut reactions_extra_raw = None;
@@ -233,7 +245,21 @@ impl From<AgentParser> for AgentImplementer {
                         CellAspect::Position => position = Some(field_info),
                         CellAspect::Velocity => velocity = Some(field_info),
                         CellAspect::Interaction => {
-                            interaction = Some(field_info);
+                            interaction = Some(field_info.clone());
+                            interaction_information = Some(field_info);
+                        }
+                        CellAspect::InteractionRaw => {
+                            interaction_raw = Some(field_info);
+                        }
+                        CellAspect::InteractionInformation => {
+                            interaction_information = Some(field_info);
+                        }
+                        CellAspect::NeighborInteraction => {
[…]
