+++ cellular_raza-concepts-derive/src/lib.rs
@@ -31,6 +31,9 @@ mod domain;
         Position,
         Velocity,
         Interaction,
+        InteractionRaw,
+        InteractionInformation,
+        NeighborInteraction,
         Reactions,
         ReactionsContact,
         ReactionsRaw,
