### 2025-05-14 4ea3bd8
do not implement for all Boxed types
