### 2025-05-14 ccb71fb
adjust test to new trait formulation
