### 2025-05-14 b79a2d6
adjust building blocks to new concepts
