### 2025-05-14 b9d5d9c
add trait bound for method
