### 2025-05-14 4777e41
add nonworking (empty) methods for possible future use
