### 2025-05-14 868f23c
include NeighborInteraction in list of Simulation Aspects
