### 2025-05-14 14dcca1
remove implementation for Boxed types
