+++ cellular_raza-concepts/src/interaction.rs
@@ -36,32 +36,3 @@ pub trait Interaction<Pos, Vel, Force, Inf = ()> {
     // TODO
     // fn contact_function(&mut self, other_cell: &C, environment: &mut Env) -> Result<(), SimulationError>;
 }
-
-impl<Pos, Vel, For, Inf> Interaction<Pos, Vel, For, Inf>
-    for Box<dyn Interaction<Pos, Vel, For, Inf>>
-{
-    fn get_interaction_information(&self) -> Inf {
-        use core::ops::Deref;
-        self.deref().get_interaction_information()
-    }
-    fn calculate_force_between(
-        &self,
-        own_pos: &Pos,
-        own_vel: &Vel,
-        ext_pos: &Pos,
-        ext_vel: &Vel,
-        ext_info: &Inf,
-    ) -> Result<(For, For), CalcError> {
-        use core::ops::Deref;
-        self.deref()
-            .calculate_force_between(own_pos, own_vel, ext_pos, ext_vel, ext_info)
-    }
-    fn is_neighbor(&self, own_pos: &Pos, ext_pos: &Pos, ext_inf: &Inf) -> Result<bool, CalcError> {
-        use core::ops::Deref;
-        self.deref().is_neighbor(own_pos, ext_pos, ext_inf)
-    }
-    fn react_to_neighbors(&mut self, neighbors: usize) -> Result<(), CalcError> {
-        use core::ops::DerefMut;
-        self.deref_mut().react_to_neighbors(neighbors)
-    }
-}
