### 2025-05-14 f7de45a
distinguish between Interaction and InteractionInformation for deriving
