+++ cellular_raza-concepts-derive/src/lib.rs
@@ -31,6 +31,8 @@ mod domain;
         Position,
         Velocity,
         Interaction,
+        InteractionRaw,
+        InteractionInformation,
         Reactions,
         ReactionsContact,
         ReactionsRaw,
