+++ cellular_raza-concepts-derive/src/cell_agent.rs
@@ -77,6 +77,8 @@ enum CellAspect {
     Velocity,
     Cycle,
     Interaction,
+    InteractionRaw,
+    InteractionInformation,
     Intracellular,
     Reactions,
     ReactionsRaw,
@@ -100,6 +102,8 @@ impl CellAspect {
                 "Velocity" => Some(CellAspect::Velocity),
                 "Cycle" => Some(CellAspect::Cycle),
                 "Interaction" => Some(CellAspect::Interaction),
+                "InteractionRaw" => Some(CellAspect::InteractionRaw),
+                "InteractionInformation" => Some(CellAspect::InteractionInformation),
                 "Intracellular" => Some(CellAspect::Intracellular),
                 "Reactions" => Some(CellAspect::Reactions),
                 "ReactionsRaw" => Some(CellAspect::ReactionsRaw),
@@ -183,7 +187,8 @@ pub struct AgentImplementer {
     mechanics_raw: Option<FieldInfo>,
     position: Option<FieldInfo>,
     velocity: Option<FieldInfo>,
-    interaction: Option<FieldInfo>,
+    interaction_raw: Option<FieldInfo>,
+    interaction_information: Option<FieldInfo>,
     intracellular: Option<FieldInfo>,
     reactions_raw: Option<FieldInfo>,
     reactions_extra_raw: Option<FieldInfo>,
@@ -199,6 +204,8 @@ impl From<AgentParser> for AgentImplementer {
         let mut position = None;
         let mut velocity = None;
         let mut interaction = None;
+        let mut interaction_raw = None;
+        let mut interaction_information = None;
         let mut intracellular = None;
         let mut reactions_raw = None;
         let mut reactions_extra_raw = None;
@@ -232,7 +239,14 @@ impl From<AgentParser> for AgentImplementer {
                         CellAspect::Position => position = Some(field_info),
                         CellAspect::Velocity => velocity = Some(field_info),
                         CellAspect::Interaction => {
-                            interaction = Some(field_info);
+                            interaction = Some(field_info.clone());
+                            interaction_information = Some(field_info);
+                        }
+                        CellAspect::InteractionRaw => {
+                            interaction_raw = Some(field_info);
+                        }
+                        CellAspect::InteractionInformation => {
+                            interaction_information = Some(field_info);
                         }
                         CellAspect::Intracellular => {
                             intracellular = Some(field_info);
@@ -273,7 +287,8 @@ impl From<AgentParser> for AgentImplementer {
             mechanics_raw,
             position,
             velocity,
-            interaction,
+            interaction_raw,
+            interaction_information,
             intracellular,
             reactions_raw,
             reactions_extra_raw,
@@ -484,7 +499,9 @@ impl AgentImplementer {
 
                     #[inline(always)]
                     fn set_velocity(&mut self, velocity: &#velocity) {
-                        <#field_type as Velocity<#tokens>>::set_velocity(&mut self.#field_name, velocity)
+                        <#field_type as Velocity<#tokens>>::set_velocity(
+                            &mut self.#field_name, velocity
+                        )
                     }
                 }
             };
@@ -493,11 +510,11 @@ impl AgentImplementer {
         TokenStream::new()
     }
 
-    pub fn implement_interaction(&self) -> TokenStream {
+    pub fn implement_interaction_raw(&self) -> TokenStream {
         let struct_name = &self.name;
         let (_, struct_ty_generics, struct_where_clause) = &self.generics.split_for_impl();
 
-        if let Some(field_info) = &self.interaction {
+        if let Some(field_info) = &self.interaction_raw {
             let field_name = &field_info.field_name;
             let field_type = &field_info.field_type;
[…]
