### 2025-05-14 f8247b5
split trait into Interaction and InteractionInformation
