+++ cellular_raza-concepts/tests/derive_cell_agent.rs
@@ -1,3 +1,5 @@
+use cellular_raza_concepts::InteractionInformation;
+
 #[test]
 fn derive_cycle() {
     use cellular_raza_concepts::*;
@@ -110,8 +112,7 @@ fn derive_interaction() {
     use cellular_raza_concepts::*;
     use cellular_raza_concepts_derive::CellAgent;
     struct InteractionModel;
-    impl cellular_raza_concepts::Interaction<f32, f32, f32> for InteractionModel {
-        fn get_interaction_information(&self) -> () {}
+    impl cellular_raza_concepts::Interaction<f32, f32, f32, ()> for InteractionModel {
         fn calculate_force_between(
             &self,
             _own_pos: &f32,
@@ -122,6 +123,9 @@ fn derive_interaction() {
         ) -> Result<(f32, f32), CalcError> {
             unimplemented!()
         }
+    }
+    impl cellular_raza_concepts::InteractionInformation<()> for InteractionModel {
+        fn get_interaction_information(&self) {}
     }
     #[derive(CellAgent)]
     struct NewAgent1 {
@@ -137,7 +141,7 @@ fn derive_interaction() {
 
 #[test]
 fn derive_interaction_generics() {
-    use cellular_raza_concepts::{CalcError, Interaction};
+    use cellular_raza_concepts::Interaction;
     use cellular_raza_concepts_derive::CellAgent;
 
     struct InteractionModel<const D: usize> {
@@ -145,9 +149,6 @@ fn derive_interaction_generics() {
     }
 
     impl<const D: usize> Interaction<f32, f32, f32, [usize; D]> for InteractionModel<D> {
-        fn get_interaction_information(&self) -> [usize; D] {
-            self.index.clone()
-        }
         fn calculate_force_between(
             &self,
             _own_pos: &f32,
@@ -157,6 +158,12 @@ fn derive_interaction_generics() {
             _ext_info: &[usize; D],
         ) -> Result<(f32, f32), cellular_raza_concepts::CalcError> {
             Ok((0.0, 0.0))
+        }
+    }
+
+    impl<const D: usize> InteractionInformation<[usize; D]> for InteractionModel<D> {
+        fn get_interaction_information(&self) -> [usize; D] {
+            self.index
         }
     }
 
