+++ cellular_raza-concepts/src/interaction.rs
@@ -1,12 +1,13 @@
 use crate::errors::CalcError;
 
-/// Trait describing force-interactions between cellular agents.
-pub trait Interaction<Pos, Vel, Force, Inf = ()> {
+/// Exposes information which is used to calculate interactions between cells and the domain.
+pub trait InteractionInformation<Inf> {
     /// Get additional information of cellular properties (ie. for cell-specific interactions).
-    /// For now, this can also be used to get the mass of the other cell-agent.
-    /// In the future, we will probably provide a custom function for this.
     fn get_interaction_information(&self) -> Inf;
+}
 
+/// Trait describing force-interactions between cellular agents.
+pub trait Interaction<Pos, Vel, Force, Inf = ()>: InteractionInformation<Inf> {
     /// Calculates the forces (velocity-derivative) on the corresponding external position given
     /// external velocity.
     /// By providing velocities, we can calculate terms that are related to friction.
