### 2025-05-14 48b82e7
require information for SubDomainForce
