### 2025-05-14 dbb4d64
implement new derivation for SubDomainForce
