### 2025-05-14 a81730e
implement new traits for building blocks
