### 2025-05-14 90645ab
aim at inlining function calls more explicitly; do not create named closures
