+++ cellular_raza-core-proc-macro/src/run_sim.rs
@@ -298,32 +298,37 @@ pub fn run_main_update(kwargs: KwargsMain) -> proc_macro2::TokenStream {
             .push(quote!(#core_path::backend::chili::local_subdomain_update_reactions_extra));
     }
 
-    let update_local_funcs = quote!(
-        let __cr_private_combined_local_subdomain_funcs = |
-            subdomain: &mut _,
-            dt,
-        | -> Result<(), #core_path::backend::chili::SimulationError> {
-            #(
-                #local_subdomain_func_names(subdomain, dt)?;
-            )*
-            Ok(())
-        };
+    let update_local_subdomain_funcs = quote!(
         sbox.run_local_subdomain_funcs(
-            __cr_private_combined_local_subdomain_funcs,
+            #[inline(always)]
+            |
+                subdomain: &mut _,
+                dt,
+            | -> Result<(), #core_path::backend::chili::SimulationError> {
+                #(
+                    #local_subdomain_func_names(subdomain, dt)?;
+                )*
+                Ok(())
+            },
             &next_time_point
         )?;
-        let __cr_private_combined_local_cell_funcs = |
-            cell: &mut _,
-            aux_storage: &mut _,
-            dt,
-            rng: &mut rand_chacha::ChaCha8Rng
-        | -> Result<(), #core_path::backend::chili::SimulationError> {
-            #(
-                #local_func_names(cell, aux_storage, dt, rng)?;
-            )*
-            Ok(())
-        };
-        sbox.run_local_cell_funcs(__cr_private_combined_local_cell_funcs, &next_time_point)?;
+    );
+    let update_local_cell_funcs = quote!(
+        sbox.run_local_cell_funcs(
+            #[inline(always)]
+            |
+                cell: &mut _,
+                aux_storage: &mut _,
+                dt,
+                rng: &mut rand_chacha::ChaCha8Rng
+            | -> Result<(), #core_path::backend::chili::SimulationError> {
+                #(
+                    #local_func_names(cell, aux_storage, dt, rng)?;
+                )*
+                Ok(())
+            },
+            &next_time_point
+        )?;
     );
 
     quote!(
@@ -345,7 +350,8 @@ pub fn run_main_update(kwargs: KwargsMain) -> proc_macro2::TokenStream {
                 #step_2
                 sbox.sync()?;
                 #step_3
-                #update_local_funcs
+                #update_local_subdomain_funcs
+                #update_local_cell_funcs
                 #step_4
                 sbox.sync()?;
                 #step_5
