### 2025-05-14 3992f3b
cleanup; report first error instead of first element of loop; this loop would not have run but simply returned the first element
