### 2025-05-14 41f7954
add Interaction Information to calculation of DomainForce
