### 2025-05-14 30bcbe4
minor bugfix; almost duplicate variable names
