### 2025-05-14 5c9f231
rename trait bound to avoid confusion with new concept trait
