### 2025-05-14 e4c930e
adjust bulding blocks to new Interaction traits
