### 2025-05-14 835353f
adjust benchmarks to new Interaction traits
