### 2025-05-14 86ee239
adjust integration tests to new Interaction traits
