### 2025-05-14 08aef38
adjust all examples (including templates) to new Interaction traits
