+++ cellular_raza-examples/semi_vertex/src/cell_properties.rs
@@ -59,8 +59,10 @@ impl Interaction<Vector2<f64>, Vector2<f64>, Vector2<f64>> for OutsideInteractio
         let force = -dir * strength * spatial_cutoff;
         Ok((-force, force))
     }
+}
 
-    fn get_interaction_information(&self) -> () {}
+impl InteractionInformation<()> for OutsideInteraction {
+    fn get_interaction_information(&self) {}
 }
 
 impl Interaction<Vector2<f64>, Vector2<f64>, Vector2<f64>> for InsideInteraction {
@@ -80,8 +82,10 @@ impl Interaction<Vector2<f64>, Vector2<f64>, Vector2<f64>> for InsideInteraction
         let force = self.potential_strength * dir / (0.5 + 0.5 * r / self.average_radius);
         Ok((-force, force))
     }
+}
 
-    fn get_interaction_information(&self) -> () {}
+impl InteractionInformation<()> for InsideInteraction {
+    fn get_interaction_information(&self) {}
 }
 
 impl<const D: usize> Cycle for MyCell<D> {
