### 2025-05-19 6e4c58c
loosen requirement on ndarray
