### 2025-05-19 cfee46f
loosen dependency on pyo3
