+++ Cargo.toml
@@ -46,7 +46,7 @@ serde = "1.0"
 tempfile = "3"
 nalgebra = { version="0.33", features=["serde-serialize", "rand"]}
 ndarray = { version = ">=0.15", features = ["blas", "serde", "serde-1"] }
-pyo3 = { version = "0.24", features = ["serde", "py-clone"]}
+pyo3 = { version = ">=0.24", features = ["serde", "py-clone"]}
 kdam = "0.6"
 approxim = "0.6.5"
 approx = { package = "approxim", version = "0.6.5" }
