### 2025-05-28 #7723 @jonaspleyer
Yes I have already archived version 0.2.3 (see comment https://github.com/openjournals/joss-reviews/issues/7723#issuecomment-2757951467). I hope that this record is satisfactory.
