### 2025-05-28 #7723 @mahfuz05062
> Yes I have already archived version 0.2.3 (see comment [#7723 (comment)](https://github.com/openjournals/joss-reviews/issues/7723#issuecomment-2757951467)). I hope that this record is satisfactory.

I was able to find the DOI link from there, but I would suggest to edit the name with the article name (as shown on the paper).
