### 2025-06-02 #7723 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
