### 2025-06-02 64c8df7
minor correction to coloring of cells
