### 2025-06-02 #7723 @jonaspleyer
We have reviewed the final draft and found two very minor mistakes. The first was a formatting error: 'cellular_raza' != `cellular_raza`. And the second one was in the description of the coloring of cellular agents in the branching pattern simulation. It should now state more accurately how the coloring is done.
