### 2025-06-07 #7723 @Kevin-Mattheus-Moerman
- [ ] Please complete the affiliation to also have the city and country.
- [ ] The archive lists the software as `v1` but this should match the release tag `0.2.3`. Please manually edit the Zenodo archive to match the license tag.
- [ ] `behaviour` is British English whereas words like `self-organization`, `customize`, `visualizing`, and `utilizes` are American English, please use one form consistently.
- [ ] The second author lacks an Orcid. Is this perhaps their Orcid: https://orcid.org/0000-0002-6371-4495, if so consider adding it.
