+++ paper.md
@@ -11,6 +11,7 @@ authors:
     affiliation: 1
   - name: Christian Fleck
     affiliation: 1
+    orcid: 0000-0002-6371-4495
 affiliations:
  - name: Freiburg Center for Data Analysis, Modeling and AI
    index: 1
