### 2025-06-08 f8d0225
formatting + adapt to american english
