### 2025-06-08 #7723 @jonaspleyer
> Please complete the affiliation to also have the city and country.

I have added the missing information. I hope that the used identifiers are correct. I could not find any information about the "city" and "country" keys in the documentation (https://joss.readthedocs.io/en/latest/paper.html#affiliations).
(commit 398553b2b2dc5a69fb46b28b1b05bc34a1ff8e15)

> The archive lists the software as v1 but this should match the release tag 0.2.3. Please manually edit the Zenodo archive to match
> the license tag.

I have updated the version tag to 0.2.3.

> behaviour is British English whereas words like self-organization, customize, visualizing, and utilizes are American English, please use one form consistently.

I have updated the text and checked for additional formulations which might be British English. Thanks for you attention to detail in catching those in the first place. 💯
(commit f8d022532a42524f22afaf71474ca96eed256542)

> The second author lacks an Orcid. Is this perhaps their Orcid: https://orcid.org/0000-0002-6371-4495, if so consider adding it.

Added it. Thanks for the heads-up. 👍
(commit 81b52ce1fd7cf7101a04be0b6279126602e4fa09)
