### 2025-06-09 52f413e
Update affiliation in paper.md
