+++ paper.md
@@ -13,10 +13,8 @@ authors:
     affiliation: 1
     orcid: 0000-0002-6371-4495
 affiliations:
-  - name: Freiburg Center for Data Analysis, Modeling and AI
-    index: 1
-    city: Freiburg im Breisgau
-    country: Germany
+  - name: Albert-Ludwigs-Universität Freiburg, Freiburg Center for Data Analysis, Modeling and AI, Freiburg im Breisgau, Germany
+    index: 1    
 date: 13 March 2025
 bibliography: paper.bib
 
