### 2025-06-09 295ab4e
update affiliation to match other publications
Co-authored-by: Kevin Mattheus Moerman <Kevin-Mattheus-Moerman@users.noreply.github.com>
