+++ paper.md
@@ -13,7 +13,7 @@ authors:
     affiliation: 1
     orcid: 0000-0002-6371-4495
 affiliations:
-  - name: Albert-Ludwigs-Universität Freiburg, Freiburg Center for Data Analysis, Modeling and AI, Freiburg im Breisgau, Germany
+  - name: Freiburg Center for Data Analysis and Modelling and AI (FDMAI), University of Freiburg, Freiburg, Germany
     index: 1    
 date: 13 March 2025
 bibliography: paper.bib
