### 2025-06-10 20c2b48
update CITATION.cff file with contents from JOSS publication
